Experimental Integrators

Missing docstring.

Missing docstring for GeometricIntegrators.Integrators.IntegratorExplicitEuler. Check Documenter's build log for details.

Missing docstring.

Missing docstring for GeometricIntegrators.Integrators.IntegratorImplicitEuler. Check Documenter's build log for details.

GeometricIntegrators.Integrators.IntegratorIRKimplicit — Type

Implicit Runge-Kutta integrator for implicit systems of equations, solving the system

\[\begin{aligned} P_{n,i} &= \vartheta (Q_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , \\ F_{n,i} &= f (Q_{n,i}, V_{n,i}) , & P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} . \end{aligned}\]

If implicit_update is set to true, the update is computed by solving

\[\vartheta(q_{n+1}) = \vartheta(q_{n}) + h \sum \limits_{i=1}^{s} b_{i} \, f (Q_{n,j}, V_{n,j}) ,\]

otherwise it is computed explicitly by

\[q_{n+1} = q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} .\]

Usually we are interested in Lagrangian systems, where

\[\begin{aligned} P_{n,i} &= \dfrac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , & F_{n,i} &= \dfrac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , \end{aligned}\]

and tableaus satisfying the symplecticity conditions

\[\begin{aligned} b_{i} \bar{a}_{ij} + \bar{b}_{j} a_{ji} &= b_{i} \bar{b}_{j} , & \bar{b}_i &= b_i . \end{aligned}\]