Equations

GeometricIntegrators.Equations.DAE — Type.

DAE: Differential Algebraic Equation

Defines a differential algebraic initial value problem

\[\begin{align*} \dot{q} (t) &= f(t, q(t)) + u(t, q(t), \lambda(t)) , & q(t_{0}) &= q_{0} , \\ 0 &= \phi (t, q(t), \lambda(t)) , & \lambda(t_{0}) &= \lambda_{0} , \end{align*}\]

with vector field $f$, projection $u$, algebraic constraint $\phi=0$, initial conditions $q_{0}$ and $\lambda_{0}$, the dynamical variable $q$ taking values in $\mathbb{R}^{m}$ and the algebraic variable $\lambda$ taking values in $\mathbb{R}^{n}$.

Fields

m: dimension of dynamical variable $q$ and the vector field $f$

n: dimension of algebraic variable $\lambda$ and the constraint $\phi$

f: function computing the vector field

u: function computing the projection

ϕ: algebraic constraint

t₀: initial time

q₀: initial condition for dynamical variable $q$

λ₀: initial condition for algebraic variable $\lambda$

The function f, providing the vector field, takes three arguments, f(t, q, fq), the functions u and ϕ, providing the projection and the algebraic constraint take four arguments, u(t, q, λ, fu) and ϕ(t, q, λ, fϕ), where t is the current time, q and λ are the current solution vectors, and fq, fu and fϕ are the vectors which hold the result of evaluating the vector field $f$, the projection $u$ and the algebraic constraint $\phi$ on t, q and λ.

Example

    function f(t, q, fq)
        fq[1] = q[1]
        fq[2] = q[2]
    end

    function u(t, q, λ, fu)
        fu[1] = +λ[1]
        fu[2] = -λ[1]
    end

    function ϕ(t, q, λ, fϕ)
        fϕ[1] = q[2] - q[1]
    end

    t₀ = 0.
    q₀ = [1., 1.]
    λ₀ = [0.]

    dae = DAE(f, u, ϕ, t₀, q₀, λ₀)

source

GeometricIntegrators.Equations.IODE — Type.

IODE: Implicit Ordinary Differential Equation

Defines an implicit initial value problem

\[\begin{align*} \dot{q} (t) &= v(t) , & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), v(t)) , & p(t_{0}) &= p_{0} , \\ p(t) &= g(t, q(t), v(t)) \end{align*}\]

with vector fields $f$ and $g$, initial conditions $(q_{0}, p_{0})$ and the solution $(q,p)$ taking values in $\mathbb{R}^{d} \times \mathbb{R}^{d}$. This is a special case of a differential algebraic equation with dynamical variables $(q,p)$ and algebraic variable $v$.

Fields

d: dimension of dynamical variables $q$ and $p$ as well as the vector fields $f$ and $g$

f: function computing the vector field

g: function determining the algebraic constraint

t₀: initial time

q₀: initial condition for q

p₀: initial condition for p

The functions f and g must have the interface

    function f(t, q, v, f)
        f[1] = ...
        f[2] = ...
        ...
    end

and

    function g(t, q, v, g)
        g[1] = ...
        g[2] = ...
        ...
    end

where t is the current time, q is the current solution vector, v is the current velocity and f and g are the vectors which hold the result of evaluating the functions $f$ and $g$ on t, q and v.

source

GeometricIntegrators.Equations.ODE — Type.

ODE: Ordinary Differential Equation

Defines an initial value problem

\[\dot{q} (t) = f(t, q(t)) , \qquad q(t_{0}) = q_{0} ,\]

with vector field $f$, initial condition $q_{0}$ and the solution $q$ taking values in $\mathbb{R}^{d}$.

Fields

d: dimension of dynamical variable $q$ and the vector field $f$

f: function computing the vector field

t₀: initial time

q₀: initial condition

The function f providing the vector field must have the interface

    function f(t, q, f)
        f[1] = ...
        f[2] = ...
        ...
    end

where t is the current time, q is the current solution vector, and f is the vector which holds the result of evaluating the vector field $f$ on t and q.

source

GeometricIntegrators.Equations.PDAE — Type.

Partitioned Differential Algebraic Equation

source

GeometricIntegrators.Equations.PODE — Type.

IODE: Partitioned Ordinary Differential Equation

Defines a partitioned initial value problem

\[\begin{align*} \dot{q} (t) &= v(t, q(t), p(t)) , & q(t_{0}) &= q_{0} , \\ \dot{p} (t) &= f(t, q(t), p(t)) , & p(t_{0}) &= p_{0} , \end{align*}\]

with vector fields $v$ and $f$, initial conditions $(q_{0}, p_{0})$ and the solution $(q,p)$ taking values in $\mathbb{R}^{d} \times \mathbb{R}^{d}$.

Fields

d: dimension of dynamical variables $q$ and $p$ as well as the vector fields $v$ and $f$

v: function computing the vector field $v$

f: function computing the vector field $f$

t₀: initial time

q₀: initial condition for q

p₀: initial condition for p

The functions v and f must have the interface

    function v(t, q, p, v)
        v[1] = ...
        v[2] = ...
        ...
    end

and

    function f(t, q, p, f)
        f[1] = ...
        f[2] = ...
        ...
    end

where t is the current time, q and p are the current solution vectors and v and f are the vectors which hold the result of evaluating the vector fields $v$ and $f$ on t, q and p.

source