Integrators

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for special additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Print error for integrators not implemented, yet.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for implicit partitioned Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for diagonally implicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for explicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for fully implicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for singly implicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for special partitioned additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for explicit partitioned Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.IntegratorDIRK — Type.

Diagonally implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorEPRK — Type.

Explicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorERK — Type.

Explicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorFIRK — Type.

Fully implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorIPRK — Type.

Implicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSARK — Type.

Special Additive Runge Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSIRK — Type.

Singly implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSPARK — Type.

Special Partitioned Additive Runge Kutta integrator.

source

GeometricIntegrators.Integrators.Solution — Type.

Create solution for implicit ODE.

source

GeometricIntegrators.Integrators.Solution — Type.

Print error for solutions of equations not implemented, yet.

source

GeometricIntegrators.Integrators.Solution — Type.

Create solution for DAE.

source

GeometricIntegrators.Integrators.Solution — Type.

Create solution for partitioned DAE.

source

GeometricIntegrators.Integrators.Solution — Type.

Create solution for partitioned ODE.

source

GeometricIntegrators.Integrators.Solution — Type.

Create solution for ODE.

source

GeometricIntegrators.Integrators.SolutionDAE — Type.

Solution of a differential algebraic equation.

source

GeometricIntegrators.Integrators.SolutionODE — Type.

SolutionODE: Solution of an ordinary differential equation

Contains all fields necessary to store the solution of an ODE.

Fields

nd: dimension of the dynamical variable $q$

nt: number of time steps to store

n0: number of initial conditions

t: time steps

x: solution x[nd, nt+1, n0] with x[:,0,:] the initial conditions

ntime: number of time steps to compute

nsave: save every nsave'th time step

source

GeometricIntegrators.Integrators.SolutionPDAE — Type.

Solution of a partitioned differential algebraic equation.

source

GeometricIntegrators.Integrators.SolutionPODE — Type.

Solution of a partitioned ordinary differential equation.

source

GeometricIntegrators.Integrators.Tableau — Type.

Holds the information for the various methods' tableaus.

source

GeometricIntegrators.Integrators.TableauDIRK — Type.

Holds the tableau of a diagonally implicit Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauEPRK — Type.

Holds the tableau of an explicit partitioned Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauERK — Type.

Holds the tableau of an explicit Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauFIRK — Type.

Holds the tableau of a fully implicit Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauGLM — Type.

Holds the tableau of a general linear method.

source

GeometricIntegrators.Integrators.TableauIPRK — Type.

Holds the tableau of a implicit partitioned Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauRK — Type.

Holds the tableau of a Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauSARK — Type.

Holds the tableau of a spezialized additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauSIRK — Type.

Holds the tableau of a singly implicit Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauSPARK — Type.

Holds the tableau of a spezialized partitioned additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.createHDF5 — Function.

Creates HDF5 file and initialises datasets for ODE solution object.

source

GeometricIntegrators.Integrators.integrate — Function.

Integrate given equation with given tableau for ntime time steps and return solution.

source

GeometricIntegrators.Integrators.integrate — Function.

Apply integrator for ntime time steps and return solution.

source

GeometricIntegrators.Integrators.integrate — Function.

Integrate ODE specified by vector field and initial condition with given tableau for ntime time steps and return solution.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE with diagonally implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate partitioned ODE with diagonally implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate partitioned ODE with explicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE with explicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate partitioned ODE with fully implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE with fully implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE with implicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate DAE with Special Additive Runge Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE with singly implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate partitioned ODE with singly implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate partitioned DAE with Special Additive Runge Kutta integrator.

source

GeometricIntegrators.Integrators.readTableauERKFromFile — Method.

Read explicit Runge-Kutta tableau from file.

source

GeometricIntegrators.Integrators.showTableau — Method.

Print Runge-Kutta tableau to standard output.

source

GeometricIntegrators.Integrators.writeSolutionToHDF5 — Function.

Append solution to HDF5 file.

source

GeometricIntegrators.Integrators.writeSolutionToHDF5 — Method.

Creates HDF5 file, writes solution to file, and closes file.

source

GeometricIntegrators.Integrators.writeTableauToFile — Method.

Write Runge-Kutta tableau to file.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of fully implicit Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of implicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersFIRK — Type.

Parameters for right-hand side function of fully implicit Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersIPRK — Type.

Parameters for right-hand side function of implicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.TableauIRK — Type.

Holds the tableau of an implicit Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauPRK — Type.

Holds the tableau of a partitioned Runge-Kutta method.

source

GeometricIntegrators.Integrators.computeStageP! — Method.

Compute P stages of explicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.computeStageQ! — Method.

Compute Q stages of explicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.getTableauERK4 — Method.

Tableau for explicit Runge-Kutta method of order four (1/6 rule)

source

GeometricIntegrators.Integrators.getTableauERK438 — Method.

Tableau for explicit Runge-Kutta method of order four (3/8 rule)

source

GeometricIntegrators.Integrators.getTableauExplicitEuler — Method.

Tableau for explicit Euler method

source

GeometricIntegrators.Integrators.getTableauExplicitMidpoint — Method.

Tableau for explicit midpoint method

source

GeometricIntegrators.Integrators.getTableauHeun — Method.

Tableau for Heun's method

source

GeometricIntegrators.Integrators.getTableauKutta — Method.

Tableau for Kutta's method of order three

source

GeometricIntegrators.Integrators.readTableauRKHeaderFromFile — Method.

Reads and parses Tableau metadata from file.

source