Methods

GeometricIntegrators.Methods.BackwardEuler — Type

Fully implicit Runge-Kutta method with TableauBackwardEuler.

GeometricIntegrators.Methods.CMDVI — Type

Midpoint Degenerate Variational Integrator.

GeometricIntegrators.Methods.CTDVI — Type

Trapezoidal Degenerate Variational Integrator.

GeometricIntegrators.Methods.CrankNicolson — Type

Diagonally implicit Runge-Kutta method with TableauCrankNicolson.

GeometricIntegrators.Methods.Crouzeix — Type

Diagonally implicit Runge-Kutta method with TableauCrouzeix.

GeometricIntegrators.Methods.DVIA — Type

Symplectic Euler-A Degenerate Variational Integrator.

GeometricIntegrators.Methods.DVIB — Type

Symplectic Euler-B Degenerate Variational Integrator.

GeometricIntegrators.Methods.ExplicitEuler — Type

Explicit Runge-Kutta method with TableauExplicitEuler.

GeometricIntegrators.Methods.ExplicitMidpoint — Type

Explicit Runge-Kutta method with TableauExplicitMidpoint.

GeometricIntegrators.Methods.ForwardEuler — Type

Explicit Runge-Kutta method with TableauForwardEuler.

GeometricIntegrators.Methods.Gauss — Type

Fully implicit Runge-Kutta method with TableauGauss.

GeometricIntegrators.Methods.Heun2 — Type

Explicit Runge-Kutta method with TableauHeun2.

GeometricIntegrators.Methods.Heun3 — Type

Explicit Runge-Kutta method with TableauHeun3.

GeometricIntegrators.Methods.ImplicitEuler — Type

Fully implicit Runge-Kutta method with TableauImplicitEuler.

GeometricIntegrators.Methods.ImplicitMidpoint — Type

Fully implicit Runge-Kutta method with TableauImplicitMidpoint.

GeometricIntegrators.Methods.KraaijevangerSpijker — Type

Diagonally implicit Runge-Kutta method with TableauKraaijevangerSpijker.

GeometricIntegrators.Methods.Kutta3 — Type

Explicit Runge-Kutta method with TableauKutta3.

GeometricIntegrators.Methods.LieA — Type

Lie-Trotter Splitting A

For a vector field $\dot{x} = f_1 (t,x) + f_2 (t,x)$, the splitting reads

\[\Phi_{\Delta t} = \varphi^{2}_{\Delta t} \circ \varphi^{1}_{\Delta t}\]

Reference:

H. F. Trotter.
On the product of semi-groups of operators.
Proceedings of the American Mathematical Society, Volume 10, Pages 545-551, 1959.
doi: 10.1090/S0002-9939-1959-0108732-6.

GeometricIntegrators.Methods.LieB — Type

Lie-Trotter Splitting B

For a vector field $\dot{x} = f_1 (t,x) + f_2 (t,x)$, the splitting reads

\[\Phi_{\Delta t} = \varphi^{1}_{\Delta t} \circ \varphi^{2}_{\Delta t}\]

Reference:

H. F. Trotter.
On the product of semi-groups of operators.
Proceedings of the American Mathematical Society, Volume 10, Pages 545-551, 1959.
doi: 10.1090/S0002-9939-1959-0108732-6.

GeometricIntegrators.Methods.LobattoIIIA — Type

Runge-Kutta method with TableauLobattoIIIA.

GeometricIntegrators.Methods.LobattoIIIAIIIB — Type

Partitioned Runge-Kutta method with TableauLobattoIIIA for $q$ and TableauLobattoIIIB for $p$.

GeometricIntegrators.Methods.LobattoIIIAIIIĀ — Type

Partitioned Runge-Kutta method with TableauLobattoIIIA for $q$ and TableauLobattoIIIĀ for $p$.

GeometricIntegrators.Methods.LobattoIIIB — Type

Runge-Kutta method with TableauLobattoIIIB.

GeometricIntegrators.Methods.LobattoIIIBIIIA — Type

Partitioned Runge-Kutta method with TableauLobattoIIIB for $q$ and TableauLobattoIIIA for $p$.

GeometricIntegrators.Methods.LobattoIIIBIIIB̄ — Type

Partitioned Runge-Kutta method with TableauLobattoIIIB for $q$ and TableauLobattoIIIB̄ for $p$.

GeometricIntegrators.Methods.LobattoIIIC — Type

Runge-Kutta method with TableauLobattoIIIC.

GeometricIntegrators.Methods.LobattoIIICIIIC̄ — Type

Partitioned Runge-Kutta method with TableauLobattoIIIC for $q$ and TableauLobattoIIIC̄ for $p$.

GeometricIntegrators.Methods.LobattoIIIC̄ — Type

Runge-Kutta method with TableauLobattoIIIC̄.

GeometricIntegrators.Methods.LobattoIIIC̄IIIC — Type

Partitioned Runge-Kutta method with TableauLobattoIIIC̄ for $q$ and TableauLobattoIIIC for $p$.

GeometricIntegrators.Methods.LobattoIIID — Type

Runge-Kutta method with TableauLobattoIIID.

GeometricIntegrators.Methods.LobattoIIIDIIID̄ — Type

Partitioned Runge-Kutta method with TableauLobattoIIID for $q$ and TableauLobattoIIID̄ for $p$.

GeometricIntegrators.Methods.LobattoIIIE — Type

Runge-Kutta method with TableauLobattoIIIE.

GeometricIntegrators.Methods.LobattoIIIEIIIĒ — Type

Partitioned Runge-Kutta method with TableauLobattoIIIE for $q$ and TableauLobattoIIIĒ for $p$.

GeometricIntegrators.Methods.LobattoIIIF — Type

Runge-Kutta method with TableauLobattoIIIF.

GeometricIntegrators.Methods.LobattoIIIFIIIF̄ — Type

Partitioned Runge-Kutta method with TableauLobattoIIIF for $q$ and TableauLobattoIIIF̄ for $p$.

GeometricIntegrators.Methods.LobattoIIIF̄ — Type

Runge-Kutta method with TableauLobattoIIIF̄.

GeometricIntegrators.Methods.LobattoIIIF̄IIIF — Type

Partitioned Runge-Kutta method with TableauLobattoIIIF̄ for $q$ and TableauLobattoIIIF for $p$.

GeometricIntegrators.Methods.LobattoIIIG — Type

Runge-Kutta method with TableauLobattoIIIG.

GeometricIntegrators.Methods.LobattoIIIGIIIḠ — Type

Partitioned Runge-Kutta method with TableauLobattoIIIG for $q$ and TableauLobattoIIIḠ for $p$.

GeometricIntegrators.Methods.Marchuk — Type

Alias for Strang

GeometricIntegrators.Methods.McLachlan2 — Type

McLachlan's 2nd order symmetric, minimum error composition method

The composition reads

\[\Phi_{\Delta t} = \varphi_{\alpha \Delta t} \circ \varphi^{*}_{(1/2 - \alpha) \Delta t} \circ \varphi_{(1/2 - \alpha) \Delta t} \circ \varphi^{*}_{\alpha \Delta t} ,\]

where the parameter $\alpha$ can be optimized, e.g., to minimize the solution error. McLachlan arrives at $\alpha = 0.1932$ as a generally useful value.

Reference:

Robert I. McLachlan.
On the Numerical Integration of Ordinary Differential Equations by Symmetric Composition Methods
SIAM Journal on Scientific Computing, Volume 16, Pages 151-168, 1995.
doi: 10.1137/0916010.

GeometricIntegrators.Methods.McLachlan4 — Type

McLachlan's 4th order symmetric, minimum error composition method

The composition reads

\[\Phi_{\Delta t} = \varphi_{\alpha_5 \Delta t} \circ \varphi^{*}_{\beta_5 \Delta t} \circ \dotsc \circ \varphi_{\alpha_2 \Delta t} \circ \varphi^{*}_{\beta_2 \Delta t} \circ \varphi_{\alpha_1 \Delta t} \circ \varphi^{*}_{\beta_1 \Delta t} ,\]

with

\[\begin{aligned} \beta_1 &= \alpha_5 = \frac{14 - \sqrt{19}}{108} , & \alpha_1 &= \beta_5 = \frac{146 + 5 \sqrt{19}}{540} , & & \\ \beta_2 &= \alpha_4 = \frac{- 23 - 20 \sqrt{19}}{270} , & \alpha_2 &= \beta_4 = \frac{-2 + 10 \sqrt{19}}{135} , & \beta_3 &= \alpha_3 = \frac{1}{5} . \end{aligned}\]

The coefficients are optimised to provide an integrator with minimal solution error.

Reference:

Robert I. McLachlan.
On the Numerical Integration of Ordinary Differential Equations by Symmetric Composition Methods
SIAM Journal on Scientific Computing, Volume 16, Pages 151-168, 1995.
doi: 10.1137/0916010.

GeometricIntegrators.Methods.PRK — Type

Partitioned Runge-Kutta Method

PRK(tableau)

GeometricIntegrators.Methods.ProjectedVPRK — Type

Projected VPRK Method

A projected Variational Runge-Kutta Method consists of a VPRK method and a projection method:

ProjectedVPRK(vprk::VPRKMethod, projection::ProjectionMethod)
ProjectedVPRK(tableau::AbstractTableau, projection::ProjectionMethod, args...; kwargs...) = ProjectedVPRK(VPRK(tableau, args...; kwargs...), projection)

GeometricIntegrators.Methods.QinZhang — Type

Diagonally implicit Runge-Kutta method with TableauQinZhang.

GeometricIntegrators.Methods.RK — Type

General Runge-Kutta Method

RK(tableau)

GeometricIntegrators.Methods.RK4 — Type

Explicit Runge-Kutta method with TableauRK4.

GeometricIntegrators.Methods.RK416 — Type

Explicit Runge-Kutta method with TableauRK416.

GeometricIntegrators.Methods.RK438 — Type

Explicit Runge-Kutta method with TableauRK438.

GeometricIntegrators.Methods.RadauIA — Type

Runge-Kutta method with TableauRadauIA.

GeometricIntegrators.Methods.RadauIB — Type

Runge-Kutta method with TableauRadauIB.

GeometricIntegrators.Methods.RadauIIA — Type

Runge-Kutta method with TableauRadauIIA.

GeometricIntegrators.Methods.RadauIIB — Type

Runge-Kutta method with TableauRadauIIB.

GeometricIntegrators.Methods.Ralston2 — Type

Explicit Runge-Kutta method with TableauRalston2.

GeometricIntegrators.Methods.Ralston3 — Type

Explicit Runge-Kutta method with TableauRalston3.

GeometricIntegrators.Methods.Runge2 — Type

Explicit Runge-Kutta method with TableauRunge2.

GeometricIntegrators.Methods.SRK3 — Type

Fully implicit Runge-Kutta method with TableauSRK3.

GeometricIntegrators.Methods.SSPRK2 — Type

Explicit Runge-Kutta method with TableauSSPRK2.

GeometricIntegrators.Methods.SSPRK3 — Type

Explicit Runge-Kutta method with TableauSSPRK3.

GeometricIntegrators.Methods.Strang — Type

Strang Splitting

For a vector field $\dot{x} = f_1 (t,x) + f_2 (t,x)$, the splitting reads

\[\Phi_{\Delta t} = \varphi^{1}_{\Delta t / 2} \circ \varphi^{2}_{\Delta t / 2} \circ \varphi^{2}_{\Delta t / 2} \circ \varphi^{1}_{\Delta t / 2}\]

For vector fields with two components, this is not the most efficient implementation. For such cases StrangA or StrangB should be used instead.

References:

Gilbert Strang.
On the construction and comparison of difference schemes.
SIAM Journal on Numerical Analysis, Volume 5, Pages 506-517, 1968.
doi: 10.1137/0705041.

Gurij Ivanovich Marchuk.
Some applications of splitting-up methods to the solution of mathematical physics problems.
Aplikace Matematiky, Volume 13, Pages 103-132, 1968.
doi: 10.21136/AM.1968.103142.

GeometricIntegrators.Methods.StrangA — Type

Strang Splitting A for a vector field $\dot{x} = f_1 (t,x) + f_2 (t,x)$.

The splitting reads

\[\Phi_{\Delta t} = \varphi^{1}_{\Delta t / 2} \circ \varphi^{2}_{\Delta t} \circ \varphi^{1}_{\Delta t / 2}\]

References:

Gilbert Strang.
On the construction and comparison of difference schemes.
SIAM Journal on Numerical Analysis, Volume 5, Pages 506-517, 1968.
doi: 10.1137/0705041.

Gurij Ivanovich Marchuk.
Some applications of splitting-up methods to the solution of mathematical physics problems.
Aplikace Matematiky, Volume 13, Pages 103-132, 1968.
doi: 10.21136/AM.1968.103142.

GeometricIntegrators.Methods.StrangB — Type

Strang Splitting B for a vector field $\dot{x} = f_1 (t,x) + f_2 (t,x)$

The splitting reads

\[\Phi_{\Delta t} = \varphi^{2}_{\Delta t / 2} \circ \varphi^{1}_{\Delta t} \circ \varphi^{2}_{\Delta t / 2}\]

References:

Gilbert Strang.
On the construction and comparison of difference schemes.
SIAM Journal on Numerical Analysis, Volume 5, Pages 506-517, 1968.
doi: 10.1137/0705041.

Gurij Ivanovich Marchuk.
Some applications of splitting-up methods to the solution of mathematical physics problems.
Aplikace Matematiky, Volume 13, Pages 103-132, 1968.
doi: 10.21136/AM.1968.103142.

GeometricIntegrators.Methods.SuzukiFractal — Type

Suzuki's 4th order "fractal" composition method

The composition reads

\[\Phi_{\Delta t} = \varphi_{\gamma_5 \Delta t} \circ \varphi_{\gamma_4 \Delta t} \circ \varphi_{\gamma_3 \Delta t} \circ \varphi_{\gamma_2 \Delta t} \circ \varphi_{\gamma_1 \Delta t}\]

with

\[\gamma_1 = \gamma_2 = \gamma_4 = \gamma_5 = \frac{1}{4 - 4^{1/(p+1)}} , \qquad \gamma_3 = - \frac{4^{1/(p+1)}}{4 - 4^{1/(p+1)}} .\]

Reference:

Masuo Suzuki
Fractal decomposition of exponential operators with applications to many-body theories and Monte Carlo simulations.
Physics Letters A, Volume 146, Pages 319-323, 1990.
doi: 10.1016/0375-9601(90)90962-N

GeometricIntegrators.Methods.TripleJump — Type

4th order "Triple Jump" composition method.

The composition reads

\[\Phi_{\Delta t} = \varphi_{\gamma_3 \Delta t} \circ \varphi_{\gamma_2 \Delta t} \circ \varphi_{\gamma_1 \Delta t}\]

with

\[\gamma_1 = \gamma_3 = \frac{1}{2 - 2^{1/(p+1)}} , \qquad \gamma_2 = - \frac{2^{1/(p+1)}}{2 - 2^{1/(p+1)}} .\]

References:

Michael Creutz and Andreas Gocksch.
Higher-order hybrid Monte Carlo algorithms.
Physical Review Letters, Volume 63, Pages 9-12, 1989.
doi: 10.1103/PhysRevLett.63.9.

Etienne Forest.
Canonical integrators as tracking codes (or how to integrate perturbation theory with tracking).
AIP Conference Proceedings, Volume 184, Pages 1106-1136, 1989.
doi: 10.1063/1.38062.

Masuo Suzuki
Fractal decomposition of exponential operators with applications to many-body theories and Monte Carlo simulations.
Physics Letters A, Volume 146, Pages 319-323, 1990.
doi: 10.1016/0375-9601(90)90962-N

Haruo Yoshida.
Construction of higher order symplectic integrators.
Physics Letters A, Volume 150, Pages 262-268, 1990.
doi: 10.1016/0375-9601(90)90092-3

GeometricIntegrators.Methods.VPRK — Type

Variational Partitioned Runge-Kutta Method

VPRK(tableau::PartitionedTableau, d=nothing)
VPRK(tableau::Tableau, args...; kwargs...) = VPRK(PartitionedTableau(tableau), args...; kwargs...)
VPRK(tableau1::Tableau, tableau2::Tableau, args...; kwargs...) = VPRK(PartitionedTableau(tableau1, tableau2), args...; kwargs...)

GeometricIntegrators.Methods.VPRKGauss — Type

Variational Partitioned Runge-Kutta Method that uses Gauss-Legendre coefficients for both $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIA — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIA for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIAIIIB — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIA for the coefficients $a_{ij}$ and TableauLobattoIIIB for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIAIIIĀ — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIA for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIB — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIB for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIBIIIA — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIB for the coefficients $a_{ij}$ and TableauLobattoIIIA for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIBIIIB̄ — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIB for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIC — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIC for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIICIIIC̄ — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIC for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIC̄ — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIC̄ for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIC̄IIIC — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIC̄ for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIID — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIID for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIDIIID̄ — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIID for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIE — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIE for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIEIIIĒ — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIE for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIF — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIF for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIFIIIF̄ — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIF for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIF̄ — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIF̄ for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIF̄IIIF — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIF̄ for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIG — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIG for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIGIIIḠ — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIG for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKRadauIIA — Type

Variational Partitioned Runge-Kutta Method that uses TableauRadauIIA for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKRadauIIB — Type

Variational Partitioned Runge-Kutta Method that uses TableauRadauIIB for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPSRK3 — Type

Variational Partitioned Runge-Kutta Method that uses the symmetric TableauSRK3 coefficients with 3 stages for both $a_{ij}$ and $\bar{a}_{ij}$.

Base.show — Method

Base.show(io::IO, ml::MethodList)
Base.show(io::IO, ::MIME"text/markdown", ml::MethodList)

Pretty-print MethodList.

GeometricIntegrators.Methods.VPRKpInternal — Method

VPRK integrator with projection on internal stages

GeometricIntegrators.Methods.VPRKpLegendre — Method

VPRK integrator with Legendre projection

GeometricIntegrators.Methods.VPRKpMidpoint — Method

VPRK integrator with Midpoint projection

GeometricIntegrators.Methods.VPRKpSecondary — Method

VPRK integrator with projection on secondary constraint

GeometricIntegrators.Methods.VPRKpStandard — Method

VPRK integrator with standard projection

GeometricIntegrators.Methods.VPRKpSymmetric — Method

VPRK integrator with symmetric projection

GeometricIntegrators.Methods.VPRKpSymplectic — Method

VPRK integrator with symplectic projection

GeometricIntegrators.Methods.VPRKpVariational — Method

VPRK integrator with variational projection

GeometricIntegrators.Methods.VPRKpVariationalP — Method

VPRK integrator with variational projection on P

GeometricIntegrators.Methods.VPRKpVariationalQ — Method

VPRK integrator with variational projection on Q