Projections

Projection Methods

GeometricIntegrators.Integrators.ProjectedMethod — Type

A ProjectedMethod consists of a ProjectionMethod and a GeometricMethod.

GeometricIntegrators.Integrators.ProjectionIntegrator — Type

The ProjectionIntegrator is the counterpart to the GeometricIntegrator for ProjectionMethods.

GeometricIntegrators.Integrators.ProjectionMethod — Type

A ProjectionMethod is an algorithm that is applied together with a geometric integrator to enforce constraints which are not automatically satisfied by the integrator. Examples include conservation of invariants like energy or the Dirac constraint in IODEs.

Projected VPRK Methods

GeometricIntegrators.Integrators.VPRKpInternal — Method

VPRK integrator with projection on internal stages

GeometricIntegrators.Integrators.VPRKpLegendre — Method

VPRK integrator with Legendre projection

GeometricIntegrators.Integrators.VPRKpMidpoint — Method

VPRK integrator with Midpoint projection

GeometricIntegrators.Integrators.VPRKpSecondary — Method

VPRK integrator with projection on secondary constraint

GeometricIntegrators.Integrators.VPRKpStandard — Method

VPRK integrator with standard projection

GeometricIntegrators.Integrators.VPRKpSymmetric — Method

VPRK integrator with symmetric projection

GeometricIntegrators.Integrators.VPRKpSymplectic — Method

VPRK integrator with symplectic projection

GeometricIntegrators.Integrators.VPRKpVariational — Method

VPRK integrator with variational projection

GeometricIntegrators.Integrators.VPRKpVariationalP — Method

VPRK integrator with variational projection on P

GeometricIntegrators.Integrators.VPRKpVariationalQ — Method

VPRK integrator with variational projection on Q