Partitioned Runge-Kutta Tableaus

The following tableaus are all implemented in RungeKutta.jl.

RungeKutta.PartitionedTableau — Type

Tableau of a Partitioned Runge-Kutta method

\[\begin{aligned} Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, v(t_{n} + c_j \Delta t, Q_{n,j}, P_{n,j}) , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, v(t_{n} + c_j \Delta t, Q_{n,i}, P_{n,i}) , \\ P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, f(t_{n} + c_j \Delta t, Q_{n,j}, P_{n,j}) , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, f(t_{n} + c_j \Delta t, Q_{n,i}, P_{n,i}) . \end{aligned}\]

Parameters:

T: datatype of coefficient arrays

Fields:

name: symbolic name of the tableau

o: order of the method

s: number of stages

q: Tableau for q

p: Tableau for p

R∞: stability function at infinity

The actual tableaus are stored in q and p:

a: coefficients $a_{ij}$ with $ 1 \le i,j \le s$

b: weights $b_{i}$ with $ 1 \le i \le s$

c: nodes $c_{i}$ with $ 1 \le i \le s$

Constructors:

PartitionedTableau{T}(name, o, q, p)
PartitionedTableau{T}(name, q, p)
PartitionedTableau(name::Symbol, q::Tableau, p::Tableau)
PartitionedTableau(name::Symbol, q::Tableau)

RungeKutta.PartitionedTableaus.PartitionedTableauGauss — Method

Partitioned Gauss-Legendre Runge-Kutta tableau with s stages

PartitionedTableauGauss(::Type{T}, s)
PartitionedTableauGauss(s) = PartitionedTableauGauss(Float64, s)

The constructor takes the number of stages s and optionally the element type T of the tableau.

This PartitionedTableau uses TableauGauss for both coefficients a and ā.

RungeKutta.PartitionedTableaus.TableauLobattoIIIAIIIB — Method

Partitioned Gauss-Lobatto IIIA-IIIB tableau with s stages

TableauLobattoIIIAIIIB(::Type{T}, s)
TableauLobattoIIIAIIIB(s) = TableauLobattoIIIAIIIB(Float64, s)

The constructor takes the number of stages s and optionally the element type T of the tableau.

This PartitionedTableau uses TableauLobattoIIIA for a and TableauLobattoIIIB for ā.

RungeKutta.PartitionedTableaus.TableauLobattoIIIAIIIĀ — Method

Tableau for Gauss-Lobatto IIIA-IIIĀ method with s stages

TableauLobattoIIIAIIIĀ(::Type{T}, s)
TableauLobattoIIIAIIIĀ(s) = TableauLobattoIIIAIIIĀ(Float64, s)

The constructor takes the number of stages s and optionally the element type T of the tableau.

This PartitionedTableau uses TableauLobattoIIIA for a and TableauLobattoIIIĀ for ā.

RungeKutta.PartitionedTableaus.TableauLobattoIIIBIIIA — Method

Tableau for Gauss-Lobatto IIIB-IIIA method with s stages

TableauLobattoIIIBIIIA(::Type{T}, s)
TableauLobattoIIIBIIIA(s) = TableauLobattoIIIBIIIA(Float64, s)

The constructor takes the number of stages s and optionally the element type T of the tableau.

This PartitionedTableau uses TableauLobattoIIIB for a and TableauLobattoIIIA for ā.

RungeKutta.PartitionedTableaus.TableauLobattoIIIBIIIB̄ — Method

Tableau for Gauss-Lobatto IIIB-IIIB̄ method with s stages

TableauLobattoIIIBIIIB̄(::Type{T}, s)
TableauLobattoIIIBIIIB̄(s) = TableauLobattoIIIBIIIB̄(Float64, s)

The constructor takes the number of stages s and optionally the element type T of the tableau.

This PartitionedTableau uses TableauLobattoIIIB for a and TableauLobattoIIIB̄ for ā.

RungeKutta.PartitionedTableaus.TableauLobattoIIICIIIC̄ — Method

Tableau for Gauss-Lobatto IIIC-IIIC̄ method with s stages

TableauLobattoIIICIIIC̄(::Type{T}, s)
TableauLobattoIIICIIIC̄(s) = TableauLobattoIIICIIIC̄(Float64, s)

The constructor takes the number of stages s and optionally the element type T of the tableau.

This PartitionedTableau uses TableauLobattoIIIC for a and TableauLobattoIIIC̄ for ā.

RungeKutta.PartitionedTableaus.TableauLobattoIIIC̄IIIC — Method

Tableau for Gauss-Lobatto IIIC̄-IIIC method with s stages

TableauLobattoIIIC̄IIIC(::Type{T}, s)
TableauLobattoIIIC̄IIIC(s) = TableauLobattoIIIC̄IIIC(Float64, s)

The constructor takes the number of stages s and optionally the element type T of the tableau.

This PartitionedTableau uses TableauLobattoIIIC̄ for a and TableauLobattoIIIC for ā.

RungeKutta.PartitionedTableaus.TableauLobattoIIIDIIID̄ — Method

Tableau for Gauss-Lobatto IIID-IIID̄ method with s stages

TableauLobattoIIIDIIID̄(::Type{T}, s)
TableauLobattoIIIDIIID̄(s) = TableauLobattoIIIDIIID̄(Float64, s)

The constructor takes the number of stages s and optionally the element type T of the tableau.

This PartitionedTableau uses TableauLobattoIIID for a and TableauLobattoIIID̄ for ā.

RungeKutta.PartitionedTableaus.TableauLobattoIIIEIIIĒ — Method

Tableau for Gauss-Lobatto IIIE-IIIĒ method with s stages

TableauLobattoIIIEIIIĒ(::Type{T}, s)
TableauLobattoIIIEIIIĒ(s) = TableauLobattoIIIEIIIĒ(Float64, s)

The constructor takes the number of stages s and optionally the element type T of the tableau.

This PartitionedTableau uses TableauLobattoIIIE for a and TableauLobattoIIIĒ for ā.

RungeKutta.PartitionedTableaus.TableauLobattoIIIFIIIF̄ — Method

Tableau for Gauss-Lobatto IIIF-IIIF̄ method with s stages

TableauLobattoIIIFIIIF̄(::Type{T}, s)
TableauLobattoIIIFIIIF̄(s) = TableauLobattoIIIFIIIF̄(Float64, s)

The constructor takes the number of stages s and optionally the element type T of the tableau.

This PartitionedTableau uses TableauLobattoIIIF for a and TableauLobattoIIIF̄ for ā.

RungeKutta.PartitionedTableaus.TableauLobattoIIIF̄IIIF — Method

Tableau for Gauss-Lobatto IIIF̄-IIIF method with s stages

TableauLobattoIIIF̄IIIF(::Type{T}, s)
TableauLobattoIIIF̄IIIF(s) = TableauLobattoIIIF̄IIIF(Float64, s)

The constructor takes the number of stages s and optionally the element type T of the tableau.

This PartitionedTableau uses TableauLobattoIIIF̄ for a and TableauLobattoIIIF for ā.

RungeKutta.PartitionedTableaus.TableauLobattoIIIGIIIḠ — Method

Tableau for Gauss-Lobatto IIIG-IIIḠ method with s stages

TableauLobattoIIIGIIIḠ(::Type{T}, s)
TableauLobattoIIIGIIIḠ(s) = TableauLobattoIIIGIIIḠ(Float64, s)

The constructor takes the number of stages s and optionally the element type T of the tableau.

This PartitionedTableau uses TableauLobattoIIIG for a and TableauLobattoIIIḠ for ā.