Tableaus

Tableaus

GeometricIntegrators.Tableaus.AbstractTableau — Type.

Holds the information for the various methods' tableaus.

source

GeometricIntegrators.Tableaus.AbstractTableauIRK — Type.

Holds the tableau of an implicit Runge-Kutta method.

source

GeometricIntegrators.Tableaus.AbstractTableauPRK — Type.

Holds the tableau of a partitioned Runge-Kutta method.

source

GeometricIntegrators.Tableaus.AbstractTableauRK — Type.

Holds the tableau of a Runge-Kutta method.

source

GeometricIntegrators.Tableaus.CoefficientsARK — Type.

Holds the coefficients of an additive Runge-Kutta method.

source

GeometricIntegrators.Tableaus.CoefficientsMRK — Type.

Holds the multiplier Runge-Kutta coefficients.

source

GeometricIntegrators.Tableaus.CoefficientsPRK — Type.

Holds the coefficients of a projective Runge-Kutta method.

source

GeometricIntegrators.Tableaus.CoefficientsRK — Type.

Holds the coefficients of a Runge-Kutta method.

source

GeometricIntegrators.Tableaus.TableauARK — Type.

Holds the tableau of a additive Runge-Kutta method.

source

GeometricIntegrators.Tableaus.TableauDIRK — Type.

Holds the tableau of a diagonally implicit Runge-Kutta method.

source

GeometricIntegrators.Tableaus.TableauEPRK — Type.

TableauEPRK: Tableau of an Explicit Partitioned Runge-Kutta method

\[\begin{align*} V_{n,i} &= \hphantom{-} \dfrac{\partial H}{\partial p} (Q_{n,i}, P_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} , \\ F_{k,i} &= - \dfrac{\partial H}{\partial q} (Q_{n,i}, P_{n,i}) , & P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, F_{n,i} , \end{align*}\]

usually satisfying the symplecticity conditions

\[\begin{align*} b_{i} \bar{a}_{ij} + b_{j} a_{ji} &= b_{i} b_{j} , & \bar{b}_i &= b_i . \end{align*}\]

source

GeometricIntegrators.Tableaus.TableauERK — Type.

Holds the tableau of an explicit Runge-Kutta method.

source

GeometricIntegrators.Tableaus.TableauFIRK — Type.

Holds the tableau of a fully implicit Runge-Kutta method.

source

GeometricIntegrators.Tableaus.TableauGLM — Type.

Holds the tableau of a general linear method.

source

GeometricIntegrators.Tableaus.TableauIPRK — Type.

TableauIPRK: Tableau of an Implicit Partitioned Runge-Kutta method

\[\begin{align*} P_{n,i} &= \dfrac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} , \\ F_{k,i} &= \dfrac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , & P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, F_{n,i} , \end{align*}\]

usually satisfying the symplecticity conditions

\[\begin{align*} b_{i} \bar{a}_{ij} + b_{j} a_{ji} &= b_{i} b_{j} , & \bar{b}_i &= b_i . \end{align*}\]

source

GeometricIntegrators.Tableaus.TableauPARK — Type.

Holds the tableau of an partitioned additive Runge-Kutta method.

source

GeometricIntegrators.Tableaus.TableauSARK — Type.

Holds the tableau of a spezialized additive Runge-Kutta method.

source

GeometricIntegrators.Tableaus.TableauSIRK — Type.

Holds the tableau of a singly implicit Runge-Kutta method.

source

GeometricIntegrators.Tableaus.TableauSPARK — Type.

Holds the tableau of a spezialized partitioned additive Runge-Kutta method.

source

GeometricIntegrators.Tableaus.TableauVPARK — Type.

Holds the tableau of an variational partitioned additive Runge-Kutta method.

source

GeometricIntegrators.Tableaus.TableauVPRK — Type.

TableauVPRK: Tableau of a Variational Partitioned Runge-Kutta method

\[\begin{align*} P_{n,i} &= \dfrac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} , \\ F_{k,i} &= \dfrac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , & P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} - d_i \lambda , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, F_{n,i} , \\ && 0 &= \sum \limits_{i=1}^{s} d_i V_i , && \end{align*}\]

satisfying the symplecticity conditions

\[\begin{align*} b_{i} \bar{a}_{ij} + b_{j} a_{ji} &= b_{i} b_{j} , & \bar{b}_i &= b_i . \end{align*}\]

source

GeometricIntegrators.Tableaus.TableauVSPARK — Type.

Holds the tableau of an variational special partitioned additive Runge-Kutta method.

source

GeometricIntegrators.Tableaus.getTableauERK4 — Method.

Tableau for explicit Runge-Kutta method of order four (1/6 rule)

source

GeometricIntegrators.Tableaus.getTableauERK438 — Method.

Tableau for explicit Runge-Kutta method of order four (3/8 rule)

source

GeometricIntegrators.Tableaus.getTableauExplicitEuler — Method.

Tableau for explicit Euler method

source

GeometricIntegrators.Tableaus.getTableauExplicitMidpoint — Method.

Tableau for explicit midpoint method

source

GeometricIntegrators.Tableaus.getTableauGLRKpSymmetric — Method.

Tableau for Gauss-Legendre method with s stages and symplectic projection.

source

GeometricIntegrators.Tableaus.getTableauGLRKpSymplectic — Method.

Tableau for Gauss-Legendre method with s stages and symplectic projection.

source

GeometricIntegrators.Tableaus.getTableauHeun — Method.

Tableau for Heun's method

source

GeometricIntegrators.Tableaus.getTableauImplicitEuler — Method.

Implicit Euler

source

GeometricIntegrators.Tableaus.getTableauImplicitMidpoint — Method.

Implicit Midpoint

source

GeometricIntegrators.Tableaus.getTableauKutta — Method.

Tableau for Kutta's method of order three

source

GeometricIntegrators.Tableaus.getTableauLobIIIA2 — Method.

Gauss-Lobatto-IIIA Runge-Kutta, s=2

source

GeometricIntegrators.Tableaus.getTableauLobIIIA3 — Method.

Gauss-Lobatto-IIIA Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauLobIIIA4 — Method.

Gauss-Lobatto-IIIA Runge-Kutta, s=4

source

GeometricIntegrators.Tableaus.getTableauLobIIIAIIIB2pSymmetric — Method.

Tableau for Gauss-Lobatto IIIA-IIIB method with two stages and symmetric projection.

source

GeometricIntegrators.Tableaus.getTableauLobIIIAIIIB2pSymplectic — Method.

Tableau for Gauss-Lobatto IIIA-IIIB method with two stages and symplectic projection.

source

GeometricIntegrators.Tableaus.getTableauLobIIIAIIIB3pSymmetric — Method.

Tableau for Gauss-Lobatto IIIA-IIIB method with three stages and symmetric projection.

source

GeometricIntegrators.Tableaus.getTableauLobIIIAIIIB3pSymplectic — Method.

Tableau for Gauss-Lobatto IIIA-IIIB method with three stages and symplectic projection.

source

GeometricIntegrators.Tableaus.getTableauLobIIIB2 — Method.

Gauss-Lobatto-IIIB Runge-Kutta, s=2

source

GeometricIntegrators.Tableaus.getTableauLobIIIB3 — Method.

Gauss-Lobatto-IIIB Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauLobIIIB4 — Method.

Gauss-Lobatto-IIIB Runge-Kutta, s=4

source

GeometricIntegrators.Tableaus.getTableauLobIIIC2 — Method.

Gauss-Lobatto-IIIC Runge-Kutta, s=2

source

GeometricIntegrators.Tableaus.getTableauLobIIIC3 — Method.

Gauss-Lobatto-IIIC Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauLobIIIC4 — Method.

Gauss-Lobatto-IIIC Runge-Kutta, s=4

source

GeometricIntegrators.Tableaus.getTableauLobIIID2 — Method.

Gauss-Lobatto-IIID Runge-Kutta, s=2

source

GeometricIntegrators.Tableaus.getTableauLobIIID3 — Method.

Gauss-Lobatto-IIID Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauLobIIID4 — Method.

Gauss-Lobatto-IIID Runge-Kutta, s=4

source

GeometricIntegrators.Tableaus.getTableauLobIIIE2 — Method.

Gauss-Lobatto-IIIE Runge-Kutta, s=2

source

GeometricIntegrators.Tableaus.getTableauLobIIIE3 — Method.

Gauss-Lobatto-IIIE Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauLobIIIE4 — Method.

Gauss-Lobatto-IIIE Runge-Kutta, s=4

source

GeometricIntegrators.Tableaus.getTableauLobIIIF2 — Method.

Gauss-Lobatto-IIIF Runge-Kutta, s=2

source

GeometricIntegrators.Tableaus.getTableauLobIIIF3 — Method.

Gauss-Lobatto-IIIF Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauLobIIIF4 — Method.

Gauss-Lobatto-IIIF Runge-Kutta, s=4

source

GeometricIntegrators.Tableaus.getTableauLobIIIG2 — Method.

Gauss-Lobatto-IIIG Runge-Kutta, s=2

source

GeometricIntegrators.Tableaus.getTableauLobIIIG3 — Method.

Gauss-Lobatto-IIIG Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauLobIIIG4 — Method.

Gauss-Lobatto-IIIG Runge-Kutta, s=4

source

GeometricIntegrators.Tableaus.getTableauRadIIA2 — Method.

Gauss-Radau-IIA Runge-Kutta, s=2

source

GeometricIntegrators.Tableaus.getTableauRadIIA3 — Method.

Gauss-Radau-IIA Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauSRK3 — Method.

Gauss-Legendre Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauSymplecticEulerA — Method.

Tableau for symplectic Euler-A method

source

GeometricIntegrators.Tableaus.getTableauSymplecticEulerB — Method.

Tableau for symplectic Euler-B method

source

GeometricIntegrators.Tableaus.getTableauVPGLRK — Method.

Tableau for variational Gauss-Legendre method with s stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIAIIIB2 — Method.

Tableau for variational Gauss-Lobatto IIIA-IIIB method with two stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIAIIIB3 — Method.

Tableau for variational Gauss-Lobatto IIIA-IIIB method with three stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIAIIIB4 — Method.

Tableau for variational Gauss-Lobatto IIIA-IIIB method with four stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIC2 — Method.

Tableau for variational Gauss-Lobatto IIIC-III method with two stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIC3 — Method.

Tableau for variational Gauss-Lobatto IIIC-III method with three stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIC4 — Method.

Tableau for variational Gauss-Lobatto IIIC-III method with four stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIID2 — Method.

Tableau for variational Gauss-Lobatto IIID method with two stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIID3 — Method.

Tableau for variational Gauss-Lobatto IIID method with three stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIID4 — Method.

Tableau for variational Gauss-Lobatto IIID method with four stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIE2 — Method.

Tableau for variational Gauss-Lobatto IIIE method with two stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIE3 — Method.

Tableau for variational Gauss-Lobatto IIIE method with three stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIE4 — Method.

Tableau for variational Gauss-Lobatto IIIE method with four stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIF2 — Method.

Tableau for variational Gauss-Lobatto IIIF method with two stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIF3 — Method.

Tableau for variational Gauss-Lobatto IIIF method with three stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIF4 — Method.

Tableau for variational Gauss-Lobatto IIIF method with four stages

source

GeometricIntegrators.Tableaus.getTableauVPSRK3 — Method.

Tableau for variational symmetric Runge-Kutta method with 3 stages

source

GeometricIntegrators.Tableaus.readTableauERKFromFile — Method.

Read explicit Runge-Kutta tableau from file.

source

GeometricIntegrators.Tableaus.writeTableauToFile — Method.

Write Runge-Kutta tableau to file.

source

GeometricIntegrators.Tableaus.CoefficientsPGLRK — Type.

Holds the coefficients of a projected Gauss-Legendre Runge-Kutta method.

source

Base.show — Method.

Print additive Runge-Kutta coefficients.

source

Base.show — Method.

Print multiplier Runge-Kutta coefficients.

source

Base.show — Method.

Print Runge-Kutta coefficients.

source

Base.show — Method.

Print projective Runge-Kutta coefficients.

source

Base.show — Method.

Print Runge-Kutta coefficients.

source

GeometricIntegrators.Tableaus.readTableauRKHeaderFromFile — Method.

Reads and parses Tableau metadata from file.

source