Partitioned Runge-Kutta Tableaus

RungeKutta.PartitionedTableaus.PartitionedTableauGauss — Method

Partitioned Gauss-Legendre Runge-Kutta tableau with s stages

PartitionedTableauGauss(::Type{T}, s)
PartitionedTableauGauss(s) = PartitionedTableauGauss(Float64, s)

The constructor takes the number of stages s and optionally the element type T of the tableau.

This partitioned tableau uses TableauGauss for both coefficients a and ā.

RungeKutta.PartitionedTableaus.TableauLobattoIIIAIIIB — Method

Partitioned Gauss-Lobatto IIIA-IIIB tableau with s stages

TableauLobattoIIIAIIIB(::Type{T}, s)
TableauLobattoIIIAIIIB(s) = TableauLobattoIIIAIIIB(Float64, s)

The constructor takes the number of stages s and optionally the element type T of the tableau.

This partitioned tableau uses TableauLobattoIIIA for a and TableauLobattoIIIB for ā.

RungeKutta.PartitionedTableaus.TableauLobattoIIIBIIIA — Method

Tableau for Gauss-Lobatto IIIB-IIIA method with s stages

TableauLobattoIIIBIIIA(::Type{T}, s)
TableauLobattoIIIBIIIA(s) = TableauLobattoIIIBIIIA(Float64, s)

The constructor takes the number of stages s and optionally the element type T of the tableau.

This partitioned tableau uses TableauLobattoIIIB for a and TableauLobattoIIIA for ā.

RungeKutta.PartitionedTableaus.TableauLobattoIIICIIIC̄ — Method

Tableau for Gauss-Lobatto IIIC-IIIC̄ method with s stages

TableauLobattoIIICIIIC̄(::Type{T}, s)
TableauLobattoIIICIIIC̄(s) = TableauLobattoIIICIIIC̄(Float64, s)

The constructor takes the number of stages s and optionally the element type T of the tableau.

This partitioned tableau uses TableauLobattoIIIC for a and TableauLobattoIIIC̄ for ā.

RungeKutta.PartitionedTableaus.TableauLobattoIIIC̄IIIC — Method

Tableau for Gauss-Lobatto IIIC̄-IIIC method with s stages

TableauLobattoIIIC̄IIIC(::Type{T}, s)
TableauLobattoIIIC̄IIIC(s) = TableauLobattoIIIC̄IIIC(Float64, s)

The constructor takes the number of stages s and optionally the element type T of the tableau.

This partitioned tableau uses TableauLobattoIIIC̄ for a and TableauLobattoIIIC for ā.