SPARK Tableaus

GeometricIntegrators.Tableaus.get_lobatto_glrk_coefficients — Function

The projective Lobatto-GLRK coefficients are implicitly given by

\[\sum \limits_{j=1}^{s} a_{ij} c_{j}^{k-1} = \frac{\bar{c}_i^k}{k} \qquad i = 1 , \, ... , \, \sigma , \; k = 1 , \, ... , \, s ,\]

where $c$ are Gauß-Legendre nodes with $s$ stages and $\bar{c}$ are Gauß-Lobatto nodes with $\sigma$ stages.

GeometricIntegrators.Tableaus.TableauHPARKGLRK — Method

Tableau for Gauss-Legendre HPARK method with s stages.

GeometricIntegrators.Tableaus.TableauHPARKLobattoIIIAIIIB — Method

SPARK tableau for Gauss-Lobatto IIIA-IIIB HPARK method with s stages.

GeometricIntegrators.Tableaus.TableauHPARKLobattoIIIBIIIA — Method

SPARK tableau for Gauss-Lobatto IIIB-IIIA method with s stages.

GeometricIntegrators.Tableaus.TableauHSPARKGLRKpSymmetric — Method

Tableau for Gauss-Legendre method with s stages and symplectic projection.

GeometricIntegrators.Tableaus.TableauHSPARKLobattoIIIAIIIBpSymmetric — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with s stages and symmetric projection.

GeometricIntegrators.Tableaus.TableauHSPARKLobattoIIIBIIIApSymmetric — Method

Tableau for Gauss-Lobatto IIIB-IIIA method with s stages and symmetric projection.

GeometricIntegrators.Tableaus.TableauSPARKGLRK — Method

SPARK tableau for Gauss-Legendre Runge-Kutta method with s stages.

GeometricIntegrators.Tableaus.TableauSPARKGLRKLobattoIIIAIIIB — Function

SPARK tableau for Gauss-Legendre/Gauss-Lobatto-IIIA-IIIB methods.

GeometricIntegrators.Tableaus.TableauSPARKGLRKLobattoIIIBIIIA — Function

SPARK tableau for Gauss-Legendre/Gauss-Lobatto-IIIB-IIIA methods.

GeometricIntegrators.Tableaus.TableauSPARKGLVPRK — Method

Tableau for Variational Gauss-Legendre method with s stages.

GeometricIntegrators.Tableaus.TableauSPARKLobABC — Method

Tableau for Gauss-Lobatto IIIA-IIIB-IIIC method with s stages.

GeometricIntegrators.Tableaus.TableauSPARKLobABD — Method

Tableau for Gauss-Lobatto IIIA-IIIB-IIID method with s stages.

GeometricIntegrators.Tableaus.TableauSPARKLobatto — Method

SPARK tableau for Gauss-Lobatto methods.

GeometricIntegrators.Tableaus.TableauSPARKLobattoIIIAIIIB — Method

SPARK tableau for Gauss-Lobatto IIIA-IIIB method with s stages.

GeometricIntegrators.Tableaus.TableauSPARKLobattoIIIBIIIA — Method

SPARK tableau for Gauss-Lobatto IIIB-IIIA method with s stages.

GeometricIntegrators.Tableaus.TableauSPARKVPRK — Method

SPARK Tableau for Variational Partitioned Runge-Kutta Methods.

GeometricIntegrators.Tableaus.TableauGausspSymplectic — Method

Tableau for Gauss-Legendre method with s stages and symplectic projection.

GeometricIntegrators.Tableaus.TableauLobattoIIIAIIIBpSymplectic — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with s stages and symplectic projection.

GeometricIntegrators.Tableaus.TableauLobattoIIIBIIIApSymplectic — Method

Tableau for Gauss-Lobatto IIIB-IIIA method with s stages and symplectic projection.

GeometricIntegrators.Tableaus.TableauVSPARKGLRKpInternal — Method

Tableau for Gauss-Legendre method with s stages and symplectic projection.

GeometricIntegrators.Tableaus.TableauVSPARKGLRKpLobattoIIIAIIIB — Method

Tableau for Gauss-Legendre method with s stages and Lobatto-IIIA-IIIB projection.

GeometricIntegrators.Tableaus.TableauVSPARKGLRKpLobattoIIIBIIIA — Method

Tableau for Gauss-Legendre method with s stages and Lobatto-IIIB-IIIA projection.

GeometricIntegrators.Tableaus.TableauVSPARKGLRKpMidpoint — Method

Tableau for Gauss-Legendre method with s stages and midpoint projection.

GeometricIntegrators.Tableaus.TableauVSPARKGLRKpModifiedInternal — Method

Tableau for Gauss-Legendre method with s stages and symplectic projection.

GeometricIntegrators.Tableaus.TableauVSPARKGLRKpModifiedLobattoIIIAIIIB — Method

Tableau for Gauss-Legendre method with s stages and symplectic projection.

GeometricIntegrators.Tableaus.TableauVSPARKGLRKpModifiedLobattoIIIBIIIA — Method

Tableau for Gauss-Legendre method with s stages and symplectic projection.

GeometricIntegrators.Tableaus.TableauVSPARKGLRKpModifiedMidpoint — Method

Tableau for Gauss-Legendre method with s stages and midpoint projection.

GeometricIntegrators.Tableaus.TableauVSPARKGLRKpSymmetric — Method

Tableau for Gauss-Legendre method with s stages and symplectic projection.

GeometricIntegrators.Tableaus.TableauVSPARKInternalProjection — Method

Consider a symplectic pair of tableaus $(a^{1}, b^{1}, c^{1})$ and $(a^{3}, b^{3}, c^{3})$, i.e., satsifying $b^{1}_{i} b^{3}_{j} = b^{1}_{i} a^{3}_{ij} + b^{3}_{j} a^{1}_{ji}$, with an arbitrary number of stages $s$. Use the same tableaus for $\tilde{a}^{1}$ and $\tilde{a}^{3}$, so that $\tilde{s} = s$, as well as

\[\begin{aligned} \begin{array}{c|cc} & \tfrac{1}{2} b^{1} \\ & \vdots \\ & \tfrac{1}{2} b^{1} \\ \hline a^{2} & \\ \end{array} && \begin{array}{c|cc} & \tfrac{1}{2} b^{3} \\ & \vdots \\ & \tfrac{1}{2} b^{3} \\ \hline a^{4} & \\ \end{array} && \begin{array}{c|cc} & \tfrac{1}{2} b^{1} \\ c^{1} & \vdots \\ & \tfrac{1}{2} b^{1} \\ \hline \tilde{a}^{2} & \tfrac{1}{2} (1 + R(\infty)) \, b^{1} \\ \end{array} && \begin{array}{c|cc} & \tfrac{1}{2} b^{3} \\ c^{3} & \vdots \\ & \tfrac{1}{2} b^{3} \\ \hline \tilde{a}^{4} & \tfrac{1}{2} (1 + R(\infty)) \, b^{3} \\ \end{array} \end{aligned}\]

Set $\omega = [0, ..., 0, 1]$ and

\[\delta_{ij} = \begin{cases} +1 & j = i , \\ -1 & j = \tilde{s} , \\ 0 & \text{else} , \end{cases}\]

so that $\Lambda_{1} = \Lambda_{2} = ... = \Lambda_{\tilde{s}}$.

This methods is constructed to satisfy the constraint on the projective stages, $\phi(\tilde{Q}_{n,i}, \tilde{P}_{n,i}) = 0$ for $i = 1, \, ..., \, \tilde{s}$. Note, however, that it violates the symplecticity conditions $b^{1}_{i} b^{4}_{j} = b^{1}_{i} a^{4}_{ij} + b^{4}_{j} \tilde{a}^{1}_{ji}$ and $b^{2}_{i} b^{3}_{j} = b^{2}_{i} \tilde{a}^{3}_{ij} + b^{3}_{j} a^{2}_{ji}$.

GeometricIntegrators.Tableaus.TableauVSPARKLobattoIIIAIIIBProjection — Method

Consider a symplectic pair of tableaus $(a^{1}, b^{1}, c^{1})$ and $(a^{3}, b^{3}, c^{3})$, i.e., satsifying $b^{1}_{i} b^{3}_{j} = b^{1}_{i} a^{3}_{ij} + b^{3}_{j} a^{1}_{ji}$, with an arbitrary number of stages $s$. For the projection, choose the Lobatto-IIIA and IIIB tableaus with $\tilde{s} = 2$ stages for $(\tilde{a}^{4}, b^{4})$ and $(\tilde{a}^{2}, b^{2})$, respectively, and choose $\tilde{a}^{1}$ and $\tilde{a}^{3}$ such that the projective stages correspond to the initial condition and the solution, i.e.,

\[\begin{aligned} \begin{array}{c|cc} 0 & 0 \\ 1 & b^{1} \\ \hline \tilde{a}^{1} & \\ \end{array} && \begin{array}{c|cc} 0 & \tfrac{1}{2} & 0 \\ 1 & \tfrac{1}{2} & 0 \\ \hline \tilde{a}^{2} & \\ \end{array} && \begin{array}{c|cc} 0 & 0 \\ 1 & b^{3} \\ \hline \tilde{a}^{3} & \\ \end{array} && \begin{array}{c|cc} 0 & 0 & 0 \\ 1 & \tfrac{1}{2} & \tfrac{1}{2} \\ \hline \tilde{a}^{4} & \\ \end{array} \end{aligned}\]

and compute $a^{2}$ and $a^{4}$ by the symplecticity conditions, that is $a^{2}_{ij} = b^{2}_{j} ( b^{3}_{i} - \tilde{a}^{3}_{ji} ) / b^{3}_{i}$ and $a^{4}_{ij} = b^{4}_{j} ( b^{1}_{i} - \tilde{a}^{1}_{ji}) / b^{1}_{i}$. Finally choose $\omega = [0, 0, 1]$ and $\delta = [-1, R_{\infty}]$, implying that $\rho = 1$. By construction, this method satisfies all symplecticity conditions, but the constraint on the projection stages, $\phi(\tilde{Q}_{n,i}, \tilde{P}_{n,i}) = 0$ for $i = 1, \, ..., \, \tilde{s}$, is not satisfied exactly, but only approximately, although with bounded error.

GeometricIntegrators.Tableaus.TableauVSPARKLobattoIIIAIIIBpLobattoIIIAIIIB — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with s stages and Lobatto-IIIA-IIIB projection.

GeometricIntegrators.Tableaus.TableauVSPARKLobattoIIIAIIIBpLobattoIIIBIIIA — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with s stages and Lobatto-IIIB-IIIA projection.

GeometricIntegrators.Tableaus.TableauVSPARKLobattoIIIAIIIBpMidpoint — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with s stages and midpoint projection.

GeometricIntegrators.Tableaus.TableauVSPARKLobattoIIIAIIIBpModifiedLobattoIIIAIIIB — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with s stages and Lobatto-IIIA-IIIB projection.

GeometricIntegrators.Tableaus.TableauVSPARKLobattoIIIAIIIBpModifiedLobattoIIIBIIIA — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with s stages and Lobatto-IIIB-IIIA projection.

GeometricIntegrators.Tableaus.TableauVSPARKLobattoIIIAIIIBpModifiedMidpoint — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with s stages and midpoint projection.

GeometricIntegrators.Tableaus.TableauVSPARKLobattoIIIAIIIBpSymmetric — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with s stages and symmetric projection.

GeometricIntegrators.Tableaus.TableauVSPARKLobattoIIIBIIIAProjection — Method

This methods is the same as TableauVSPARKLobattoIIIAIIIBProjection, except for using Lobatto-IIIA and IIIB tableaus with $\tilde{s} = 2$ stages for $(\tilde{a}^{2}, b^{2})$, and $(\tilde{a}^{4}, b^{4})$ respectively, instead of the other way around.

GeometricIntegrators.Tableaus.TableauVSPARKLobattoIIIBIIIApLobattoIIIAIIIB — Method

Tableau for Gauss-Lobatto IIIB-IIIA method with s stages and Lobatto-IIIA-IIIB projection.

GeometricIntegrators.Tableaus.TableauVSPARKLobattoIIIBIIIApLobattoIIIBIIIA — Method

Tableau for Gauss-Lobatto IIIB-IIIA method with s stages and Lobatto-IIIB-IIIA projection.

GeometricIntegrators.Tableaus.TableauVSPARKLobattoIIIBIIIApMidpoint — Method

Tableau for Gauss-Lobatto IIIB-IIIA method with s stages and midpoint projection.

GeometricIntegrators.Tableaus.TableauVSPARKLobattoIIIBIIIApModifiedLobattoIIIAIIIB — Method

Tableau for Gauss-Lobatto IIIB-IIIA method with s stages and Lobatto-IIIA-IIIB projection.

GeometricIntegrators.Tableaus.TableauVSPARKLobattoIIIBIIIApModifiedLobattoIIIBIIIA — Method

Tableau for Gauss-Lobatto IIIB-IIIA method with s stages and Lobatto-IIIB-IIIA projection.

GeometricIntegrators.Tableaus.TableauVSPARKLobattoIIIBIIIApModifiedMidpoint — Method

Tableau for Gauss-Lobatto IIIB-IIIA method with s stages and midpoint projection.

GeometricIntegrators.Tableaus.TableauVSPARKLobattoIIIBIIIApSymmetric — Method

Tableau for Gauss-Lobatto IIIB-IIIA method with s stages and symmetric projection.

GeometricIntegrators.Tableaus.TableauVSPARKMidpointProjection — Method

Consider a symplectic pair of tableaus $(a^{1}, b^{1}, c^{1})$ and $(a^{3}, b^{3}, c^{3})$, i.e., satsifying $b^{1}_{i} b^{3}_{j} = b^{1}_{i} a^{3}_{ij} + b^{3}_{j} a^{1}_{ji}$, with an arbitrary number of stages $s$. For the projection, choose the tableau with $\tilde{s} = 1$ and $\rho = 0$, such that $\tilde{Q}_{n,1} = \tfrac{1}{2} ( q_{n} + q_{n+1})$, $\tilde{P}_{n,1} = \tfrac{1}{2} ( p_{n} + p_{n+1})$, i.e.,

\[\begin{aligned} \begin{array}{c|cc} & \tfrac{1}{2} \\ & \vdots \\ & \tfrac{1}{2} \\ \hline a^{2} & \\ \end{array} && \begin{array}{c|cc} & \tfrac{1}{2} \\ & \vdots \\ & \tfrac{1}{2} \\ \hline a^{4} & \\ \end{array} && \begin{array}{c|cc} \tfrac{1}{2} & \tfrac{1}{2} \\ \hline \tilde{a}^{2} & \tfrac{1}{2} (1 + R(\infty))\\ \end{array} && \begin{array}{c|cc} \tfrac{1}{2} & \tfrac{1}{2} \\ \hline \tilde{a}^{4} & \tfrac{1}{2} (1 + R(\infty))\\ \end{array} \end{aligned}\]

The coefficients $\tilde{a}^{1}$ and $\tilde{a}^{3}$ are determined by the symplecticity conditions, specifically $a^{4}_{ij} = b^{4}_{j} ( b^{1}_{i} - \tilde{a}^{1}_{ji}) / b^{1}_{i}$ and $a^{2}_{ij} = b^{2}_{j} ( b^{3}_{i} - \tilde{a}^{3}_{ji} ) / b^{3}_{i}$, and $\omega = [0, 1]$.

GeometricIntegrators.Tableaus.TableauVSPARKModifiedInternalProjection — Function

Consider a symplectic pair of tableaus $(a^{1}, b^{1}, c^{1})$ and $(a^{3}, b^{3}, c^{3})$, i.e., satsifying $b^{1}_{i} b^{3}_{j} = b^{1}_{i} a^{3}_{ij} + b^{3}_{j} a^{1}_{ji}$, with an arbitrary number of stages $s$, and set

\[\begin{aligned} \begin{array}{c|cc} & \tfrac{1}{2} b^{1} \\ & \vdots \\ & \tfrac{1}{2} b^{1} \\ \hline a^{2} & \\ \end{array} && \begin{array}{c|cc} & \tfrac{1}{2} b^{4} \\ & \vdots \\ & \tfrac{1}{2} b^{4} \\ \hline a^{4} & \\ \end{array} && \begin{array}{c|cc} & \tfrac{1}{2} b^{1} \\ c^{1} & \vdots \\ & \tfrac{1}{2} b^{1} \\ \hline \tilde{a}^{2} & \tfrac{1}{2} (1 + R(\infty)) \, b^{1} \\ \end{array} && \begin{array}{c|cc} & \tfrac{1}{2} b^{3} \\ c^{3} & \vdots \\ & \tfrac{1}{2} b^{3} \\ \hline \tilde{a}^{4} & \tfrac{1}{2} (1 + R(\infty)) \, b^{3} \\ \end{array} \end{aligned}\]

Note that by this definition $\tilde{s} = s$. The coefficients $\tilde{a}^{1}$ and $\tilde{a}^{3}$ are determined by the (modified) symplecticity conditions, specifically $a^{4}_{ij} = b^{3}_{j} ( b^{1}_{i} - \tilde{a}^{1}_{ji}) / b^{1}_{i}$ and $a^{2}_{ij} = b^{1}_{j} ( b^{3}_{i} - \tilde{a}^{3}_{ji} ) / b^{3}_{i}$, where $b^{2}$ has been replaced with $b^{1}$ and $b^{4}$ with $b^{3}$, respectively. Set $\omega = [0, ..., 0, 1]$ and

\[\delta_{ij} = \begin{cases} +1 & j = i , \\ -1 & j = \tilde{s} , \\ 0 & \text{else} , \end{cases}\]

so that $\Lambda_{1} = \Lambda_{2} = ... = \Lambda_{\tilde{s}}$.

Note that this method satisfies the symplecticity conditions $b^{1}_{i} b^{4}_{j} = b^{1}_{i} a^{4}_{ij} + b^{4}_{j} \tilde{a}^{1}_{ji}$ and $b^{2}_{i} b^{3}_{j} = b^{2}_{i} \tilde{a}^{3}_{ij} + b^{3}_{j} a^{2}_{ji}$ only if $R(\infty) = 1$ due to the definitions of $b^{2}$ and $b^{4}$. Moreover, it does usually not satisfy the constraint on the projective stages, $\phi(\tilde{Q}_{n,i}, \tilde{P}_{n,i}) = 0$ for $i = 1, \, ..., \, \tilde{s}$, exactly, but only approximately with bounded error, thus implying a residual in the symplecticity equation even if $R(\infty) = 1$.

GeometricIntegrators.Tableaus.TableauVSPARKModifiedLobattoIIIAIIIBProjection — Method

Consider a symplectic pair of tableaus $(a^{1}, b^{1}, c^{1})$ and $(a^{3}, b^{3}, c^{3})$, i.e., satsifying $b^{1}_{i} b^{3}_{j} = b^{1}_{i} a^{3}_{ij} + b^{3}_{j} a^{1}_{ji}$, with an arbitrary number of stages $s$. For the projection, choose the Lobatto-IIIA and IIIB tableaus with $\tilde{s} = 2$ stages for $(\tilde{a}^{4}, b^{4})$ and $(\tilde{a}^{2}, b^{2})$, respectively.

The coefficients $\tilde{a}^{1}$ and $\tilde{a}^{3}$ are determined by the relations

\[\begin{aligned} \sum \limits_{j=1}^{s} \tilde{a}^{1}_{ij} (c_{j}^{1})^{k-1} &= \frac{(c_{i}^{2})^k}{k} , \qquad & \sum \limits_{j=1}^{s} \tilde{a}^{3}_{ij} (c_{j}^{3})^{k-1} &= \frac{(c_{i}^{4})^k}{k} , \qquad & i &= 1 , \, ... , \, \tilde{s} , \qquad & k &= 1 , \, ... , \, s . \end{aligned}\]

The coefficients $a^{2}$ and $a^{4}$ by the symplecticity conditions, that is $a^{2}_{ij} = b^{2}_{j} ( b^{3}_{i} - \tilde{a}^{3}_{ji} ) / b^{3}_{i}$ and $a^{4}_{ij} = b^{4}_{j} ( b^{1}_{i} - \tilde{a}^{1}_{ji}) / b^{1}_{i}$. Finally choose $\omega = [0, 0, 1]$ and $\delta = [-1, R_{\infty}]$, implying that $\rho = 1$. By construction, this method satisfies all symplecticity conditions, but the constraint on the projection stages, $\phi(\tilde{Q}_{n,i}, \tilde{P}_{n,i}) = 0$ for $i = 1, \, ..., \, \tilde{s}$, is not satisfied exactly, but only approximately, although with bounded error.

GeometricIntegrators.Tableaus.TableauVSPARKModifiedLobattoIIIBIIIAProjection — Method

This methods is the same as TableauVSPARKModifiedLobattoIIIAIIIBProjection, except for using Lobatto-IIIA and IIIB tableaus with $\tilde{s} = 2$ stages for $(\tilde{a}^{2}, b^{2})$, and $(\tilde{a}^{4}, b^{4})$ respectively, instead of the other way around.

GeometricIntegrators.Tableaus.TableauVSPARKModifiedMidpointProjection — Method

Consider a symplectic pair of tableaus $(a^{1}, b^{1}, c^{1})$ and $(a^{3}, b^{3}, c^{3})$, i.e., satsifying $b^{1}_{i} b^{3}_{j} = b^{1}_{i} a^{3}_{ij} + b^{3}_{j} a^{1}_{ji}$, with an arbitrary number of stages $s$. For the projection, choose the tableau with $\tilde{s} = 1$ and $\rho = 0$, such that $\tilde{Q}_{n,1} = \tfrac{1}{2} ( q_{n} + q_{n+1})$, $\tilde{P}_{n,1} = \tfrac{1}{2} ( p_{n} + p_{n+1})$, i.e.,

\[\begin{aligned} \begin{array}{c|cc} \tfrac{1}{2} & \tfrac{1}{2} b^{1} \\ \hline \tilde{a}^{1} & \\ \end{array} && \begin{array}{c|cc} \tfrac{1}{2} & \tfrac{1}{2} \\ \hline \tilde{a}^{2} & \tfrac{1}{2} ( 1 + R (\infty) ) \\ \end{array} && \begin{array}{c|cc} \tfrac{1}{2} & \tfrac{1}{2} b^{3} \\ \hline \tilde{a}^{3} & \\ \end{array} && \begin{array}{c|cc} \tfrac{1}{2} & \tfrac{1}{2} \\ \hline \tilde{a}^{4} & \tfrac{1}{2} ( 1 + R (\infty) ) \\ \end{array} \end{aligned}\]

The coefficients $a^{2}$ and $a^{4}$ are determined by the symplecticity conditions, specifically $a^{4}_{ij} = b^{4}_{j} ( b^{1}_{i} - \tilde{a}^{1}_{ji}) / b^{1}_{i}$ and $a^{2}_{ij} = b^{2}_{j} ( b^{3}_{i} - \tilde{a}^{3}_{ji} ) / b^{3}_{i}$, and $\omega = [0, 1]$.

GeometricIntegrators.Tableaus.TableauVSPARKSymmetricProjection — Method

Consider a symplectic pair of tableaus $(a^{1}, b^{1}, c^{1})$ and $(a^{3}, b^{3}, c^{3})$, i.e., satsifying $b^{1}_{i} b^{3}_{j} = b^{1}_{i} a^{3}_{ij} + b^{3}_{j} a^{1}_{ji}$, with an arbitrary number of stages $s$. For the projection, choose the tableau with $\tilde{s} = 2$ and $\rho = 1$, such that $\tilde{Q}_{n,1} = q_{n}$, $\tilde{Q}_{n,2} = q_{n+1}$, $\tilde{P}_{n,1} = p_{n}$, $\tilde{P}_{n,2} = p_{n+1}$, i.e.,

\[\begin{aligned} \begin{array}{c|cc} 0 & 0 \\ 1 & b^{1} \\ \hline \tilde{a}^{1} & \\ \end{array} && \begin{array}{c|cc} & 0 & 0 \\ & \tfrac{1}{2} & \tfrac{1}{2} \\ \hline \tilde{a}^{2} & \tfrac{1}{2} & \tfrac{1}{2} \\ \end{array} && \begin{array}{c|cc} 0 & 0 \\ 1 & b^{3} \\ \hline \tilde{a}^{3} & \\ \end{array} && \begin{array}{c|cc} & 0 & 0 \\ & \tfrac{1}{2} & \tfrac{1}{2} \\ \hline \tilde{a}^{4} & \tfrac{1}{2} & \tfrac{1}{2} \\ \end{array} \end{aligned}\]

The coefficients $a^{2}$ and $a^{4}$ are determined by the symplecticity conditions, specifically $a^{4}_{ij} = b^{4}_{j} ( b^{1}_{i} - \tilde{a}^{1}_{ji}) / b^{1}_{i}$ and $a^{2}_{ij} = b^{2}_{j} ( b^{3}_{i} - \tilde{a}^{3}_{ji} ) / b^{3}_{i}$. Further choose $\omega = [1, 1, 0]$ and $\delta = [-1, R_{\infty}]$, so that $\tilde{\Lambda}_{n,1} = R_{\infty} \tilde{\Lambda}_{n,2}$ and

\[\tilde{P}_{n,1} - \vartheta (\tilde{Q}_{n,1}) + R_{\infty} ( \tilde{P}_{n,2} - \vartheta (\tilde{Q}_{n,2}) ) = 0 .\]

Due to the particular choice of projective stages, this is equivalent to

\[p_{n} - \vartheta (q_{n}) + R_{\infty} ( p_{n+1} - \vartheta (q_{n+1}) ) = 0 ,\]

so that the constraint $\phi(q_{n+1}, p_{n+1}) = 0$ is satisfied if $\phi(q_{n}, p_{n}) = 0$. Note that the choice of $\tilde{a}^{2}$ and $\tilde{a}^{4}$ violates the symplecticity condition $b^{2}_{i} b^{4}_{j} = b^{2}_{i} \tilde{a}^{4}_{ij} + b^{4}_{j} \tilde{a}^{2}_{ji}$.