Integrators

Common

GeometricIntegrators.Integrators.get_internal_variables — Method

get_internal_variables(::Integrator) = NamedTuple()

Returns a NamedTuple containing all internal variables of an integrator that shall be stored in an AtomicSolution. If there is no method for a specific integrator implemented an empty NamedTuple() is returned.

GeometricIntegrators.Integrators.integrate_step! — Method

Performs one time step with a given integrator.

integrate_step!(integrator::Integrator, asol::AtomicSolution)

The function accepts two arguments: an integrator and an appropriate AtomicSolution, which contains the state of the system at the beginning and the end of the time step and possibly additional information like solver output or the solution at internal stages of a Runge-Kutta method.

GeometricIntegrators.Integrators.AbstractTableau — Type

Holds the information for the various methods' tableaus.

GeometricIntegrators.Integrators.create_internal_stage_matrix — Method

Create a vector of S solution matrices of type DT to store the solution of S internal stages for a problem with DxM dimensions.

GeometricIntegrators.Integrators.create_internal_stage_vector — Method

Create a vector of S solution vectors of type DT to store the solution of S internal stages for a problem with D dimensions.

GeometricIntegrators.Integrators.create_internal_stage_vector_with_zero — Method

Create a vector of (S,M+1) solution vectors of type DT to store the solution of S internal stages and M random processes for a problem with D dimensions.

GeometricIntegrators.Integrators.create_internal_stage_vector_with_zero — Method

Create a vector of S+1 solution vectors of type DT to store the solution of S internal stages and the solution of the previous timestep for a problem with D dimensions.

GeometricIntegrators.Integrators.create_nonlinear_solver — Function

Create nonlinear solver object for a system of N equations with data type DT. The function $f(x)=0$ to be solved for is determined by a julia function function_stages!(x, b, params), where x is the current solution and b is the output vector, s.th. $b = f(x)$. params are a set of parameters depending on the equation and integrator that is used. The solver type is obtained from the config dictionary (:nls_solver).

GeometricIntegrators.Integrators.Integrator — Type

Integrator(equation, tableau, Δt)

Universal constructor for Runge-Kutta (standard, partitioned, variational), SPARK and splitting integrators that automatically selects the appropriate integrator based on the equation and tableau types.

GeometricIntegrators.Integrators.IntegratorConstructor — Function

IntegratorConstructor(DT, D)
IntegratorConstructor(DT, D, tableau)

Returns a closure for various integrator constructors that is needed for composition methods based on the tableau type. If not tableau is provided, a closure for the constructor of an exact solution is returned.

DT: data type of the solution / state vector

D: dimension of the state vector

GeometricIntegrators.Integrators.integrate — Function

integrate(equation, integrator, ntime; kwargs...)
integrate(equation, tableau, Δt, ntime; kwargs...)

Integrate an equation with integrator for ntime time steps and return the solution. If a tableau and a time step Δt is passed instead of an integrator, the appropriate integrator is created automatically.

Some convenience methods exist for the integration of ODEs,

integrate(v, x₀, tableau, Δt, ntime; t₀=0., kwargs...)

where v is the function for the vector field, x₀ the initial condition and t₀ the initial time, and for PODEs

integrate(v, f, q₀, p₀, tableau, Δt, ntime; t₀=0., kwargs...)

with vector fields v and f and initial conditions q₀ and p₀.

GeometricIntegrators.Integrators.integrate! — Function

integrate!(integrator, solution)

Perform the actual integration for a given integrator and solution. Solve all time steps for all initial conditions.

integrate!(integrator, solution, m₁, m₂)

Solve all time steps for initial conditions m with m₁ ≤ m ≤ m₂.

integrate!(integrator, solution, m₁, m₂, n₁, n₂)

Solve time steps n with n₁ ≤ n ≤ n₂ for initial conditions m with m₁ ≤ m ≤ m₂.

integrate!(integrator, solution, atomic_solution, m, n)

Solve one time step n for one initial condition m.

Initial Guesses

GeometricIntegrators.Integrators.aitken_neville — Method

Compute p(x) where p is the unique polynomial of degree length(xi), such that p(x[i]) = y[i]) for all i.

ti: interpolation nodes
xi: interpolation values
t:  evaluation point
x:  evaluation value

GeometricIntegrators.Integrators.euler_extrapolation — Method

Euler extrapolation method with arbitrary order p.

v:  function to compute vector field
t₀: initial time
t₁: final   time
x₀: initial value
x₁: final   value
s:  number of interpolations (order p=s+1)

TODO This is probably broken!

GeometricIntegrators.Integrators.midpoint_extrapolation — Method

Midpoint extrapolation method with arbitrary order p.

v:  function to compute vector field
f:  function to compute force  field
t₀: initial time
t₁: final   time
q₀: initial positions
p₀: initial momenta
q₁: final   positions
p₁: final   momenta
s:  number of interpolations (order p=2s+2)

GeometricIntegrators.Integrators.midpoint_extrapolation — Method

Midpoint extrapolation method with arbitrary order p.

v:  function to compute vector field
t₀: initial time
t₁: final   time
x₀: initial value
x₁: final   value
s:  number of interpolations (order p=2s+2)

GeometricIntegrators.Integrators.InitialGuessODE — Type

InitialGuessODE: Initial guess for ordinary differential equations

Fields

int: interpolation structure

v: vector field

Δt: time step

s: number of extrapolation stages (for initialisation)

GeometricIntegrators.Integrators.initialize! — Method

Initialise initial guess, i.e., given t₀, t₁, q₁ compute q₀, v₀, v₁.

GeometricIntegrators.Integrators.update_vector_fields! — Method

compute vector field of new solution

GeometricIntegrators.Integrators.InitialGuessPODE — Type

InitialGuessPODE: Initial guess for partitioned ordinary differential equations

Fields

int: interpolation structure

v: vector field for q

f: vector field for p

Δt: time step

s: number of extrapolation stages (for initialisation)

GeometricIntegrators.Integrators.initialize! — Method

Initialise initial guess, i.e., given t₀, t₁, q₁ compute q₀, v₀, v₁.

Splitting Methods

GeometricIntegrators.Integrators.IntegratorComposition — Type

Composition integrator for the solution of initial value problems

\[\dot{q} (t) = v(t, q(t)) , \qquad q(t_{0}) = q_{0} ,\]

whose vector field $v$ is given as a sum of vector fields

\[v (t) = v_1 (t) + ... + v_r (t) .\]

IntegratorComposition has three constructors:

IntegratorComposition{DT,D}(integrators::Tuple, Δt)
IntegratorComposition(equation::SODE, constructors::Tuple, tableau::AbstractTableauSplitting, Δt)
IntegratorComposition(equation::SODE, tableau::AbstractTableauSplitting, Δt)

In the first constructor, DT is the data type of the state vector and D the dimension of the system. In the second and third constructor, this information is extracted from the equation. The tuple integrators contains the integrators for each substep. Each integrator is instantiated with appropriately scaled time step size $\Delta t = c_i \tau$ to match the corresponding splitting scheme. In the second constructor, the tuple constructors contains constructors for the integrators of each step of the composition. The integrators are constructed according to the tableau and time step \Delta t and passed to the first constructor. The third constructor assumes that the exact solution is used for each splitting step. It thus constructs a composition method that is equivalent to a plain IntegratorSplitting.

In order to include exact solutions in the composition, the IntegratorExactODE implements the general integrator interface.

GeometricIntegrators.Integrators.IntegratorSplitting — Type

Splitting integrator for the solution of initial value problems

\[\dot{q} (t) = v(t, q(t)) , \qquad q(t_{0}) = q_{0} ,\]

whose vector field $v$ is given as a sum of vector fields

\[v (t) = v_1 (t) + ... + v_r (t) .\]

IntegratorSplitting has two constructors:

IntegratorSplitting{DT,D}(solutions::Tuple, f::Vector{Int}, c::Vector, Δt)
IntegratorSplitting(equation::SODE, tableau::AbstractTableauSplitting, Δt)

In the first constructor, DT is the data type of the state vector and D the dimension of the system. In the second constructor, this information is extracted from the equation. The tuple solutions contains functions implementing the flow (exact solution) of the vector fields v_i. The vectors f and c define the actual splitting method: f is a vector of indices of the flows in the split equation to be solved and c is a vector of the same size f that contains the coefficients for each splitting step, i.e., the resulting integrator has the form

\[\varphi_{\tau} = \phi_{c[s] \tau}^{v_{f[s]}} \circ \dotsc \circ \phi_{c[2] \tau}^{v_{f[2]}} \circ \phi_{c[1] \tau}^{v_{f[1]}} .\]

In the second constructor, these vectors are constructed from the tableau and the equation.

GeometricIntegrators.Integrators.IntegratorExactODE — Type

Exact solution of an ODE.

GeometricIntegrators.Integrators.TableauSplitting — Type

Tableau for general splitting methods for vector fields with two components A and B.

Integrator:

\[\varphi_{\tau} = \varphi_{b_s \tau}^{B} \circ \varphi_{a_s \tau}^{A} \circ \dotsc \circ \varphi_{b_1 \tau}^{B} \circ \varphi_{a_1 \tau}^{A}\]

GeometricIntegrators.Integrators.TableauSplittingGS — Type

Tableau for symmetric splitting methods with general stages. See McLachlan, Quispel, 2003, Equ. (4.11).

Basic method: Lie composition

\[\begin{aligned} \varphi_{\tau}^{A} &= \varphi_{\tau}^{v_1} \circ \varphi_{\tau}^{v_2} \circ \dotsc \circ \varphi_{\tau}^{v_{r-1}} \circ \varphi_{\tau}^{v_r} \\ \varphi_{\tau}^{B} &= \varphi_{\tau}^{v_r} \circ \varphi_{\tau}^{v_{r-1}} \circ \dotsc \circ \varphi_{\tau}^{v_2} \circ \varphi_{\tau}^{v_1} \end{aligned}\]

Integrator:

\[\varphi_{\tau}^{GS} = \varphi_{a_1 \tau}^{A} \circ \varphi_{b_1 \tau}^{B} \circ \dotsc \circ \varphi_{b_1 \tau}^{B} \circ \varphi_{a_1 \tau}^{A}\]

GeometricIntegrators.Integrators.TableauSplittingNS — Type

Tableau for non-symmetric splitting methods. See McLachlan, Quispel, 2003, Equ. (4.10). The methods A and B are the composition of all vector fields in the SODE and its adjoint, respectively.

Basic method: Lie composition

\[\begin{aligned} \varphi_{\tau}^{A} &= \varphi_{\tau}^{v_1} \circ \varphi_{\tau}^{v_2} \circ \dotsc \circ \varphi_{\tau}^{v_{r-1}} \circ \varphi_{\tau}^{v_r} \\ \varphi_{\tau}^{B} &= \varphi_{\tau}^{v_r} \circ \varphi_{\tau}^{v_{r-1}} \circ \dotsc \circ \varphi_{\tau}^{v_2} \circ \varphi_{\tau}^{v_1} \end{aligned}\]

Integrator:

\[\varphi_{\tau}^{NS} = \varphi_{b_s \tau}^{B} \circ \varphi_{a_s \tau}^{A} \circ \dotsc \circ \varphi_{b_1 \tau}^{B} \circ \varphi_{a_1 \tau}^{A}\]

GeometricIntegrators.Integrators.TableauSplittingSS — Type

Tableau for symmetric splitting methods with symmetric stages. See McLachlan, Quispel, 2003, Equ. (4.6).

Basic method: symmetric Strang composition

\[\varphi_{\tau}^{A} = \varphi_{\tau/2}^{v_1} \circ \varphi_{\tau/2}^{v_2} \circ \dotsc \circ \varphi_{\tau/2}^{v_{r-1}} \circ \varphi_{\tau/2}^{v_r} \circ \varphi_{\tau/2}^{v_r} \circ \varphi_{\tau/2}^{v_{r-1}} \circ \dotsc \circ \varphi_{\tau/2}^{v_2} \circ \varphi_{\tau/2}^{v_1}\]

Integrator:

\[\varphi_{\tau}^{SS} = \varphi_{a_1 \tau}^{A} \circ \varphi_{a_2 \tau}^{A} \circ \dotsc \circ \varphi_{a_s \tau}^{A} \circ \dotsc \circ \varphi_{a_2 \tau}^{A} \circ \varphi_{a_1 \tau}^{A}\]

Runge-Kutta Methods

GeometricIntegrators.Integrators.AbstractTableauPRK — Type

Holds the tableau of a partitioned Runge-Kutta method.

GeometricIntegrators.Integrators.AbstractTableauRK — Type

Holds the tableau of a Runge-Kutta method.

GeometricIntegrators.Integrators.IntegratorCacheERK — Type

Explicit Runge-Kutta integrator cache.

GeometricIntegrators.Integrators.IntegratorERK — Type

Explicit Runge-Kutta integrator.

GeometricIntegrators.Integrators.ParametersERK — Type

Parameters for right-hand side function of explicit Runge-Kutta methods.

GeometricIntegrators.Integrators.IntegratorCacheDIRK — Type

Diagonally implicit Runge-Kutta integrator cache.

GeometricIntegrators.Integrators.IntegratorDIRK — Type

Diagonally implicit Runge-Kutta integrator.

GeometricIntegrators.Integrators.ParametersDIRK — Type

Parameters for right-hand side function of diagonally implicit Runge-Kutta methods.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of fully implicit Runge-Kutta methods.

GeometricIntegrators.Integrators.initial_guess! — Method

Compute initial guess for internal stages.

GeometricIntegrators.Integrators.initialize! — Method

Initialise initial guess

GeometricIntegrators.Integrators.IntegratorCacheFIRK — Type

Fully implicit Runge-Kutta integrator cache.

Fields

q̃: initial guess of solution

ṽ: initial guess of vector field

s̃: holds shift due to periodicity of solution

Q: internal stages of solution

V: internal stages of vector field

Y: vector field of internal stages

GeometricIntegrators.Integrators.IntegratorFIRK — Type

Fully implicit Runge-Kutta integrator.

GeometricIntegrators.Integrators.ParametersFIRK — Type

Parameters for right-hand side function of fully implicit Runge-Kutta methods.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of fully implicit Runge-Kutta methods.

GeometricIntegrators.Integrators.jacobian! — Method

Compute stages of fully implicit Runge-Kutta methods.

GeometricIntegrators.Integrators.IntegratorCacheEPRK — Type

Explicit Runge-Kutta integrator cache.

GeometricIntegrators.Integrators.IntegratorEPRK — Type

Explicit partitioned Runge-Kutta integrator solving the system

\[\begin{aligned} V_{n,i} &= v (Q_{n,i}, P_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} , \\ F_{n,i} &= f (Q_{n,i}, P_{n,i}) , & P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, F_{n,i} . \end{aligned}\]

Usually we are interested in Hamiltonian systems, where

\[\begin{aligned} V_{n,i} &= \dfrac{\partial H}{\partial p} (Q_{n,i}, P_{n,i}) , & F_{n,i} &= - \dfrac{\partial H}{\partial q} (Q_{n,i}, P_{n,i}) , \end{aligned}\]

and tableaus satisfying the symplecticity conditions

\[\begin{aligned} b_{i} \bar{a}_{ij} + \bar{b}_{j} a_{ji} &= b_{i} \bar{b}_{j} , & \bar{b}_i &= b_i . \end{aligned}\]

GeometricIntegrators.Integrators.ParametersEPRK — Type

Parameters for right-hand side function of explicit partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.computeStageP! — Method

Compute P stages of explicit partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.computeStageQ! — Method

Compute Q stages of explicit partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.IntegratorCacheIPRK — Type

Implicit partitioned Runge-Kutta integrator cache.

Fields

q̃: initial guess of q

p̃: initial guess of p

ṽ: initial guess of v

f̃: initial guess of f

s̃: holds shift due to periodicity of solution

Q: internal stages of q

P: internal stages of p

V: internal stages of v

F: internal stages of f

Y: vector field of internal stages of q

Z: vector field of internal stages of p

GeometricIntegrators.Integrators.IntegratorIPRK — Type

Implicit partitioned Runge-Kutta integrator solving the system

\[\begin{aligned} V_{n,i} &= v (Q_{n,i}, P_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} , \\ F_{n,i} &= f (Q_{n,i}, P_{n,i}) , & P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, F_{n,i} , \end{aligned}\]

Usually we are interested in Hamiltonian systems, where

\[\begin{aligned} V_{n,i} &= \dfrac{\partial H}{\partial p} (Q_{n,i}, P_{n,i}) , & F_{n,i} &= - \dfrac{\partial H}{\partial q} (Q_{n,i}, P_{n,i}) , \end{aligned}\]

and tableaus satisfying the symplecticity conditions

\[\begin{aligned} b_{i} \bar{a}_{ij} + \bar{b}_{j} a_{ji} &= b_{i} \bar{b}_{j} , & \bar{b}_i &= b_i . \end{aligned}\]

GeometricIntegrators.Integrators.ParametersIPRK — Type

Parameters for right-hand side function of implicit partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of implicit partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.IntegratorFLRK — Type

Formal Lagrangian Runge-Kutta integrator.

GeometricIntegrators.Integrators.ParametersFLRK — Type

Parameters for right-hand side function of formal Lagrangian Runge-Kutta methods.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of formal Lagrangian Runge-Kutta methods.

GeometricIntegrators.Integrators.CoefficientsPGLRK — Type

Holds the coefficients of a projected Gauss-Legendre Runge-Kutta method.

GeometricIntegrators.Integrators.IntegratorPGLRK — Type

Projected Gauss-Legendre Runge-Kutta integrator.

Reference: LUIGI BRUGNANO, FELICE IAVERNARO, AND DONATO TRIGIANTE.
    ENERGY- AND QUADRATIC INVARIANTS–PRESERVING INTEGRATORS BASED
    UPON GAUSS COLLOCATION FORMULAE.
    SIAM J. NUMER. ANAL. Vol. 50, No. 6, pp. 2897–2916, 2012.

GeometricIntegrators.Integrators.ParametersPGLRK — Type

Parameters for right-hand side function of projected Gauss-Legendre Runge-Kutta methods.

Base.show — Method

Print Runge-Kutta coefficients.

Base.show — Method

Print Projected Gauss-Legendre Runge-Kutta integrator.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of projected Gauss-Legendre Runge-Kutta methods.

SPARK Methods

GeometricIntegrators.Integrators.SPARK.AbstractParametersSPARK — Type

Parameters for right-hand side function of Specialised Partitioned Additive Runge-Kutta methods.

GeometricIntegrators.Integrators.SPARK.AbstractTableauSPARK — Type

Holds the tableau of an Specialised Partitioned Additive Runge-Kutta method for Variational systems.

GeometricIntegrators.Integrators.SPARK.CoefficientsARK — Type

Holds the coefficients of an additive Runge-Kutta method.

GeometricIntegrators.Integrators.SPARK.CoefficientsIRK — Type

Holds the coefficients of an additive Runge-Kutta method.

GeometricIntegrators.Integrators.SPARK.CoefficientsMRK — Type

Holds the multiplier Runge-Kutta coefficients.

GeometricIntegrators.Integrators.SPARK.CoefficientsPRK — Type

Holds the coefficients of a projective Runge-Kutta method.

GeometricIntegrators.Integrators.SPARK.CoefficientsSPARK — Type

Holds the coefficients of a SPARK method.

GeometricIntegrators.Integrators.SPARK.IntegratorCacheSPARK — Type

Cache of a Specialised Partitioned Additive Runge-Kutta integrator.

Fields

n: time step number

t: time of current time step

t̄: time of previous time step

q: current solution of q

q̄: previous solution of q

p: current solution of p

p̄: previous solution of p

v: vector field of q

v̄: vector field of q̄

f: vector field of p

f̄: vector field of p̄

q̃: initial guess of q

p̃: initial guess of p

ṽ: initial guess of v

f̃: initial guess of f

s̃: holds shift due to periodicity of solution

Q: internal stages of q

P: internal stages of p

V: internal stages of v

F: internal stages of f

Y: vector field of internal stages of q

Z: vector field of internal stages of p

GeometricIntegrators.Integrators.SPARK.IntegratorHPARK — Type

Partitioned Additive Runge-Kutta integrator for Hamiltonian systems subject to Dirac constraints EXPERIMENTAL.

This integrator solves the following system of equations for the internal stages,

\[\begin{aligned} Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} V_{n,j} + h \sum \limits_{j=1}^{r} \alpha_{ij} U_{n,j} , & i &= 1, ..., s , \\ P_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} a_{ij} F_{n,j} + h \sum \limits_{j=1}^{r} \alpha_{ij} G_{n,j} , & i &= 1, ..., s , \\ \tilde{Q}_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} \tilde{a}_{ij} V_{n,j} + h \sum \limits_{j=1}^{r} \tilde{\alpha}_{ij} U_{n,j} , & i &= 1, ..., r , \\ \tilde{P}_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} \tilde{a}_{ij} F_{n,j} + h \sum \limits_{j=1}^{r} \tilde{\alpha}_{ij} G_{n,j} , & i &= 1, ..., r , \\ \tilde{\Phi}_{n,i} &= 0 , & i &= 1, ..., r , \end{aligned}\]

with definitions

\[\begin{aligned} V_{n,i} &= \hphantom{-} \frac{\partial H}{\partial p} (Q_{n,i}, P_{n,i}) , & i &= 1, ..., s , \\ F_{n,i} &= - \frac{\partial H}{\partial q} (Q_{n,i}, P_{n,i}) , & i &= 1, ..., s , \\ U_{n,i} &= \hphantom{-} \frac{\partial \phi}{\partial p} (\tilde{Q}_{n,i}, \tilde{P}_{n,i})^{T} \Lambda_{n,i} , & i &= 1, ..., r , \\ G_{n,i} &= - \frac{\partial \phi}{\partial q} (\tilde{Q}_{n,i}, \tilde{P}_{n,i})^{T} \Lambda_{n,i} , & i &= 1, ..., r , \\ \tilde{\Phi}_{n,i} &= \phi(\tilde{Q}_{n,i}, \tilde{P}_{n,i}) , & i &= 1, ..., r , \end{aligned}\]

and update rule

\[\begin{aligned} q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} V_{n,i} + h \sum \limits_{i=1}^{r} \beta_{i} U_{n,i} , \\ p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} b_{i} F_{n,i} + h \sum \limits_{i=1}^{r} \beta_{i} G_{n,i} . \end{aligned}\]

GeometricIntegrators.Integrators.SPARK.IntegratorHSPARK — Type

Specialised Partitioned Additive Runge-Kutta integrator for Hamiltonian systems EXPERIMENTAL.

This integrator solves the following system of equations for the internal stages,

\[\begin{aligned} Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} V_{n,j} + h \sum \limits_{j=1}^{r} \alpha_{ij} U_{n,j} , & i &= 1, ..., s , \\ P_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} a_{ij} F_{n,j} + h \sum \limits_{j=1}^{r} \alpha_{ij} G_{n,j} , & i &= 1, ..., s , \\ \tilde{Q}_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} \tilde{a}_{ij} V_{n,j} + h \sum \limits_{j=1}^{r} \tilde{\alpha}_{ij} U_{n,j} , & i &= 1, ..., r , \\ \tilde{P}_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} \tilde{a}_{ij} F_{n,j} + h \sum \limits_{j=1}^{r} \tilde{\alpha}_{ij} G_{n,j} , & i &= 1, ..., r , \\ 0 &= \sum \limits_{j=1}^{r} \omega_{ij} \tilde{\Phi}_{n,j} , & i &= 1, ..., r-1 , \end{aligned}\]

with definitions

\[\begin{aligned} V_{n,i} &= \hphantom{-} \frac{\partial H}{\partial p} (Q_{n,i}, P_{n,i}) , & i &= 1, ..., s , \\ F_{n,i} &= - \frac{\partial H}{\partial q} (Q_{n,i}, P_{n,i}) , & i &= 1, ..., s , \\ U_{n,i} &= \hphantom{-} \frac{\partial \phi}{\partial p} (\tilde{Q}_{n,i}, \tilde{P}_{n,i})^{T} \Lambda_{n,i} , & i &= 1, ..., r , \\ G_{n,i} &= - \frac{\partial \phi}{\partial q} (\tilde{Q}_{n,i}, \tilde{P}_{n,i})^{T} \Lambda_{n,i} , & i &= 1, ..., r , \\ \tilde{\Phi}_{n,i} &= \phi(\tilde{Q}_{n,i}, \tilde{P}_{n,i}) , & i &= 1, ..., r , \end{aligned}\]

and update rule

\[\begin{aligned} q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} V_{n,i} + h \sum \limits_{i=1}^{r} \beta_{i} U_{n,i} , \\ p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} b_{i} F_{n,i} + h \sum \limits_{i=1}^{r} \beta_{i} G_{n,i} , \\ 0 &= \phi (q_{n+1}, p_{n+1}) . \end{aligned}\]

GeometricIntegrators.Integrators.SPARK.IntegratorHSPARKprimary — Type

Specialised Partitioned Additive Runge-Kutta integrator for Hamiltonian systems EXPERIMENTAL.

This integrator solves the following system of equations for the internal stages,

\[\begin{aligned} Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} V_{n,j} + h \sum \limits_{j=1}^{r} \alpha_{ij} U_{n,j} , & i &= 1, ..., s , \\ P_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} a_{ij} F_{n,j} + h \sum \limits_{j=1}^{r} \alpha_{ij} G_{n,j} , & i &= 1, ..., s , \\ \tilde{Q}_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} \tilde{a}_{ij} V_{n,j} + h \sum \limits_{j=1}^{r} \tilde{\alpha}_{ij} U_{n,j} , & i &= 1, ..., r , \\ \tilde{P}_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} \tilde{a}_{ij} F_{n,j} + h \sum \limits_{j=1}^{r} \tilde{\alpha}_{ij} G_{n,j} , & i &= 1, ..., r , \\ 0 &= \sum \limits_{j=1}^{r} \omega_{ij} \tilde{\Phi}_{n,j} , & i &= 1, ..., r-1 , \end{aligned}\]

with definitions

\[\begin{aligned} V_{n,i} &= \hphantom{-} \frac{\partial H}{\partial p} (Q_{n,i}, P_{n,i}) , & i &= 1, ..., s , \\ F_{n,i} &= - \frac{\partial H}{\partial q} (Q_{n,i}, P_{n,i}) , & i &= 1, ..., s , \\ U_{n,i} &= \hphantom{-} \frac{\partial \phi}{\partial p} (\tilde{Q}_{n,i}, \tilde{P}_{n,i})^{T} \Lambda_{n,i} , & i &= 1, ..., r , \\ G_{n,i} &= - \frac{\partial \phi}{\partial q} (\tilde{Q}_{n,i}, \tilde{P}_{n,i})^{T} \Lambda_{n,i} , & i &= 1, ..., r , \\ \tilde{\Phi}_{n,i} &= \phi(\tilde{Q}_{n,i}, \tilde{P}_{n,i}) , & i &= 1, ..., r , \end{aligned}\]

and update rule

\[\begin{aligned} q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} V_{n,i} + h \sum \limits_{i=1}^{r} \beta_{i} U_{n,i} , \\ p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} b_{i} F_{n,i} + h \sum \limits_{i=1}^{r} \beta_{i} G_{n,i} , \\ 0 &= \phi (q_{n+1}, p_{n+1}) . \end{aligned}\]

GeometricIntegrators.Integrators.SPARK.IntegratorHSPARKsecondary — Type

Specialised Partitioned Additive Runge-Kutta integrator for Hamiltonian systems subject to Dirac constraints with projection on secondary constraint EXPERIMENTAL.

This integrator solves the following system of equations for the internal stages,

\[\begin{aligned} Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} V_{n,j} + h \sum \limits_{j=1}^{r} \alpha_{ij} U_{n,j} , & i &= 1, ..., s , \\ P_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} a_{ij} F_{n,j} + h \sum \limits_{j=1}^{r} \alpha_{ij} G_{n,j} , & i &= 1, ..., s , \\ \tilde{Q}_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} \tilde{a}_{ij} V_{n,j} + h \sum \limits_{j=1}^{r} \tilde{\alpha}_{ij} U_{n,j} , & i &= 1, ..., r , \\ \tilde{P}_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} \tilde{a}_{ij} F_{n,j} + h \sum \limits_{j=1}^{r} \tilde{\alpha}_{ij} G_{n,j} , & i &= 1, ..., r , \\ 0 &= \sum \limits_{j=1}^{r} \omega_{ij} \tilde{\Phi}_{n,j} , & i &= 1, ..., r-1 , \end{aligned}\]

with definitions

\[\begin{aligned} V_{n,i} &= \hphantom{-} \frac{\partial H}{\partial p} (Q_{n,i}, P_{n,i}) , & i &= 1, ..., s , \\ F_{n,i} &= - \frac{\partial H}{\partial q} (Q_{n,i}, P_{n,i}) , & i &= 1, ..., s , \\ U_{n,i} &= \hphantom{-} \frac{\partial \phi}{\partial p} (\tilde{Q}_{n,i}, \tilde{P}_{n,i})^{T} \Lambda_{n,i} , & i &= 1, ..., r , \\ G_{n,i} &= - \frac{\partial \phi}{\partial q} (\tilde{Q}_{n,i}, \tilde{P}_{n,i})^{T} \Lambda_{n,i} , & i &= 1, ..., r , \\ \tilde{\Phi}_{n,i} &= \phi(\tilde{Q}_{n,i}, \tilde{P}_{n,i}) , & i &= 1, ..., r , \end{aligned}\]

and update rule

\[\begin{aligned} q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} V_{n,i} + h \sum \limits_{i=1}^{r} \beta_{i} U_{n,i} , \\ p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} b_{i} F_{n,i} + h \sum \limits_{i=1}^{r} \beta_{i} G_{n,i} , \\ 0 &= \phi (q_{n+1}, p_{n+1}) . \end{aligned}\]

GeometricIntegrators.Integrators.SPARK.IntegratorSLRK — Type

Specialised Partitioned Additive Runge-Kutta integrator for degenerate variational systems with projection on secondary constraint.

This integrator solves the following system of equations for the internal stages,

\[\begin{aligned} Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a^1_{ij} V_{n,j} + h \sum \limits_{j=1}^{\sigma} a^2_{ij} \tilde{\Lambda}_{n,j} , & i &= 1, ..., s , \\ P_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} \bar{a}^1_{ij} F^1_{n,j} + h \sum \limits_{j=1}^{s} \bar{a}^2_{ij} F^2_{n,j} + h \sum \limits_{j=1}^{\sigma} \bar{a}^3_{ij} \tilde{F}^3_{n,j} , & i &= 1, ..., s , \\ 0 &= \Phi_{n,i} , & i &= 1, ..., s , \end{aligned}\]

the projective stages

\[\begin{aligned} \tilde{Q}_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} \alpha^1_{ij} \tilde{V}_{n,j} + h \sum \limits_{j=1}^{\sigma} \alpha^2_{ij} \tilde{\Lambda}_{n,j} , & i &= 1, ..., \sigma , \\ \tilde{P}_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} \bar{\alpha}^1_{ij} F^1_{n,j} + h \sum \limits_{j=1}^{s} \bar{\alpha}^2_{ij} \tilde{F}^2_{n,j} + h \sum \limits_{j=1}^{\sigma} \bar{\alpha}^3_{ij} \tilde{F}^3_{n,j} , & i &= 1, ..., \sigma , \\ 0 &= \tilde{\Phi}_{n,i} , & i &= 1, ..., \sigma , \\ 0 &= \sum \limits_{j=1}^{\sigma} \omega_{ij} \tilde{\Psi}_{n,i} , & i &= 1, ..., \sigma-1 , \end{aligned}\]

and update rule

\[\begin{aligned} q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b^1_{i} V_{n,i} + h \sum \limits_{i=1}^{\sigma} b^2_{i} \tilde{\Lambda}_{n,i} , \\ p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} b^1_{i} F^1_{n,i} + h \sum \limits_{i=1}^{s} b^2_{i} F^2_{n,i} + h \sum \limits_{i=1}^{\sigma} b^3_{i} \tilde{F}^3_{n,i} , \\ 0 &= \phi (q_{n+1}, p_{n+1}) , \end{aligned}\]

with definitions

\[\begin{aligned} F^1_{n,i} &= \nabla H (Q_{n,i}) , & i &= 1, ..., s , \\ F^2_{n,i} &= \nabla \vartheta (Q_{n,i}) \cdot V_{n,i} , & i &= 1, ..., s , \\ \tilde{F}^3_{n,i} &= \nabla \phi (\tilde{Q}_{n,i}, \tilde{P}_{n,i}) \cdot \tilde{\Lambda}_{n,i} , & i &= 1, ..., \sigma , \\ \Phi_{n,i} &= \phi (Q_{n,i}, P_{n,i}) = P_{n,i} - \vartheta (Q_{n,i}) , & i &= 1, ..., s , \\ \tilde{\Phi}_{n,i} &= \phi (\tilde{Q}_{n,i}, \tilde{P}_{n,i}) = \tilde{P}_{n,i} - \vartheta (\tilde{Q}_{n,i}) , & i &= 1, ..., \sigma , \\ \tilde{\Psi}_{n,i} &= \psi (\tilde{Q}_{n,i}, \tilde{V}_{n,i}, \tilde{P}_{n,i}, \tilde{F}_{n,i}) = \tilde{F}_{n,i} - \tilde{V}_{n,i} \cdot \nabla \vartheta (\tilde{Q}_{n,i}) , & i &= 1, ..., \sigma , \end{aligned}\]

so that

\[\begin{aligned} F^1_{n,i} + F^2_{n,i} &= \frac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , & i &= 1, ..., s , \\ \phi (Q_{n,i}, P_{n,i}) &= P_{n,i} - \frac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , & i &= 1, ..., s , \\ \psi (\tilde{Q}_{n,i}, \tilde{V}_{n,i}, \tilde{P}_{n,i}, \tilde{F}_{n,i}) &= \tilde{F}_{n,i} - \tilde{V}_{n,i} \cdot \nabla \frac{\partial L}{\partial v} (\tilde{Q}_{n,i}, \tilde{V}_{n,i}) , & i &= 1, ..., \sigma . \end{aligned}\]

GeometricIntegrators.Integrators.SPARK.IntegratorSPARK — Type

Specialised Partitioned Additive Runge-Kutta integrator for index-two DAE systems.

This integrator solves the following system of equations for the internal stages,

\[\begin{aligned} Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} V_{n,j} + h \sum \limits_{j=1}^{r} \alpha_{ij} U_{n,j} , & i &= 1, ..., s , \\ P_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} a_{ij} F_{n,j} + h \sum \limits_{j=1}^{r} \alpha_{ij} G_{n,j} , & i &= 1, ..., s , \\ \tilde{Q}_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} \tilde{a}_{ij} V_{n,j} + h \sum \limits_{j=1}^{r} \tilde{\alpha}_{ij} U_{n,j} , & i &= 1, ..., r , \\ \tilde{P}_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} \tilde{a}_{ij} F_{n,j} + h \sum \limits_{j=1}^{r} \tilde{\alpha}_{ij} G_{n,j} , & i &= 1, ..., r , \\ 0 &= \sum \limits_{j=1}^{r} \omega_{ij} \tilde{\Phi}_{n,j} , & i &= 1, ..., r-1 , \end{aligned}\]

with definitions

\[\begin{aligned} P_{n,i} &= \frac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , & i &= 1, ..., s , \\ F_{n,i} &= \frac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , & i &= 1, ..., s , \\ U_{n,i} &= \hphantom{-} \frac{\partial \phi}{\partial p} (\tilde{Q}_{n,i}, \tilde{P}_{n,i})^{T} \Lambda_{n,i} , & i &= 1, ..., r , \\ G_{n,i} &= - \frac{\partial \phi}{\partial q} (\tilde{Q}_{n,i}, \tilde{P}_{n,i})^{T} \Lambda_{n,i} , & i &= 1, ..., r , \\ \tilde{\Phi}_{n,i} &= \phi(\tilde{Q}_{n,i}, \tilde{P}_{n,i}) , & i &= 1, ..., r , \end{aligned}\]

and update rule

\[\begin{aligned} q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} V_{n,i} + h \sum \limits_{i=1}^{r} \beta_{i} U_{n,i} , \\ p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} b_{i} F_{n,i} + h \sum \limits_{i=1}^{r} \beta_{i} G_{n,i} , \\ 0 &= \phi (q_{n+1}, p_{n+1}) . \end{aligned}\]

GeometricIntegrators.Integrators.SPARK.IntegratorVPARK — Type

Variational partitioned additive Runge-Kutta integrator.

This integrator solves the following system of equations for the internal stages,

\[\begin{aligned} Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} V_{n,j} + h \sum \limits_{j=1}^{r} \alpha_{ij} U_{n,j} , & i &= 1, ..., s , \\ P_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} a_{ij} F_{n,j} + h \sum \limits_{j=1}^{r} \alpha_{ij} G_{n,j} , & i &= 1, ..., s , \\ \tilde{Q}_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} \tilde{a}_{ij} V_{n,j} + h \sum \limits_{j=1}^{r} \tilde{\alpha}_{ij} U_{n,j} , & i &= 1, ..., r , \\ \tilde{P}_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} \tilde{a}_{ij} F_{n,j} + h \sum \limits_{j=1}^{r} \tilde{\alpha}_{ij} G_{n,j} , & i &= 1, ..., r , \\ \tilde{\Phi}_{n,i} &= 0 , & i &= 1, ..., r , \end{aligned}\]

with definitions

\[\begin{aligned} P_{n,i} &= \frac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , & i &= 1, ..., s , \\ F_{n,i} &= \frac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , & i &= 1, ..., s , \\ U_{n,i} &= \hphantom{-} \frac{\partial \phi}{\partial p} (\tilde{Q}_{n,i}, \tilde{P}_{n,i})^{T} \Lambda_{n,i} , & i &= 1, ..., r , \\ G_{n,i} &= - \frac{\partial \phi}{\partial q} (\tilde{Q}_{n,i}, \tilde{P}_{n,i})^{T} \Lambda_{n,i} , & i &= 1, ..., r , \\ \tilde{\Phi}_{n,i} &= \phi(\tilde{Q}_{n,i}, \tilde{P}_{n,i}) , & i &= 1, ..., r , \end{aligned}\]

and update rule

\[\begin{aligned} q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} V_{n,i} + h \sum \limits_{i=1}^{r} \beta_{i} U_{n,i} , \\ p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} b_{i} F_{n,i} + h \sum \limits_{i=1}^{r} \beta_{i} G_{n,i} . \end{aligned}\]

GeometricIntegrators.Integrators.SPARK.IntegratorVSPARK — Type

Specialised Partitioned Additive Runge-Kutta integrator for Variational systems.

This integrator solves the following system of equations for the internal stages,

\[\begin{aligned} Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} V_{n,j} + h \sum \limits_{j=1}^{r} \alpha_{ij} U_{n,j} , & i &= 1, ..., s , \\ P_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} a_{ij} F_{n,j} + h \sum \limits_{j=1}^{r} \alpha_{ij} G_{n,j} , & i &= 1, ..., s , \\ \tilde{Q}_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} \tilde{a}_{ij} V_{n,j} + h \sum \limits_{j=1}^{r} \tilde{\alpha}_{ij} U_{n,j} , & i &= 1, ..., r , \\ \tilde{P}_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} \tilde{a}_{ij} F_{n,j} + h \sum \limits_{j=1}^{r} \tilde{\alpha}_{ij} G_{n,j} , & i &= 1, ..., r , \\ 0 &= \sum \limits_{j=1}^{r} \omega_{ij} \tilde{\Phi}_{n,j} , & i &= 1, ..., r-1 , \end{aligned}\]

with definitions

\[\begin{aligned} P_{n,i} &= \frac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , & i &= 1, ..., s , \\ F_{n,i} &= \frac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , & i &= 1, ..., s , \\ U_{n,i} &= \hphantom{-} \frac{\partial \phi}{\partial p} (\tilde{Q}_{n,i}, \tilde{P}_{n,i})^{T} \Lambda_{n,i} , & i &= 1, ..., r , \\ G_{n,i} &= - \frac{\partial \phi}{\partial q} (\tilde{Q}_{n,i}, \tilde{P}_{n,i})^{T} \Lambda_{n,i} , & i &= 1, ..., r , \\ \tilde{\Phi}_{n,i} &= \phi(\tilde{Q}_{n,i}, \tilde{P}_{n,i}) , & i &= 1, ..., r , \end{aligned}\]

and update rule

\[\begin{aligned} q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} V_{n,i} + h \sum \limits_{i=1}^{r} \beta_{i} U_{n,i} , \\ p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} b_{i} F_{n,i} + h \sum \limits_{i=1}^{r} \beta_{i} G_{n,i} , \\ 0 &= \phi (q_{n+1}, p_{n+1}) . \end{aligned}\]

GeometricIntegrators.Integrators.SPARK.IntegratorVSPARKprimary — Type

Specialised Partitioned Additive Runge-Kutta integrator for Variational systems.

This integrator solves the following system of equations for the internal stages,

\[\begin{aligned} Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} V_{n,j} + h \sum \limits_{j=1}^{r} \alpha_{ij} U_{n,j} , & i &= 1, ..., s , \\ P_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} a_{ij} F_{n,j} + h \sum \limits_{j=1}^{r} \alpha_{ij} G_{n,j} , & i &= 1, ..., s , \\ \tilde{Q}_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} \tilde{a}_{ij} V_{n,j} + h \sum \limits_{j=1}^{r} \tilde{\alpha}_{ij} U_{n,j} , & i &= 1, ..., r , \\ \tilde{P}_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} \tilde{a}_{ij} F_{n,j} + h \sum \limits_{j=1}^{r} \tilde{\alpha}_{ij} G_{n,j} , & i &= 1, ..., r , \\ 0 &= \sum \limits_{j=1}^{r} \omega_{ij} \tilde{\Phi}_{n,j} , & i &= 1, ..., r-1 , \\ 0 &= \sum \limits_{i=1}^{r} \tilde{d}_i \, \Lambda_{n,i} , \end{aligned}\]

with definitions

\[\begin{aligned} P_{n,i} &= \frac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , & i &= 1, ..., s , \\ F_{n,i} &= \frac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , & i &= 1, ..., s , \\ U_{n,i} &= \hphantom{-} \frac{\partial \phi}{\partial p} (\tilde{Q}_{n,i}, \tilde{P}_{n,i})^{T} \Lambda_{n,i} , & i &= 1, ..., r , \\ G_{n,i} &= - \frac{\partial \phi}{\partial q} (\tilde{Q}_{n,i}, \tilde{P}_{n,i})^{T} \Lambda_{n,i} , & i &= 1, ..., r , \\ \tilde{\Phi}_{n,i} &= \phi(\tilde{Q}_{n,i}, \tilde{P}_{n,i}) , & i &= 1, ..., r , \end{aligned}\]

and update rule

\[\begin{aligned} q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} V_{n,i} + h \sum \limits_{i=1}^{r} \beta_{i} U_{n,i} , \\ p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} b_{i} F_{n,i} + h \sum \limits_{i=1}^{r} \beta_{i} G_{n,i} . \end{aligned}\]

GeometricIntegrators.Integrators.SPARK.IntegratorVSPARKsecondary — Type

Specialised Partitioned Additive Runge-Kutta integrator for degenerate variational systems with projection on secondary constraint.

This integrator solves the following system of equations for the internal stages,

\[\begin{aligned} Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a^1_{ij} V_{n,j} + h \sum \limits_{j=1}^{\sigma} a^2_{ij} \tilde{\Lambda}_{n,j} , & i &= 1, ..., s , \\ P_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} \bar{a}^1_{ij} F^1_{n,j} + h \sum \limits_{j=1}^{s} \bar{a}^2_{ij} F^2_{n,j} + h \sum \limits_{j=1}^{\sigma} \bar{a}^3_{ij} \tilde{F}^3_{n,j} , & i &= 1, ..., s , \\ 0 &= \Phi_{n,i} , & i &= 1, ..., s , \end{aligned}\]

the projective stages

\[\begin{aligned} \tilde{Q}_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} \alpha^1_{ij} \tilde{V}_{n,j} + h \sum \limits_{j=1}^{\sigma} \alpha^2_{ij} \tilde{\Lambda}_{n,j} , & i &= 1, ..., \sigma , \\ \tilde{P}_{n,i} &= p_{n} + h \sum \limits_{j=1}^{s} \bar{\alpha}^1_{ij} F^1_{n,j} + h \sum \limits_{j=1}^{s} \bar{\alpha}^2_{ij} \tilde{F}^2_{n,j} + h \sum \limits_{j=1}^{\sigma} \bar{\alpha}^3_{ij} \tilde{F}^3_{n,j} , & i &= 1, ..., \sigma , \\ 0 &= \tilde{\Phi}_{n,i} , & i &= 1, ..., \sigma , \\ 0 &= \sum \limits_{j=1}^{\sigma} \omega_{ij} \tilde{\Psi}_{n,i} , & i &= 1, ..., \sigma-1 , \end{aligned}\]

and update rule

\[\begin{aligned} q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b^1_{i} V_{n,i} + h \sum \limits_{i=1}^{\sigma} b^2_{i} \tilde{\Lambda}_{n,i} , \\ p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} b^1_{i} F^1_{n,i} + h \sum \limits_{i=1}^{s} b^2_{i} F^2_{n,i} + h \sum \limits_{i=1}^{\sigma} b^3_{i} \tilde{F}^3_{n,i} , \\ 0 &= \phi (q_{n+1}, p_{n+1}) , \end{aligned}\]

with definitions

\[\begin{aligned} F^1_{n,i} &= \nabla H (Q_{n,i}) , & i &= 1, ..., s , \\ F^2_{n,i} &= \nabla \vartheta (Q_{n,i}) \cdot V_{n,i} , & i &= 1, ..., s , \\ \tilde{F}^3_{n,i} &= \nabla \phi (\tilde{Q}_{n,i}, \tilde{P}_{n,i}) \cdot \tilde{\Lambda}_{n,i} , & i &= 1, ..., \sigma , \\ \Phi_{n,i} &= \phi (Q_{n,i}, P_{n,i}) = P_{n,i} - \vartheta (Q_{n,i}) , & i &= 1, ..., s , \\ \tilde{\Phi}_{n,i} &= \phi (\tilde{Q}_{n,i}, \tilde{P}_{n,i}) = \tilde{P}_{n,i} - \vartheta (\tilde{Q}_{n,i}) , & i &= 1, ..., \sigma , \\ \tilde{\Psi}_{n,i} &= \psi (\tilde{Q}_{n,i}, \tilde{V}_{n,i}, \tilde{P}_{n,i}, \tilde{F}_{n,i}) = \tilde{F}_{n,i} - \tilde{V}_{n,i} \cdot \nabla \vartheta (\tilde{Q}_{n,i}) , & i &= 1, ..., \sigma , \end{aligned}\]

so that

\[\begin{aligned} F^1_{n,i} + F^2_{n,i} &= \frac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , & i &= 1, ..., s , \\ \phi (Q_{n,i}, P_{n,i}) &= P_{n,i} - \frac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , & i &= 1, ..., s , \\ \psi (\tilde{Q}_{n,i}, \tilde{V}_{n,i}, \tilde{P}_{n,i}, \tilde{F}_{n,i}) &= \tilde{F}_{n,i} - \tilde{V}_{n,i} \cdot \nabla \frac{\partial L}{\partial v} (\tilde{Q}_{n,i}, \tilde{V}_{n,i}) , & i &= 1, ..., \sigma . \end{aligned}\]

GeometricIntegrators.Integrators.SPARK.ParametersHPARK — Type

Parameters for right-hand side function of Hamiltonian Partitioned Additive Runge-Kutta methods.

GeometricIntegrators.Integrators.SPARK.ParametersSLRK — Type

Parameters for right-hand side function of Specialised Partitioned Additive Runge-Kutta methods for Variational systems.

GeometricIntegrators.Integrators.SPARK.ParametersVPARK — Type

Parameters for right-hand side function of Variational Partitioned Additive Runge-Kutta methods.

GeometricIntegrators.Integrators.SPARK.ParametersVSPARKsecondary — Type

Parameters for right-hand side function of Specialised Partitioned Additive Runge-Kutta methods for Variational systems.

GeometricIntegrators.Integrators.SPARK.TableauHPARK — Type

Holds the tableau of a Hamiltonian Partitioned Additive Runge-Kutta methods.

GeometricIntegrators.Integrators.SPARK.TableauSLRK — Type

Holds the tableau of an Specialised Partitioned Additive Runge-Kutta method for Variational systems.

GeometricIntegrators.Integrators.SPARK.TableauSPARK — Type

Holds the tableau of an Specialised Partitioned Additive Runge-Kutta method.

GeometricIntegrators.Integrators.SPARK.TableauVPARK — Type

Holds the tableau of an Variational Partitioned Additive Runge-Kutta method.

GeometricIntegrators.Integrators.SPARK.TableauVSPARKsecondary — Type

Holds the tableau of an Specialised Partitioned Additive Runge-Kutta method for Variational systems.

Base.show — Method

Print additive Runge-Kutta coefficients.

Base.show — Method

Print additive Runge-Kutta coefficients.

Base.show — Method

Print multiplier Runge-Kutta coefficients.

Base.show — Method

Print projective Runge-Kutta coefficients.

Base.show — Method

Print SPARK coefficients.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of specialised partitioned additive Runge-Kutta methods for variational systems.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of specialised partitioned additive Runge-Kutta methods for variational systems.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of specialised partitioned additive Runge-Kutta methods for variational systems.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of specialised partitioned additive Runge-Kutta methods for variational systems.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of specialised partitioned additive Runge-Kutta methods for variational systems.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of variational partitioned additive Runge-Kutta methods.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of specialised partitioned additive Runge-Kutta methods for variational systems.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of variational partitioned additive Runge-Kutta methods.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of specialised partitioned additive Runge-Kutta methods for variational systems.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of specialised partitioned additive Runge-Kutta methods for variational systems.

VPRK Methods

GeometricIntegrators.Integrators.VPRK.AbstractParametersVPRK — Type

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.VPRK.IntegratorCacheVPRK — Type

Variational partitioned Runge-Kutta integrator cache.

Fields

q̃: initial guess of q

p̃: initial guess of p

ṽ: initial guess of v

f̃: initial guess of f

s̃: holds shift due to periodicity of solution

Q: internal stages of q

P: internal stages of p

V: internal stages of v

F: internal stages of f

Y: integral of vector field of internal stages of q

Z: integral of vector field of internal stages of p

GeometricIntegrators.Integrators.VPRK.IntegratorVPRK — Type

Variational partitioned Runge-Kutta integrator.

GeometricIntegrators.Integrators.VPRK.IntegratorVPRKdegenerate — Type

Variational partitioned Runge-Kutta integrator.

GeometricIntegrators.Integrators.VPRK.IntegratorVPRKpInternal — Type

Variational partitioned Runge-Kutta integrator.

GeometricIntegrators.Integrators.VPRK.IntegratorVPRKpLegendre — Type

Variational special partitioned additive Runge-Kutta integrator.

GeometricIntegrators.Integrators.VPRK.IntegratorVPRKpMidpoint — Type

Variational partitioned Runge-Kutta integrator.

GeometricIntegrators.Integrators.VPRK.IntegratorVPRKpSecondary — Type

Variational partitioned Runge-Kutta integrator with projection on secondary constraint.

The VPRK integrator solves the following system of equations for the internal stages,

\[\begin{aligned} Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, \big( V_{n,j} + \Lambda_{n,j} \big) , \\ P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, \big( F_{n,j} + \nabla \vartheta (Q_{n,j})^{T} \cdot \Lambda_{n,j} \big) - d_i \lambda , \\ 0 &= \sum \limits_{i=1}^{s} d_i V_i , \\ 0 &= \sum \limits_{j=1}^{s} \omega_{ij} \Psi_{n,j} , \end{aligned}\]

with definitions

\[\begin{aligned} P_{n,i} &= \dfrac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , \\ F_{n,i} &= \dfrac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , \\ \Psi_{n,i} &= \psi(Q_{n,i}, V_{n,i}, P_{n,i}, F_{n,i}) , \end{aligned}\]

and update rule

\[\begin{aligned} q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, \big( V_{n,i} + \Lambda_{n,i} \big) , \\ p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, \big( F_{n,i} + \nabla \vartheta (Q_{n,j})^{T} \cdot \Lambda_{n,j} \big) , \\ 0 &= \phi (q_{n+1}, p_{n+1}) , \end{aligned}\]

satisfying the symplecticity conditions

\[\begin{aligned} b_{i} \bar{a}_{ij} + \bar{b}_{j} a_{ji} &= b_{i} \bar{b}_{j} , & \bar{b}_i &= b_i , \end{aligned}\]

the primary constraint,

\[\begin{aligned} \phi(q,p) = p - \vartheta (q) = 0 , \end{aligned}\]

at the final solution $(q_{n+1}, p_{n+1})$, and super positions of the secondary constraints,

\[\begin{aligned} \psi(q,\dot{q},p,\dot{p}) = \dot{p} - \dot{q} \cdot \nabla \vartheta (q) = \big( \nabla \vartheta (q) - \nabla \vartheta^{T} (q) \big) \cdot \dot{q} - \nabla H (q) = 0, \end{aligned}\]

which, evaluated at the internal stages, read

\[\begin{aligned} \Psi_{n,j} = \big( \nabla \vartheta (Q_{n,j}) - \nabla \vartheta^{T} (Q_{n,j}) \big) \cdot V_{n,j} - \nabla H (Q_{n,j}) . \end{aligned}\]

Here, $\omega$ is a $(s-1) \times s$ matrix, chosen such that the resulting method has optimal order. The vector $d$ is zero for Gauss-Legendre methods and needs to be chosen appropriately for Gauss-Lobatto methods (for details see documentation of VPRK methods).

GeometricIntegrators.Integrators.VPRK.IntegratorVPRKpStandard — Type

Variational partitioned Runge-Kutta integrator.

GeometricIntegrators.Integrators.VPRK.IntegratorVPRKpStandard — Method

Variational partitioned Runge-Kutta integrator with standard projection.

GeometricIntegrators.Integrators.VPRK.IntegratorVPRKpSymmetric — Type

Variational partitioned Runge-Kutta integrator.

GeometricIntegrators.Integrators.VPRK.IntegratorVPRKpTableau — Type

Projected Variational Gauss-Legendre Runge-Kutta integrator.

GeometricIntegrators.Integrators.VPRK.IntegratorVPRKpVariational — Type

Variational partitioned Runge-Kutta integrator.

GeometricIntegrators.Integrators.VPRK.ParametersVPRK — Type

Parameters for right-hand side function of Variational Partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.VPRK.ParametersVPRKdegenerate — Type

Parameters for right-hand side function of Variational Partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.VPRK.ParametersVPRKpInternal — Type

Parameters for right-hand side function of Variational Partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.VPRK.ParametersVPRKpLegendre — Type

Parameters for right-hand side function of Variational Partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.VPRK.ParametersVPRKpMidpoint — Type

Parameters for right-hand side function of Variational Partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.VPRK.ParametersVPRKpSecondary — Type

Parameters for right-hand side function of Variational Partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.VPRK.ParametersVPRKpStandard — Type

Parameters for right-hand side function of Variational Partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.VPRK.ParametersVPRKpSymmetric — Type

Parameters for right-hand side function of Variational Partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.VPRK.ParametersVPRKpTableau — Type

Parameters for right-hand side function of projected Gauss-Legendre Runge-Kutta methods.

GeometricIntegrators.Integrators.VPRK.ParametersVPRKpVariational — Type

Parameters for right-hand side function of Variational Partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.VPRK.TableauVPRK — Type

TableauVPRK: Tableau of a Variational Partitioned Runge-Kutta method

\[\begin{aligned} P_{n,i} &= \dfrac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} , \\ F_{n,i} &= \dfrac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , & P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} - d_i \lambda , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, F_{n,i} , \\ && 0 &= \sum \limits_{i=1}^{s} d_i V_i , && \end{aligned}\]

satisfying the symplecticity conditions

\[\begin{aligned} b_{i} \bar{a}_{ij} + b_{j} a_{ji} &= b_{i} b_{j} , & \bar{b}_i &= b_i . \end{aligned}\]

GeometricIntegrators.Integrators.VPRK.IntegratorVPRKpSymplectic — Method

Variational partitioned Runge-Kutta integrator with symplectic projection.

GeometricIntegrators.Integrators.VPRK.IntegratorVPRKpVariationalP — Method

Variational partitioned Runge-Kutta integrator with variational projection on $(p_{n}, q_{n+1})$.

GeometricIntegrators.Integrators.VPRK.IntegratorVPRKpVariationalQ — Method

Variational partitioned Runge-Kutta integrator with variational projection on $(q_{n}, p_{n+1})$.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of variational partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.function_stages! — Method

Compute solution of degenerate symplectic partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of variational partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of variational special partitioned additive Runge-Kutta methods.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of variational partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of variational partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of projected variational partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of variational partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of projected variational partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of projected Gauss-Legendre Runge-Kutta methods.

Galerkin Variational Integrators

GeometricIntegrators.Integrators.IntegratorCGVI — Type

Continuous Galerkin Variational Integrator.

GeometricIntegrators.Integrators.ParametersCGVI — Type

ParametersCGVI: Parameters for right-hand side function of continuous Galerkin variational Integrator.

Parameters

Θ: function of the noncanonical one-form (∂L/∂v)

f: function of the force (∂L/∂q)

Δt: time step

b: weights of the quadrature rule

c: nodes of the quadrature rule

x: nodes of the basis

m: mass matrix

a: derivative matrix

r₀: reconstruction coefficients at the beginning of the interval

r₁: reconstruction coefficients at the end of the interval

t: initial time

q: solution of q at time t

p: solution of p at time t

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of variational partitioned Runge-Kutta methods.

GeometricIntegrators.Integrators.IntegratorCacheDGVI — Type

Nonlinear function cache for Discontinuous Galerkin Variational Integrator.

Parameters

ST: data type

D: number of dimensions

S: number of degrees of freedom

R: number of nodes of quadrature formula

Fields

X: degrees of freedom

Q: solution at quadrature nodes

V: velocity at quadrature nodes

P: one-form at quadrature nodes

F: forces at quadrature nodes

q: current solution of $q_{n}$

q⁻: current solution of $q_{n}^{-}$

q⁺: current solution of $q_{n}^{+}$

q̄: current solution of $q_{n+1}$

q̄⁻: current solution of $q_{n+1}^{-}$

q̄⁺: current solution of $q_{n+1}^{+}$

ϕ: average of the solution at $t_{n}$

ϕ̅: average of the solution at $t_{n+1}$

λ: jump of the solution at $t_{n}$

λ̄: jump of the solution at $t_{n+1}$

θ: one-form evaluated across at $t_{n}$

Θ̅: one-form evaluated across at $t_{n+1}$

g: projection evaluated across at $t_{n}$

ḡ: projection evaluated across at $t_{n+1}$

GeometricIntegrators.Integrators.IntegratorDGVI — Type

IntegratorDGVI: Discontinuous Galerkin Variational Integrator.

The DGVI integrators arise from the discretization of the action integral

\[\mathcal{A} [q] = \int \limits_{0}^{T} L(q(t), \dot{q}(t)) \, dt ,\]

with $L$ a fully degenerate Lagrangian of the form

\[L(q, \dot{q}) = \vartheta (q) \cdot \dot{q} - H(q) ,\]

where $\vartheta (q)$ denotes the Cartan one-form and $H(q)$ the Hamiltonian, which is usually given by the total energy of the system.

Discretization

Within each interval $(t_{n}, t_{n+1})$ a piecewise-polynomial approximation $q_h$ of the trajectory $q$ is constructed using $S$ basis functions $\varphi_{i}$,

\[q_h(t) \vert_{(t_{n}, t_{n+1})} = \sum \limits_{i=1}^{S} x_{n,i} \, \bar{\varphi}_{n,i} (t) ,\]

where $\bar{\varphi}_{n,i} (t)$ is a rescaled basis function, defined by

\[\bar{\varphi}_{n,i} (t) = \varphi_{i} \bigg( \frac{t - t_{n}}{t_{n+1} - t_{n}} \bigg) ,\]

and it is assumed that $\varphi_{i}$ is compactly supported on $[0,1]$. These approximations $q_h(t)$ are not assumed to be continuous across interval boundaries $t_{n}$ but usuaslly have jumps.

The integral over $(t_{n}, t_{n+1})$ is approximated by a quadrature rule with $R$ nodes $c_i$ and weights $b_i$. Denote by $m$ and $a$ mass and derivative matrices, respectively, whose elements are given by

\[m_{ij} = \varphi_j (c_i) , \qquad a_{ij} = \varphi_j' (c_i) , \qquad i = 1, ..., R , \; j = 1, ..., S .\]

With that, the solution and its time derivative at the quadrature points can be written as

\[Q_{n,i} \equiv q_h(t_n + c_i h) = m_{ij} x_{n,j} , \qquad V_{n,i} \equiv \dot{q}_h (t_n + c_i h) = a_{ij} x_{n,j} ,\]

where

\[x_{n} = ( x_{n,1}, ..., x_{n,S} )^T\]

is the vector containing the degrees of freedom of $q_h \vert_{[t_{n}, t_{n+1}]}$. The limits of $q_h(t)$ at $t_{n}$ and $t_{n+1}$ are given by

\[q_{n}^{+} = \lim \limits_{t \downarrow t_{n}} q_h(t) = \sum \limits_{j=1}^{S} r^{+}_{j} \delta x_{n,j} , \hspace{3em} q_{n+1}^{-} = \lim \limits_{t \uparrow t_{n+1}} q_h(t) = \sum \limits_{j=1}^{S} r^{-}_{j} \delta x_{n,j} .\]

The discrete action reads

\[\mathcal{A}_d [x_d] = h \sum \limits_{n=0}^{N-1} \bigg[ \sum \limits_{i=1}^{R} b_i \big[ \vartheta (Q_{n,i}) \cdot V_{n,i} - H(Q_{n,i}) \big] \\ + \frac{\vartheta (q_n) + \vartheta (q_n^+)}{2} \cdot (q_n^+ - q_n) + \frac{\vartheta (q_{n+1}^-) + \vartheta (q_{n+1})}{2} \cdot (q_{n+1} - q_{n+1}^-) \bigg] ,\]

so that using the relations

\[\delta Q_{n,i} = m_{ij} \delta x_{n,j} , \qquad \delta V_{n,i} = \frac{a_{ij}}{h} \delta x_{n,j} , \qquad \delta q_{n}^- = r^{-}_{j} \delta x_{n-1,j} , \qquad \delta q_{n}^+ = r^{+}_{j} \delta x_{n,j} ,\]

the discrete action principle leads to the discrete equations of motion,

\[0 = \sum \limits_{i=1}^{R} b_i \big[ h m_{ij} \nabla \vartheta (Q_{n,i}) \cdot V_{n,i} + a_{ij} \vartheta (Q_{n,i}) - h m_{ij} \nabla H(Q_{n,i}) \big] \\ + r^{+}_{j} \, \frac{ \vartheta ( q_{n } ) + \vartheta( q_{n }^+ ) }{2} - r^{-}_{j} \, \frac{ \vartheta ( q_{n+1}^- ) + \vartheta( q_{n+1} ) }{2} \big] \\ + h r^{+}_{j} \, \nabla \vartheta (q_{n }^+) \cdot (q_{n }^+ - q_{n } ) + h r^{-}_{j} \, \nabla \vartheta (q_{n+1}^-) \cdot (q_{n+1} - q_{n+1}^-) ,\]

and

\[\vartheta(q_{n}^+) = \vartheta (q_{n}^-) + \nabla \vartheta (q_{n}) \cdot (q_{n}^+ - q_{n}^-) ,\]

for all $n$ and all $j$. Let us introduce the variable $p_n$ as

\[p_{n} = \vartheta (q_{n}^-) + \nabla \vartheta (q_{n}) \cdot (q_{n} - q_{n}^-) ,\]

so that

\[\vartheta(q_{n}^+) = p_{n} + \nabla \vartheta (q_{n}) \cdot (q_{n}^+ - q_{n}) .\]

Then the above equations provide a map $(q_{n}, p_{n}) \mapsto (q_{n+1}, p_{n+1})$. In order to solve these equations, initial conditions $q_{0}$ and $p_{0} = \vartheta(q_{0})$ have to be prescribed.

Fields

equation: Implicit Ordinary Differential Equation

basis: piecewise polynomial basis

quadrature: numerical quadrature rule

Δt: time step

params: ParametersDGVI

solver: nonlinear solver

iguess: initial guess

q: current solution vector for trajectory

p: current solution vector for one-form

cache: temporary variables for nonlinear solver

GeometricIntegrators.Integrators.ParametersDGVI — Type

ParametersDGVI: Parameters for right-hand side function of Discontinuous Galerkin Variational Integrator.

Parameters

DT: data type

TT: parameter type

D: dimension of the system

S: number of basis nodes

R: number of quadrature nodes

Fields

Θ: function of the noncanonical one-form (∂L/∂v)

f: function of the force (∂L/∂q)

g: function of the projection ∇ϑ(q)⋅v

Δt: time step

b: quadrature weights

c: quadrature nodes

m: mass matrix

a: derivative matrix

r⁻: reconstruction coefficients, jump lhs value

r⁺: reconstruction coefficients, jump rhs value

t: initial time

q: solution of q at time t

q⁻: solution of q⁻ at time t

q⁺: solution of q⁺ at time t

GeometricIntegrators.Integrators.compute_stages_p! — Method

Compute one-form and forces at quadrature nodes.

GeometricIntegrators.Integrators.compute_stages_q! — Method

Compute solution at quadrature nodes and across jump.

GeometricIntegrators.Integrators.compute_stages_v! — Method

Compute velocities at quadrature nodes.

GeometricIntegrators.Integrators.function_stages! — Method

Compute stages of variational partitioned Runge-Kutta methods.