Integrators

Integrators

GeometricIntegrators.Integrators.AbstractTableau — Type.

Holds the information for the various methods' tableaus.

source

GeometricIntegrators.Integrators.AbstractTableauIRK — Type.

Holds the tableau of an implicit Runge-Kutta method.

source

GeometricIntegrators.Integrators.AbstractTableauPRK — Type.

Holds the tableau of a partitioned Runge-Kutta method.

source

GeometricIntegrators.Integrators.AbstractTableauRK — Type.

Holds the tableau of a Runge-Kutta method.

source

GeometricIntegrators.Integrators.CoefficientsARK — Type.

Holds the coefficients of an additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.CoefficientsMRK — Type.

Holds the multiplier Runge-Kutta coefficients.

source

GeometricIntegrators.Integrators.CoefficientsPGLRK — Type.

Holds the coefficients of a projected Gauss-Legendre Runge-Kutta method.

source

GeometricIntegrators.Integrators.CoefficientsPRK — Type.

Holds the coefficients of a projective Runge-Kutta method.

source

GeometricIntegrators.Integrators.CoefficientsRK — Type.

Holds the coefficients of a Runge-Kutta method.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for special additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Print error for integrators not implemented, yet.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for variational partitioned additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for variational special partitioned additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for Projected Gauss-Legendre Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for variational partitioned Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for diagonally implicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for explicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for fully implicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for partitioned additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for special partitioned additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for explicit partitioned Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for implicit partitioned Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for stochastic fully implicit partitioned Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for stochastic explicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for stochastic fully implicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for weak explicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for weak fully implicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for splitting tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for stochastic fully implicit split partitioned Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for formal Lagrangian Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.IntegratorCGVI — Type.

Continuous Galerkin Variational Integrator.

source

GeometricIntegrators.Integrators.IntegratorDGVI — Type.

IntegratorDGVI: Discontinuous Galerkin Variational Integrator.

Parameters

Fields

equation: Implicit Ordinary Differential Equation

basis: piecewise polynomial basis

quadrature: numerical quadrature rule

Δt: time step

params: ParametersDGVI

solver: nonlinear solver

iguess: initial guess

q: current solution vector for trajectory

p: current solution vector for one-form

cache: temporary variables for nonlinear solver

source

GeometricIntegrators.Integrators.IntegratorDGVIEXP — Type.

IntegratorDGVIEXP: Discontinuous Galerkin Variational Integrator.

Parameters

Fields

equation: Implicit Ordinary Differential Equation

basis: piecewise polynomial basis

quadrature: numerical quadrature rule

Δt: time step

params: ParametersDGVIEXP

solver: nonlinear solver

iguess: initial guess

q: current solution vector for trajectory

p: current solution vector for one-form

cache: temporary variables for nonlinear solver

source

GeometricIntegrators.Integrators.IntegratorDGVIP0 — Type.

IntegratorDGVIP0: Discontinuous Galerkin Variational Integrator.

Parameters

Fields

equation: Implicit Ordinary Differential Equation

basis: piecewise polynomial basis

quadrature: numerical quadrature rule

Δt: time step

params: ParametersDGVIP0

solver: nonlinear solver

iguess: initial guess

q: current solution vector for trajectory

p: current solution vector for one-form

cache: temporary variables for nonlinear solver

source

GeometricIntegrators.Integrators.IntegratorDGVIP1 — Type.

IntegratorDGVIP1: Discontinuous Galerkin Variational Integrator.

Parameters

Fields

equation: Implicit Ordinary Differential Equation

basis: piecewise polynomial basis

quadrature: numerical quadrature rule

Δt: time step

params: ParametersDGVIP1

solver: nonlinear solver

iguess: initial guess

q: current solution vector for trajectory

p: current solution vector for one-form

cache: temporary variables for nonlinear solver

source

GeometricIntegrators.Integrators.IntegratorDGVIPI — Type.

IntegratorDGVIPI: Discontinuous Galerkin Variational Integrator.

Parameters

Fields

equation: Implicit Ordinary Differential Equation

basis: piecewise polynomial basis

quadrature: numerical quadrature rule

jump: jump across discontinuity

Δt: time step

params: ParametersDGVIPI

solver: nonlinear solver

iguess: initial guess

q: current solution vector for trajectory

q⁻: current solution vector for trajectory, lhs of jump

q⁺: current solution vector for trajectory, rhs of jump

cache: temporary variables for nonlinear solver

source

GeometricIntegrators.Integrators.IntegratorDIRK — Type.

Diagonally implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorEPRK — Type.

Explicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorERK — Type.

Explicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorFIRK — Type.

Fully implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorFLRK — Type.

Formal Lagrangian Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorGPARK — Type.

Special Partitioned Additive Runge Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorHSPARK — Type.

Hamiltonian Specialised Partitioned Additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorIPRK — Type.

Implicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorPARK — Type.

Implicit partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorPGLRK — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSERK — Type.

Stochastic Explicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSIPRK — Type.

Stochastic implicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSIRK — Type.

Stochastic implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSISPRK — Type.

Stochastic implicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSPARK — Type.

Variational special partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSplitting — Type.

Splitting integrator.

source

GeometricIntegrators.Integrators.IntegratorSplitting — Method.

Construct splitting integrator for symmetric splitting tableau with general stages.

source

GeometricIntegrators.Integrators.IntegratorSplitting — Method.

Construct splitting integrator for non-symmetric splitting tableau with general stages.

source

GeometricIntegrators.Integrators.IntegratorSplitting — Method.

Construct splitting integrator for symmetric splitting tableau with symmetric stages.

source

GeometricIntegrators.Integrators.IntegratorVPARK — Type.

Variational partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRK — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKdegenerate — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpLegendre — Type.

Variational special partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpMidpoint — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpSecondary — Type.

Variational partitioned Runge-Kutta integrator with projection on secondary constraint.

\[\begin{align*} P_{n,i} &= \dfrac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, \big( V_{n,j} + \Lambda_{n,j} \big) , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, \big( V_{n,i} + \Lambda_{n,i} \big) , \\ F_{n,i} &= \dfrac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , & P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, \big( F_{n,j} + \nabla \vartheta (Q_{n,j}) \cdot \Lambda_{n,j} \big) - d_i \lambda , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, \big( F_{n,i} + \nabla \vartheta (Q_{n,j}) \cdot \Lambda_{n,j} \big) , \\ 0 &= \sum \limits_{i=1}^{s} d_i V_i , & 0 &= \sum \limits_{j=1}^{s} \omega_{ij} \Psi_{n,j} , & 0 &= \phi (q_{n+1}, p_{n+1}) , \end{align*}\]

satisfying the symplecticity conditions

\[\begin{align*} b_{i} \bar{a}_{ij} + \bar{b}_{j} a_{ji} &= b_{i} \bar{b}_{j} , & \bar{b}_i &= b_i , \end{align*}\]

the primary constraint,

\[\begin{align*} \phi(q,p) = p - \vartheta (q) = 0 , \end{align*}\]

at the final solution $(q_{n+1}, p_{n+1})$, and super positions of the secondary constraints,

\[\begin{align*} \psi(q,\dot{q},p,\dot{p}) = \dot{p} - \dot{q} \cdot \nabla \vartheta (q) = \big( \nabla \vartheta (q) - \nabla \vartheta^{T} (q) \big) \cdot \dot{q} - \nabla H (q) = 0, \end{align*}\]

at the internal stages,

\[\begin{align*} \Psi_{n,j} = \big( \nabla \vartheta (Q_{n,j}) - \nabla \vartheta^{T} (Q_{n,j}) \big) \cdot V_{n,j} - \nabla H (Q_{n,j}) . \end{align*}\]

Here, $\omega$ is a $(s-1) \times s$ matrix, chosen such that the resulting method has optimal order. The vector $d$ is zero for Gauss-Legendre methods and needs to be chosen appropriately for Gauss-Lobatto methods (for details see documentation of VPRK methods).

source

GeometricIntegrators.Integrators.IntegratorVPRKpStandard — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpSymmetric — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpVariational — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVSPARK — Type.

Variational Specialised Partitioned Additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorWERK — Type.

Stochastic Explicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorWIRK — Type.

Stochastic implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.TableauARK — Type.

Holds the tableau of a additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauDIRK — Type.

Holds the tableau of a diagonally implicit Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauEPRK — Type.

TableauEPRK: Tableau of an Explicit Partitioned Runge-Kutta method

\[\begin{align*} V_{n,i} &= \hphantom{-} \dfrac{\partial H}{\partial p} (Q_{n,i}, P_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} , \\ F_{n,i} &= - \dfrac{\partial H}{\partial q} (Q_{n,i}, P_{n,i}) , & P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, F_{n,i} , \end{align*}\]

usually satisfying the symplecticity conditions

\[\begin{align*} b_{i} \bar{a}_{ij} + b_{j} a_{ji} &= b_{i} b_{j} , & \bar{b}_i &= b_i . \end{align*}\]

source

GeometricIntegrators.Integrators.TableauERK — Type.

Holds the tableau of an explicit Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauFIRK — Type.

Holds the tableau of a fully implicit Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauGLM — Type.

Holds the tableau of a general linear method.

source

GeometricIntegrators.Integrators.TableauGPARK — Type.

Holds the tableau of a spezialized partitioned additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauHSPARK — Type.

Holds the tableau of an Hamiltonian Specialised Partitioned Additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauIPRK — Type.

TableauIPRK: Tableau of an Implicit Partitioned Runge-Kutta method

\[\begin{align*} P_{n,i} &= \dfrac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} , \\ F_{n,i} &= \dfrac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , & P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, F_{n,i} , \end{align*}\]

usually satisfying the symplecticity conditions

\[\begin{align*} b_{i} \bar{a}_{ij} + b_{j} a_{ji} &= b_{i} b_{j} , & \bar{b}_i &= b_i . \end{align*}\]

source

GeometricIntegrators.Integrators.TableauPARK — Type.

Holds the tableau of an partitioned additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauSARK — Type.

Holds the tableau of a spezialized additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauSERK — Type.

Holds the tableau of a stochastic explicit Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauSIPRK — Type.

Holds the tableau of a stochastic implicit partitioned Runge-Kutta method. qdrift, pdrift hold the RK coefficients for the drift part, and qdiff, pdiff hold the RK coefficients for the diffusion part of the SDE.

source

GeometricIntegrators.Integrators.TableauSIRK — Type.

Holds the tableau of a stochastic implicit Runge-Kutta method. qdrift holds the RK coefficients for the drift part, and qdiff holds the RK coefficients for the diffusion part of the SDE.

source

GeometricIntegrators.Integrators.TableauSISPRK — Type.

Holds the tableau of a stochastic implicit split partitioned Runge-Kutta method. qdrift, pdrift1, pdrift2 hold the RK coefficients for the drift parts, and qdiff, pdiff1, pdiff2 hold the RK coefficients for the diffusion part of the SDE.

source

GeometricIntegrators.Integrators.TableauSPARK — Type.

Holds the tableau of an Specialised Partitioned Additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauSplittingGS — Type.

Tableau for symmetric splitting methods with general stages.

source

GeometricIntegrators.Integrators.TableauSplittingNS — Type.

Tableau for non-symmetric splitting methods.

source

GeometricIntegrators.Integrators.TableauSplittingSS — Type.

Tableau for symmetric splitting methods with symmetric stages.

source

GeometricIntegrators.Integrators.TableauVPARK — Type.

Holds the tableau of an variational partitioned additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauVPRK — Type.

TableauVPRK: Tableau of a Variational Partitioned Runge-Kutta method

\[\begin{align*} P_{n,i} &= \dfrac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} , \\ F_{n,i} &= \dfrac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , & P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} - d_i \lambda , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, F_{n,i} , \\ && 0 &= \sum \limits_{i=1}^{s} d_i V_i , && \end{align*}\]

satisfying the symplecticity conditions

\[\begin{align*} b_{i} \bar{a}_{ij} + b_{j} a_{ji} &= b_{i} b_{j} , & \bar{b}_i &= b_i . \end{align*}\]

source

GeometricIntegrators.Integrators.TableauVSPARK — Type.

Holds the tableau of an Variational Specialised Partitioned Additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauWERK — Type.

Holds the tableau of a weak explicit Runge-Kutta method.

According to Andreas Rossler, "Second order Runge-Kutta methods for Stratonovich stochastic differential equations", BIT Numerical Mathematics (2007) 47, equation (5.1)

source

GeometricIntegrators.Integrators.TableauWIRK — Type.

Holds the tableau of a weak implicit Runge-Kutta method.

According to Wang, Hong, Xu, "Construction of Symplectic Runge-Kutta Methods for Stochastic Hamiltonian Systems", Commun. Comput. Phys. 21(1), 2017

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of fully implicit Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of fully implicit Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of implicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute solution of degenerate symplectic partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of projected variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of projected variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational special partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of formal Lagrangian Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of implicit Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of stochastic implicit Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of stochastic implicit split partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of weak implicit Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of Hamiltonian Specialised Partitioned Additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational special partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.initialize! — Method.

Initialize stochastic integrator for the sample paths k with k₁ ≤ k ≤ k₂, initial conditions m with m₁ ≤ m ≤ m₂ and time step 0.

source

GeometricIntegrators.Integrators.initialize! — Method.

Initialize integrator for initial conditions m with m₁ ≤ m ≤ m₂ and time step 0.

source

GeometricIntegrators.Integrators.integrate — Function.

Apply integrator for ntime time steps and return solution.

source

GeometricIntegrators.Integrators.integrate — Function.

Integrate given equation with given tableau for ntime time steps and return solution.

source

GeometricIntegrators.Integrators.integrate — Function.

Integrate ODE specified by vector field and initial condition with given tableau for ntime time steps and return solution.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate SDE for the sample paths k with k₁ ≤ k ≤ k₂ and initial conditions m with m₁ ≤ m ≤ m₂.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate SDE for all sample paths and initial conditions.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE for initial conditions m with m₁ ≤ m ≤ m₂.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE for all initial conditions.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate partitioned DAE with Special Additive Runge Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE for initial conditions m with m₁ ≤ m ≤ m₂ for time steps n with n₁ ≤ n ≤ n₂.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate SDE for the sample paths k with k₁ ≤ k ≤ k₂, the initial conditions m with m₁ ≤ m ≤ m₂, and the time steps n with n₁ ≤ n ≤ n₂.

source

GeometricIntegrators.Integrators.readTableauERKFromFile — Method.

Read explicit Runge-Kutta tableau from file.

source

GeometricIntegrators.Integrators.writeTableauToFile — Method.

Write Runge-Kutta tableau to file.

source

GeometricIntegrators.Integrators.AbstractTableauERK — Type.

Holds the tableau of an explicit Runge-Kutta method.

source

GeometricIntegrators.Integrators.NonlinearFunctionCacheDGVI — Type.

Nonlinear function cache for Discontinuous Galerkin Variational Integrator.

Parameters

ST: data type

D: number of dimensions

S: number of degrees of freedom

R: number of nodes of quadrature formula

Fields

X: degrees of freedom

Q: solution at quadrature nodes

V: velocity at quadrature nodes

P: one-form at quadrature nodes

F: forces at quadrature nodes

q: current solution of qₙ

q⁻: current solution of qₙ⁻

q⁺: current solution of qₙ⁺

q̅: current solution of qₙ₊₁

q̅⁻: current solution of qₙ₊₁⁻

q̅⁺: current solution of qₙ₊₁⁺

ϕ: average of the solution at tₙ

ϕ̅: average of the solution at tₙ₊₁

λ: jump of the solution at tₙ

λ̅: jump of the solution at tₙ₊₁

θ: one-form evaluated across at tₙ

Θ̅: one-form evaluated across at tₙ₊₁

g: projection evaluated across at tₙ

g̅: projection evaluated across at tₙ₊₁

source

GeometricIntegrators.Integrators.NonlinearFunctionCacheDGVIPI — Type.

Nonlinear function cache for Discontinuous Galerkin Variational Integrator.

Parameters

X: degrees of freedom

Q: solution at quadrature nodes

V: velocity at quadrature nodes

P: one-form at quadrature nodes

F: forces at quadrature nodes

q: current solution of qₙ

q⁻: current solution of qₙ⁻

q⁺: current solution of qₙ⁺

q̅: current solution of qₙ₊₁

q̅⁻: current solution of qₙ₊₁⁻

q̅⁺: current solution of qₙ₊₁⁺

λ: jump of the solution at tₙ

λ̅: jump of the solution at tₙ₊₁

ϕ: solution evaluated across the jump at tₙ

ϕ̅: solution evaluated across the jump at tₙ₊₁

θ: one-form evaluated across the jump at tₙ

Θ̅: one-form evaluated across the jump at tₙ₊₁

g: projection evaluated across the jump at tₙ

g̅: projection evaluated across the jump at tₙ₊₁

source

GeometricIntegrators.Integrators.NonlinearFunctionCacheSIPRK — Type.

Structure for holding the internal stages Q, the values of the drift vector and the diffusion matrix evaluated at the internal stages VQ=v(Q), BQ=B(Q), and the increments Y = Δta_driftv(Q) + a_diffB(Q)ΔW

source

GeometricIntegrators.Integrators.NonlinearFunctionCacheSIRK — Type.

Structure for holding the internal stages Q, the values of the drift vector and the diffusion matrix evaluated at the internal stages VQ=v(Q), BQ=B(Q), and the increments Y = Δta_driftv(Q) + a_diffB(Q)ΔW

source

GeometricIntegrators.Integrators.NonlinearFunctionCacheSISPRK — Type.

Structure for holding the internal stages Q, the values of the drift vector and the diffusion matrix evaluated at the internal stages VQ=v(Q), BQ=B(Q), and the increments Y = Δta_driftv(Q) + a_diffB(Q)ΔW

source

GeometricIntegrators.Integrators.ParametersCGVI — Type.

ParametersCGVI: Parameters for right-hand side function of continuous Galerkin variational Integrator.

Parameters

Θ: function of the noncanonical one-form (∂L/∂v)

f: function of the force (∂L/∂q)

Δt: time step

b: weights of the quadrature rule

c: nodes of the quadrature rule

x: nodes of the basis

m: mass matrix

a: derivative matrix

r₀: reconstruction coefficients at the beginning of the interval

r₁: reconstruction coefficients at the end of the interval

t: current time

q: current solution of q

p: current solution of p

source

GeometricIntegrators.Integrators.ParametersDGVI — Type.

ParametersDGVI: Parameters for right-hand side function of Discontinuous Galerkin Variational Integrator.

Parameters

DT: data type

TT: parameter type

D: dimension of the system

S: number of basis nodes

R: number of quadrature nodes

Fields

Θ: function of the noncanonical one-form (∂L/∂v)

f: function of the force (∂L/∂q)

g: function of the projection ∇ϑ(q)⋅v

Δt: time step

b: quadrature weights

c: quadrature nodes

m: mass matrix

a: derivative matrix

r⁻: reconstruction coefficients, jump lhs value

r⁺: reconstruction coefficients, jump rhs value

t: current time

q: current solution of qₙ

q⁻: current solution of qₙ⁻

q⁺: current solution of qₙ⁺

source

GeometricIntegrators.Integrators.ParametersDGVIEXP — Type.

ParametersDGVIEXP: Parameters for right-hand side function of Discontinuous Galerkin Variational Integrator.

Parameters

DT: data type

TT: parameter type

D: dimension of the system

S: number of basis nodes

R: number of quadrature nodes

Fields

Θ: function of the noncanonical one-form (∂L/∂v)

f: function of the force (∂L/∂q)

g: function of the projection ∇ϑ(q)⋅v

Δt: time step

b: quadrature weights

c: quadrature nodes

m: mass matrix

a: derivative matrix

r⁻: reconstruction coefficients, jump lhs value

r⁺: reconstruction coefficients, jump rhs value

t: current time

q: current solution of qₙ

q⁻: current solution of qₙ⁻

q⁺: current solution of qₙ⁺

source

GeometricIntegrators.Integrators.ParametersDGVIP0 — Type.

ParametersDGVIP0: Parameters for right-hand side function of Discontinuous Galerkin Variational Integrator.

Parameters

DT: data type

TT: parameter type

D: dimension of the system

S: number of basis nodes

R: number of quadrature nodes

Fields

Θ: function of the noncanonical one-form (∂L/∂v)

f: function of the force (∂L/∂q)

g: function of the projection ∇ϑ(q)⋅v

Δt: time step

b: quadrature weights

c: quadrature nodes

m: mass matrix

a: derivative matrix

r⁻: reconstruction coefficients, jump lhs value

r⁺: reconstruction coefficients, jump rhs value

t: current time

q: current solution of qₙ

q⁻: current solution of qₙ⁻

θ: one-form ϑ evaluated on qₙ

θ⁻: one-form ϑ evaluated on qₙ⁻

source

GeometricIntegrators.Integrators.ParametersDGVIP1 — Type.

ParametersDGVIP1: Parameters for right-hand side function of Discontinuous Galerkin Variational Integrator.

Parameters

DT: data type

TT: parameter type

D: dimension of the system

S: number of basis nodes

R: number of quadrature nodes

Fields

Θ: function of the noncanonical one-form (∂L/∂v)

f: function of the force (∂L/∂q)

g: function of the projection ∇ϑ(q)⋅v

Δt: time step

b: quadrature weights

c: quadrature nodes

m: mass matrix

a: derivative matrix

r⁻: reconstruction coefficients, jump lhs value

r⁺: reconstruction coefficients, jump rhs value

t: current time

q: current solution of qₙ

q⁻: current solution of qₙ⁻

q⁺: current solution of qₙ⁺

source

GeometricIntegrators.Integrators.ParametersDGVIPI — Type.

ParametersDGVIPI: Parameters for right-hand side function of Discontinuous Galerkin Variational Integrator with Path Integral approximation of the jump.

Parameters

DT: data type

TT: parameter type

D: dimension of the system

S: number of basis nodes

QR: number of quadrature nodes

FR: number of quadrature nodes for the discontinuity

Fields

Θ: function of the noncanonical one-form (∂L/∂v)

f: function of the force (∂L/∂q)

g: function of the projection ∇ϑ(q)⋅v

Δt: time step

b: quadrature weights

c: quadrature nodes

m: mass matrix

a: derivative matrix

r⁻: reconstruction coefficients, jump lhs value

r⁺: reconstruction coefficients, jump rhs value

β: weights of the quadrature rule for the discontinuity

γ: nodes of the quadrature rule for the discontinuity

μ⁻: mass vector for the lhs jump value

μ⁺: mass vector for the rhs jump value

α⁻: derivative vector for the discontinuity lhs value

α⁺: derivative vector for the discontinuity rhs value

ρ⁻: reconstruction coefficients for central jump value, lhs value

ρ⁺: reconstruction coefficients for central jump value, rhs value

t: current time

q: current solution of qₙ

q⁻: current solution of qₙ⁻

q⁺: current solution of qₙ⁺

source

GeometricIntegrators.Integrators.ParametersDIRK — Type.

Parameters for right-hand side function of diagonally implicit Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersFIRK — Type.

Parameters for right-hand side function of fully implicit Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersFLRK — Type.

Parameters for right-hand side function of formal Lagrangian Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersHSPARK — Type.

Parameters for right-hand side function of Hamiltonian Specialised Partitioned Additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersIPRK — Type.

Parameters for right-hand side function of implicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersPARK — Type.

Parameters for right-hand side function of partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersPGLRK — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersSIPRK — Type.

Parameters for right-hand side function of implicit Runge-Kutta methods. A - if positive, the upper bound of the Wiener process increments; if A=0.0, no truncation

source

GeometricIntegrators.Integrators.ParametersSIRK — Type.

Parameters for right-hand side function of implicit Runge-Kutta methods. A - if positive, the upper bound of the Wiener process increments; if A=0.0, no truncation

source

GeometricIntegrators.Integrators.ParametersSISPRK — Type.

Parameters for right-hand side function of implicit Runge-Kutta methods. A - if positive, the upper bound of the Wiener process increments; if A=0.0, no truncation

source

GeometricIntegrators.Integrators.ParametersSPARK — Type.

Parameters for right-hand side function of Specialised Partitioned Additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersVPARK — Type.

Parameters for right-hand side function of variational partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersVPRK — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersVPRKdegenerate — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersVPRKpInternal — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersVPRKpLegendre — Type.

Parameters for right-hand side function of variational special partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersVPRKpMidpoint — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersVPRKpSecondary — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersVPRKpStandard — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersVPRKpSymmetric — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersVPRKpVariational — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersVSPARK — Type.

Parameters for right-hand side function of Variational Specialised Partitioned Additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.ParametersWIRK — Type.

Parameters for right-hand side function of weak implicit Runge-Kutta methods.

source

Base.show — Method.

Print additive Runge-Kutta coefficients.

source

Base.show — Method.

Print multiplier Runge-Kutta coefficients.

source

Base.show — Method.

Print Runge-Kutta coefficients.

source

Base.show — Method.

Print projective Runge-Kutta coefficients.

source

Base.show — Method.

Print Runge-Kutta coefficients.

source

GeometricIntegrators.Integrators.aitken_neville — Method.

Compute p(x) where p is the unique polynomial of degree length(xi), such that p(x[i]) = y[i]) for all i.

ti: interpolation nodes
xi: interpolation values
t:  evaluation point
x:  evaluation value

source

GeometricIntegrators.Integrators.computeStageP! — Method.

Compute P stages of explicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.computeStageQ! — Method.

Compute Q stages of explicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.compute_stages! — Method.

Unpacks the data stored in x = (Y[1,1], Y[2,1], ... Y[D,1], Y[1,2], ..., Z[1,1], Z[2,1], ... Z[D,1], Z[1,2], ...) into the matrix Y, Z, calculates the internal stages Q, P, the values of the RHS of the SDE ( vi(Q,P), fi(Q,P), Bi(Q,P) and Gi(Q,P) ), and assigns them to VQPi, FQPi, BQPi and GQPi. Unlike for FIRK, here Y = Δt adrift v(Q,P) + adiff B(Q,P) ΔW, Z = Δt ̃a1drift f1(Q,P) + Δt ̃a2drift f2(Q,P) + ̃a1diff G1(Q,P) ΔW + ̃a2diff G2(Q,P) ΔW.

source

GeometricIntegrators.Integrators.compute_stages! — Method.

Unpacks the data stored in x = (Y[1,1], Y[2,1], ... Y[D,1], Y[1,2], ..., Z[1,1], Z[2,1], ... Z[D,1], Z[1,2], ...) into the matrix Y, Z, calculates the internal stages Q, P, the values of the RHS of the SDE ( v(Q,P), f(Q,P), B(Q,P) and G(Q,P) ), and assigns them to VQP, FQP, BQP and GQP. Unlike for FIRK, here Y = Δt adrift v(Q,P) + adiff B(Q,P) ΔW, Z = Δt ̃adrift v(Q,P) + ̃adiff B(Q,P) ΔW.

source

GeometricIntegrators.Integrators.compute_stages! — Method.

Unpacks the data stored in x = (Y[1,1], Y[2,1], ... Y[D,1], Y[1,2], ...) into the matrix Y, calculates the internal stages Q, the values of the RHS of the SDE ( v(Q) and B(Q) ), and assigns them to VQ and BQ. Unlike for FIRK, here Y = Δt a v(Q) + ̃a B(Q) ΔW

source

GeometricIntegrators.Integrators.compute_stages! — Method.

Unpacks the data stored in x = (Y0[1,1], Y0[2,1], ... Y0[D,1], ... Y0[D,S], Y1[1,1,1], Y1[2,1,1], ... Y1[D,1,1], Y1[1,2,1], Y1[2,2,1], ... Y1[D,2,1], ... Y1[D,M,S] ) into the matrices Y0 and Y1, calculates the internal stages Q0 and Q1, the values of the RHS of the SDE ( v(Q0) and B(Q1) ), and assigns them to VQ and BQ. Unlike for FIRK, here Y = Δt a v(Q) + ̃a B(Q) ΔW

source

GeometricIntegrators.Integrators.compute_stages_p! — Method.

Compute one-form and forces at quadrature nodes.

source

GeometricIntegrators.Integrators.compute_stages_p! — Method.

Compute one-form and forces at quadrature nodes.

source

GeometricIntegrators.Integrators.compute_stages_p! — Method.

Compute one-form and forces at quadrature nodes.

source

GeometricIntegrators.Integrators.compute_stages_p! — Method.

Compute one-form and forces at quadrature nodes.

source

GeometricIntegrators.Integrators.compute_stages_p! — Method.

Compute one-form and forces at quadrature nodes.

source

GeometricIntegrators.Integrators.compute_stages_q! — Method.

Compute solution at quadrature nodes and across jump.

source

GeometricIntegrators.Integrators.compute_stages_q! — Method.

Compute solution at quadrature nodes and across jump.

source

GeometricIntegrators.Integrators.compute_stages_q! — Method.

Compute solution at quadrature nodes and across jump.

source

GeometricIntegrators.Integrators.compute_stages_q! — Method.

Compute solution at quadrature nodes and across jump.

source

GeometricIntegrators.Integrators.compute_stages_q! — Method.

Compute solution at quadrature nodes and across jump.

source

GeometricIntegrators.Integrators.compute_stages_v! — Method.

Compute velocities at quadrature nodes.

source

GeometricIntegrators.Integrators.compute_stages_v! — Method.

Compute velocities at quadrature nodes.

source

GeometricIntegrators.Integrators.compute_stages_v! — Method.

Compute velocities at quadrature nodes.

source

GeometricIntegrators.Integrators.compute_stages_v! — Method.

Compute velocities at quadrature nodes.

source

GeometricIntegrators.Integrators.compute_stages_v! — Method.

Compute velocities at quadrature nodes.

source

GeometricIntegrators.Integrators.create_internal_stage_vector — Method.

Create a vector of S solution matrices of type DT to store the solution of S internal stages for a problem with DxM dimensions.

source

GeometricIntegrators.Integrators.create_internal_stage_vector — Method.

Create a vector of S solution vectors of type DT to store the solution of S internal stages for a problem with D dimensions.

source

GeometricIntegrators.Integrators.create_internal_stage_vector_with_zero — Method.

Create a vector of S+1 solution vectors of type DT to store the solution of S internal stages and the solution of the previous timestep for a problem with D dimensions.

source

GeometricIntegrators.Integrators.create_nonlinear_solver — Method.

Create nonlinear solver object for a system of N equations with data type DT. The function $f(x)=0$ to be solved for is determined by a julia function function_stages!(x, b, params), where x is the current solution and b is the output vector, s.th. $b = f(x)$. params are a set of parameters depending on the equation and integrator that is used. The solver type is obtained from the config dictionary (:nls_solver).

source

GeometricIntegrators.Integrators.create_solution_vector — Method.

Create a solution vector of type TwicePrecision{DT} for a problem with D dimensions, NS' sample paths, andNI` independent initial conditions.

source

GeometricIntegrators.Integrators.create_solution_vector — Method.

Create a solution vector of type TwicePrecision{DT} for a problem with D dimensions and M independent initial conditions.

source

GeometricIntegrators.Integrators.euler_extrapolation — Method.

Euler extrapolation method with arbitrary order p.

v:  function to compute vector field
t₀: initial time
t₁: final   time
x₀: initial value
x₁: final   value
s:  number of interpolations (order p=s+1)

TODO This is probably broken!

source

GeometricIntegrators.Integrators.initial_guess! — Method.

Compute initial guess for internal stages.

source

GeometricIntegrators.Integrators.initial_guess! — Method.

This function computes initial guesses for Y, Z and assigns them to int.solver.x The prediction is calculated using an explicit integrator.

source

GeometricIntegrators.Integrators.initial_guess! — Method.

This function computes initial guesses for Y and assigns them to int.solver.x The prediction is calculated using an explicit integrator.

source

GeometricIntegrators.Integrators.initial_guess! — Method.

This function computes initial guesses for Y, Z and assigns them to int.solver.x The prediction is calculated using an explicit integrator.

source

GeometricIntegrators.Integrators.initial_guess! — Method.

This function computes initial guesses for Y and assigns them to int.solver.x The prediction is calculated using an explicit integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate SDE with explicit Runge-Kutta integrator. Calculating the n-th time step of the explicit integrator for the sample path r and the initial condition m

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate SDE with explicit Runge-Kutta integrator. Calculating the n-th time step of the explicit integrator for the sample path r and the initial condition m

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate PSDE with a stochastic implicit partitioned Runge-Kutta integrator. Integrating the k-th sample path for the m-th initial condition

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate SDE with a stochastic implicit Runge-Kutta integrator. Integrating the k-th sample path for the m-th initial condition

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate PSDE with a stochastic implicit partitioned Runge-Kutta integrator. Integrating the k-th sample path for the m-th initial condition

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate SDE with a stochastic implicit Runge-Kutta integrator. Integrating the k-th sample path for the m-th initial condition

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with splitting integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with diagonally implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with fully implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate DAE with variational special partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate DAE with variational special partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate DAE with variational partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate DAE with variational special partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate DAE with partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate partitioned ODE with explicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with explicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with fully implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with implicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate PODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.midpoint_extrapolation — Method.

Midpoint extrapolation method with arbitrary order p.

v:  function to compute vector field
f:  function to compute force  field
t₀: initial time
t₁: final   time
q₀: initial positions
p₀: initial momenta
q₁: final   positions
p₁: final   momenta
s:  number of interpolations (order p=2s+2)

source

GeometricIntegrators.Integrators.midpoint_extrapolation — Method.

Midpoint extrapolation method with arbitrary order p.

v:  function to compute vector field
t₀: initial time
t₁: final   time
x₀: initial value
x₁: final   value
s:  number of interpolations (order p=2s+2)

source

GeometricIntegrators.Integrators.readTableauRKHeaderFromFile — Method.

Reads and parses Tableau metadata from file.

source