Integrators

Common

Initial Guesses

Extrapolation Methods

The extrapolation routines are exclusively used for computing initial guesses and are usually not called directly by the user.

Splitting Methods

Runge-Kutta Methods

GeometricIntegrators.Integrators.IntegratorCacheERK — Type

Explicit Runge-Kutta integrator cache.

GeometricIntegrators.Integrators.IntegratorERK — Type

Explicit Runge-Kutta integrator solving the system

\[\begin{aligned} V_{n,i} &= v (t_i, Q_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} . \end{aligned}\]

GeometricIntegrators.Integrators.IntegratorCacheDIRK — Type

Diagonally implicit Runge-Kutta integrator cache.

GeometricIntegrators.Integrators.IntegratorDIRK — Type

Diagonally implicit Runge-Kutta integrator solving the system

\[\begin{aligned} V_{n,i} &= v (t_i, Q_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} . \end{aligned}\]

GeometricIntegrators.Integrators.IntegratorCacheEPRK — Type

Explicit Runge-Kutta integrator cache.

GeometricIntegrators.Integrators.IntegratorEPRK — Type

Explicit partitioned Runge-Kutta integrator solving the system

\[\begin{aligned} V_{n,i} &= v (t_i, Q_{n,i}, P_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} , \\ F_{n,i} &= f (t_i, Q_{n,i}, P_{n,i}) , & P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, F_{n,i} . \end{aligned}\]

Usually we are interested in Hamiltonian systems, where

\[\begin{aligned} V_{n,i} &= \dfrac{\partial H}{\partial p} (Q_{n,i}, P_{n,i}) , & F_{n,i} &= - \dfrac{\partial H}{\partial q} (Q_{n,i}, P_{n,i}) , \end{aligned}\]

and tableaus satisfying the symplecticity conditions

\[\begin{aligned} b_{i} \bar{a}_{ij} + \bar{b}_{j} a_{ji} &= b_{i} \bar{b}_{j} , & \bar{b}_i &= b_i . \end{aligned}\]

GeometricIntegrators.Integrators.IntegratorCacheIPRK — Type

Implicit partitioned Runge-Kutta integrator cache.

Fields

q̃: initial guess of q

p̃: initial guess of p

ṽ: initial guess of v

f̃: initial guess of f

s̃: holds shift due to periodicity of solution

Q: internal stages of q

P: internal stages of p

V: internal stages of v

F: internal stages of f

Y: vector field of internal stages of q

Z: vector field of internal stages of p

GeometricIntegrators.Integrators.IntegratorIPRK — Type

Implicit partitioned Runge-Kutta integrator solving the system

\[\begin{aligned} V_{n,i} &= v (t_i, Q_{n,i}, P_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} , \\ F_{n,i} &= f (t_i, Q_{n,i}, P_{n,i}) , & P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, F_{n,i} , \end{aligned}\]

Usually we are interested in Hamiltonian systems, where

\[\begin{aligned} V_{n,i} &= \dfrac{\partial H}{\partial p} (Q_{n,i}, P_{n,i}) , & F_{n,i} &= - \dfrac{\partial H}{\partial q} (Q_{n,i}, P_{n,i}) , \end{aligned}\]

and tableaus satisfying the symplecticity conditions

\[\begin{aligned} b_{i} \bar{a}_{ij} + \bar{b}_{j} a_{ji} &= b_{i} \bar{b}_{j} , & \bar{b}_i &= b_i . \end{aligned}\]

Variational Integrators

GeometricIntegrators.Integrators.IntegratorVPRK — Type

Variational Partitioned Runge-Kutta Integrator.

\[\begin{aligned} P_{n,i} &= \dfrac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} , \\ F_{n,i} &= \dfrac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , & P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} - d_i \lambda , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, F_{n,i} , \\ && 0 &= \sum \limits_{i=1}^{s} d_i V_i , && \end{aligned}\]

satisfying the symplecticity conditions

\[\begin{aligned} b_{i} \bar{a}_{ij} + b_{j} a_{ji} &= b_{i} b_{j} , & \bar{b}_i &= b_i . \end{aligned}\]

Degenerate Variational Integrators

GeometricIntegrators.Integrators.IntegratorCacheDVIA — Type

Degenerate variational integrator cache.

Fields

q: internal stages of solution

v: internal stages of vector field

Θ: implicit function evaluated on solution

f: vector field of implicit function

GeometricIntegrators.Integrators.IntegratorDVIA — Type

Symplectic Euler-A Degenerate Variational Integrator.

GeometricIntegrators.Integrators.IntegratorCacheDVIB — Type

Degenerate variational integrator cache.

Fields

q: internal stages of solution

v: internal stages of vector field

Θ: implicit function evaluated on solution

f: vector field of implicit function

GeometricIntegrators.Integrators.IntegratorDVIB — Type

Symplectic Euler-B Degenerate Variational Integrator.

GeometricIntegrators.Integrators.IntegratorCMDVI — Type

Midpoint Degenerate Variational Integrator.

GeometricIntegrators.Integrators.IntegratorCacheCMDVI — Type

Degenerate variational integrator cache.

Fields

q: internal stages of solution

v: internal stages of vector field

Θ: implicit function evaluated on solution

f: vector field of implicit function

GeometricIntegrators.Integrators.IntegratorCTDVI — Type

Trapezoidal Degenerate Variational Integrator.

GeometricIntegrators.Integrators.IntegratorCacheCTDVI — Type

Degenerate variational integrator cache.

Fields

q: internal stages of solution

v: internal stages of vector field

Θ: implicit function evaluated on solution

f: vector field of implicit function

Galerkin Variational Integrators