Integrators

Integrators

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for special additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Print error for integrators not implemented, yet.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for variational partitioned additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for variational special partitioned additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for variational partitioned Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for diagonally implicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for explicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for fully implicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for singly implicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for partitioned additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for special partitioned additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for explicit partitioned Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for implicit partitioned Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for splitting tableau.

source

GeometricIntegrators.Integrators.IntegratorCGVI — Type.

Continuous Galerkin Variational Integrator.

source

GeometricIntegrators.Integrators.IntegratorDGVI — Type.

Discontinuous Galerkin Variational Integrator.

source

GeometricIntegrators.Integrators.IntegratorDIRK — Type.

Diagonally implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorEPRK — Type.

Explicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorERK — Type.

Explicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorFIRK — Type.

Fully implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorIPRK — Type.

Implicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorPARK — Type.

Implicit partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSIRK — Type.

Singly implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSPARK — Type.

Special Partitioned Additive Runge Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSplitting — Type.

Splitting integrator.

source

GeometricIntegrators.Integrators.IntegratorSplitting — Method.

Construct splitting integrator for symmetric splitting tableau with general stages.

source

GeometricIntegrators.Integrators.IntegratorSplitting — Method.

Construct splitting integrator for non-symmetric splitting tableau with general stages.

source

GeometricIntegrators.Integrators.IntegratorSplitting — Method.

Construct splitting integrator for symmetric splitting tableau with symmetric stages.

source

GeometricIntegrators.Integrators.IntegratorVPARK — Type.

Variational partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRK — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpMidpoint — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpSecondary — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpStandard — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpSymmetric — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpVariational — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVSPARK — Type.

Variational special partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.initialize! — Method.

Initialize integrator for initial conditions m with m₁ ≤ m ≤ m₂ and time step 0.

source

GeometricIntegrators.Integrators.integrate — Function.

Integrate ODE specified by vector field and initial condition with given tableau for ntime time steps and return solution.

source

GeometricIntegrators.Integrators.integrate — Function.

Apply integrator for ntime time steps and return solution.

source

GeometricIntegrators.Integrators.integrate — Function.

Integrate given equation with given tableau for ntime time steps and return solution.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE for initial conditions m with m₁ ≤ m ≤ m₂.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE for all initial conditions.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE with diagonally implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate partitioned ODE with diagonally implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE with singly implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate partitioned ODE with singly implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate partitioned DAE with Special Additive Runge Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE for initial conditions m with m₁ ≤ m ≤ m₂ for time steps n with n₁ ≤ n ≤ n₂.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of variational partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of variational special partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of fully implicit Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of implicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of projected variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of projected variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersCGVI — Type.

NonlinearFunctionParametersCGVI: Parameters for right-hand side function of continuous Galerkin variational Integrator.

Parameters

Θ: function of the noncanonical one-form (∂L/∂v)

f: function of the force (∂L/∂q)

Δt: time step

b: weights of the quadrature rule

c: nodes of the quadrature rule

x: nodes of the basis

m: mass matrix

a: derivative matrix

r₀: reconstruction coefficients at the beginning of the interval

r₁: reconstruction coefficients at the end of the interval

t: current time

q: current solution of q

p: current solution of p

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersDGVI — Type.

NonlinearFunctionParametersDGVI: Parameters for right-hand side function of discontinuous Galerkin variational Integrator.

Parameters

Θ: function of the noncanonical one-form (∂L/∂v)

f: function of the force (∂L/∂q)

Δt: time step

b: quadrature weights

c: quadrature nodes

m: mass matrix

a: derivative matrix

r₀: reconstruction coefficients, left-hand side

r₁: reconstruction coefficients, right-hand side

β: weights of the quadrature rule for the flux

γ: nodes of the quadrature rule for the flux

μ₀: mass vector for the lhs flux

μ₁: mass vector for the rhs flux

α₀: derivative vector for the lhs flux

α₁: derivative vector for the rhs flux

t: current time

q: current solution of q

p: current solution of p

D: dimension of the system

S: number of basis nodes

R: number of quadrature nodes

P: number of quadrature nodes for the flux

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersFIRK — Type.

Parameters for right-hand side function of fully implicit Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersIPRK — Type.

Parameters for right-hand side function of implicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersPARK — Type.

Parameters for right-hand side function of partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPARK — Type.

Parameters for right-hand side function of variational partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPRK — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPRKpMidpoint — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPRKpSecondary — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPRKpStandard — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPRKpSymmetric — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPRKpVariational — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVSPARK — Type.

Parameters for right-hand side function of variational special partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.aitken_neville — Method.

Compute p(x) where p is the unique polynomial of degree length(xi), such that p(x[i]) = y[i]) for all i.

ti: interpolation nodes
xi: interpolation values
t:  evaluation point
x:  evaluation value

source

GeometricIntegrators.Integrators.computeStageP! — Method.

Compute P stages of explicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.computeStageQ! — Method.

Compute Q stages of explicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.euler_extrapolation — Method.

Euler extrapolation method with arbitrary order p.

v:  function to compute vector field
t₀: initial time
t₁: final   time
x₀: initial value
x₁: final   value
s:  number of interpolations (order p=s+1)

TODO This is probably broken!

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate DAE with partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate DAE with variational partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate DAE with variational special partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate partitioned ODE with explicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with explicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with fully implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with implicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with splitting integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.midpoint_extrapolation — Method.

Midpoint extrapolation method with arbitrary order p.

v:  function to compute vector field
f:  function to compute force  field
t₀: initial time
t₁: final   time
q₀: initial positions
p₀: initial momenta
q₁: final   positions
p₁: final   momenta
s:  number of interpolations (order p=2s+2)

source

GeometricIntegrators.Integrators.midpoint_extrapolation — Method.

Midpoint extrapolation method with arbitrary order p.

v:  function to compute vector field
t₀: initial time
t₁: final   time
x₀: initial value
x₁: final   value
s:  number of interpolations (order p=2s+2)

source