Integrators

Integrators

GeometricIntegrators.Integrators.AbstractTableau — Type.

Holds the information for the various methods' tableaus.

source

GeometricIntegrators.Integrators.AbstractTableauIRK — Type.

Holds the tableau of an implicit Runge-Kutta method.

source

GeometricIntegrators.Integrators.AbstractTableauPRK — Type.

Holds the tableau of a partitioned Runge-Kutta method.

source

GeometricIntegrators.Integrators.AbstractTableauRK — Type.

Holds the tableau of a Runge-Kutta method.

source

GeometricIntegrators.Integrators.CoefficientsARK — Type.

Holds the coefficients of an additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.CoefficientsMRK — Type.

Holds the multiplier Runge-Kutta coefficients.

source

GeometricIntegrators.Integrators.CoefficientsPGLRK — Type.

Holds the coefficients of a projected Gauss-Legendre Runge-Kutta method.

source

GeometricIntegrators.Integrators.CoefficientsPRK — Type.

Holds the coefficients of a projective Runge-Kutta method.

source

GeometricIntegrators.Integrators.CoefficientsRK — Type.

Holds the coefficients of a Runge-Kutta method.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for special additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Print error for integrators not implemented, yet.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for variational partitioned additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for variational special partitioned additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for Projected Gauss-Legendre Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for variational partitioned Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for diagonally implicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for explicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for fully implicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for singly implicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for partitioned additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for special partitioned additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for explicit partitioned Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for implicit partitioned Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for splitting tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for formal Lagrangian Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.IntegratorCGVI — Type.

Continuous Galerkin Variational Integrator.

source

GeometricIntegrators.Integrators.IntegratorDGVI — Type.

Discontinuous Galerkin Variational Integrator.

source

GeometricIntegrators.Integrators.IntegratorDIRK — Type.

Diagonally implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorEPRK — Type.

Explicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorERK — Type.

Explicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorFIRK — Type.

Fully implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorFLRK — Type.

Formal Lagrangian Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorGPARK — Type.

Special Partitioned Additive Runge Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorHSPARK — Type.

Hamiltonian Specialised Partitioned Additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorIPRK — Type.

Implicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorPARK — Type.

Implicit partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorPGLRK — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSIRK — Type.

Singly implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSPARK — Type.

Variational special partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSplitting — Type.

Splitting integrator.

source

GeometricIntegrators.Integrators.IntegratorSplitting — Method.

Construct splitting integrator for symmetric splitting tableau with general stages.

source

GeometricIntegrators.Integrators.IntegratorSplitting — Method.

Construct splitting integrator for non-symmetric splitting tableau with general stages.

source

GeometricIntegrators.Integrators.IntegratorSplitting — Method.

Construct splitting integrator for symmetric splitting tableau with symmetric stages.

source

GeometricIntegrators.Integrators.IntegratorVPARK — Type.

Variational partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRK — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpLegendre — Type.

Variational special partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpMidpoint — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpSecondary — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpStandard — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpSymmetric — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpVariational — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVSPARK — Type.

Variational Specialised Partitioned Additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.TableauARK — Type.

Holds the tableau of a additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauDIRK — Type.

Holds the tableau of a diagonally implicit Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauEPRK — Type.

TableauEPRK: Tableau of an Explicit Partitioned Runge-Kutta method

\[\begin{align*} V_{n,i} &= \hphantom{-} \dfrac{\partial H}{\partial p} (Q_{n,i}, P_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} , \\ F_{k,i} &= - \dfrac{\partial H}{\partial q} (Q_{n,i}, P_{n,i}) , & P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, F_{n,i} , \end{align*}\]

usually satisfying the symplecticity conditions

\[\begin{align*} b_{i} \bar{a}_{ij} + b_{j} a_{ji} &= b_{i} b_{j} , & \bar{b}_i &= b_i . \end{align*}\]

source

GeometricIntegrators.Integrators.TableauERK — Type.

Holds the tableau of an explicit Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauFIRK — Type.

Holds the tableau of a fully implicit Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauGLM — Type.

Holds the tableau of a general linear method.

source

GeometricIntegrators.Integrators.TableauGPARK — Type.

Holds the tableau of a spezialized partitioned additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauHSPARK — Type.

Holds the tableau of an Hamiltonian Specialised Partitioned Additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauIPRK — Type.

TableauIPRK: Tableau of an Implicit Partitioned Runge-Kutta method

\[\begin{align*} P_{n,i} &= \dfrac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} , \\ F_{k,i} &= \dfrac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , & P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, F_{n,i} , \end{align*}\]

usually satisfying the symplecticity conditions

\[\begin{align*} b_{i} \bar{a}_{ij} + b_{j} a_{ji} &= b_{i} b_{j} , & \bar{b}_i &= b_i . \end{align*}\]

source

GeometricIntegrators.Integrators.TableauPARK — Type.

Holds the tableau of an partitioned additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauSARK — Type.

Holds the tableau of a spezialized additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauSIRK — Type.

Holds the tableau of a singly implicit Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauSPARK — Type.

Holds the tableau of an Specialised Partitioned Additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauSplittingGS — Type.

Tableau for symmetric splitting methods with general stages.

source

GeometricIntegrators.Integrators.TableauSplittingNS — Type.

Tableau for non-symmetric splitting methods.

source

GeometricIntegrators.Integrators.TableauSplittingSS — Type.

Tableau for symmetric splitting methods with symmetric stages.

source

GeometricIntegrators.Integrators.TableauVPARK — Type.

Holds the tableau of an variational partitioned additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.TableauVPRK — Type.

TableauVPRK: Tableau of a Variational Partitioned Runge-Kutta method

\[\begin{align*} P_{n,i} &= \dfrac{\partial L}{\partial v} (Q_{n,i}, V_{n,i}) , & Q_{n,i} &= q_{n} + h \sum \limits_{j=1}^{s} a_{ij} \, V_{n,j} , & q_{n+1} &= q_{n} + h \sum \limits_{i=1}^{s} b_{i} \, V_{n,i} , \\ F_{k,i} &= \dfrac{\partial L}{\partial q} (Q_{n,i}, V_{n,i}) , & P_{n,i} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{a}_{ij} \, F_{n,j} - d_i \lambda , & p_{n+1} &= p_{n} + h \sum \limits_{i=1}^{s} \bar{b}_{i} \, F_{n,i} , \\ && 0 &= \sum \limits_{i=1}^{s} d_i V_i , && \end{align*}\]

satisfying the symplecticity conditions

\[\begin{align*} b_{i} \bar{a}_{ij} + b_{j} a_{ji} &= b_{i} b_{j} , & \bar{b}_i &= b_i . \end{align*}\]

source

GeometricIntegrators.Integrators.TableauVSPARK — Type.

Holds the tableau of an Variational Specialised Partitioned Additive Runge-Kutta method.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of Hamiltonian Specialised Partitioned Additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational special partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of fully implicit Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of formal Lagrangian Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of implicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational special partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of projected variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of projected variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.initialize! — Method.

Initialize integrator for initial conditions m with m₁ ≤ m ≤ m₂ and time step 0.

source

GeometricIntegrators.Integrators.integrate — Function.

Integrate given equation with given tableau for ntime time steps and return solution.

source

GeometricIntegrators.Integrators.integrate — Function.

Apply integrator for ntime time steps and return solution.

source

GeometricIntegrators.Integrators.integrate — Function.

Integrate ODE specified by vector field and initial condition with given tableau for ntime time steps and return solution.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE for initial conditions m with m₁ ≤ m ≤ m₂.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE for all initial conditions.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE with diagonally implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate partitioned ODE with diagonally implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate partitioned DAE with Special Additive Runge Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE with singly implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate partitioned ODE with singly implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE for initial conditions m with m₁ ≤ m ≤ m₂ for time steps n with n₁ ≤ n ≤ n₂.

source

GeometricIntegrators.Integrators.readTableauERKFromFile — Method.

Read explicit Runge-Kutta tableau from file.

source

GeometricIntegrators.Integrators.writeTableauToFile — Method.

Write Runge-Kutta tableau to file.

source

GeometricIntegrators.Integrators.AbstractTableauERK — Type.

Holds the tableau of an explicit Runge-Kutta method.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersCGVI — Type.

NonlinearFunctionParametersCGVI: Parameters for right-hand side function of continuous Galerkin variational Integrator.

Parameters

Θ: function of the noncanonical one-form (∂L/∂v)

f: function of the force (∂L/∂q)

Δt: time step

b: weights of the quadrature rule

c: nodes of the quadrature rule

x: nodes of the basis

m: mass matrix

a: derivative matrix

r₀: reconstruction coefficients at the beginning of the interval

r₁: reconstruction coefficients at the end of the interval

t: current time

q: current solution of q

p: current solution of p

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersDGVI — Type.

NonlinearFunctionParametersDGVI: Parameters for right-hand side function of discontinuous Galerkin variational Integrator.

Parameters

Θ: function of the noncanonical one-form (∂L/∂v)

f: function of the force (∂L/∂q)

Δt: time step

b: quadrature weights

c: quadrature nodes

m: mass matrix

a: derivative matrix

r₀: reconstruction coefficients, left-hand side

r₁: reconstruction coefficients, right-hand side

β: weights of the quadrature rule for the flux

γ: nodes of the quadrature rule for the flux

μ₀: mass vector for the lhs flux

μ₁: mass vector for the rhs flux

α₀: derivative vector for the lhs flux

α₁: derivative vector for the rhs flux

t: current time

q: current solution of q

p: current solution of p

D: dimension of the system

S: number of basis nodes

R: number of quadrature nodes

P: number of quadrature nodes for the flux

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersFIRK — Type.

Parameters for right-hand side function of fully implicit Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersFLRK — Type.

Parameters for right-hand side function of formal Lagrangian Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersHSPARK — Type.

Parameters for right-hand side function of Hamiltonian Specialised Partitioned Additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersIPRK — Type.

Parameters for right-hand side function of implicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersPARK — Type.

Parameters for right-hand side function of partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersPGLRK — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersSPARK — Type.

Parameters for right-hand side function of Specialised Partitioned Additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPARK — Type.

Parameters for right-hand side function of variational partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPRK — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPRKpLegendre — Type.

Parameters for right-hand side function of variational special partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPRKpMidpoint — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPRKpSecondary — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPRKpStandard — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPRKpSymmetric — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPRKpVariational — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVSPARK — Type.

Parameters for right-hand side function of Variational Specialised Partitioned Additive Runge-Kutta methods.

source

Base.show — Method.

Print additive Runge-Kutta coefficients.

source

Base.show — Method.

Print multiplier Runge-Kutta coefficients.

source

Base.show — Method.

Print Runge-Kutta coefficients.

source

Base.show — Method.

Print projective Runge-Kutta coefficients.

source

Base.show — Method.

Print Runge-Kutta coefficients.

source

GeometricIntegrators.Integrators.aitken_neville — Method.

Compute p(x) where p is the unique polynomial of degree length(xi), such that p(x[i]) = y[i]) for all i.

ti: interpolation nodes
xi: interpolation values
t:  evaluation point
x:  evaluation value

source

GeometricIntegrators.Integrators.computeStageP! — Method.

Compute P stages of explicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.computeStageQ! — Method.

Compute Q stages of explicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.create_nonlinear_solver — Method.

Create nonlinear solver object for a system of N equations with data type DT. The function $f(x)=0$ to be solved for is determined by a julia function function_stages!(x, b, params), where x is the current solution and b is the output vector, s.th. $b = f(x)$. params are a set of parameters depending on the equation and integrator that is used. The solver type is obtained from the config dictionary (:nls_solver).

source

GeometricIntegrators.Integrators.create_solution_vector_double_double — Method.

Create a solution vector of type Double{DT} for a problem with D dimensions and M independent initial conditions.

source

GeometricIntegrators.Integrators.euler_extrapolation — Method.

Euler extrapolation method with arbitrary order p.

v:  function to compute vector field
t₀: initial time
t₁: final   time
x₀: initial value
x₁: final   value
s:  number of interpolations (order p=s+1)

TODO This is probably broken!

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate DAE with partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate DAE with variational special partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate DAE with variational partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate partitioned ODE with explicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with explicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate DAE with variational special partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with implicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with splitting integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate DAE with variational special partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with fully implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with fully implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.midpoint_extrapolation — Method.

Midpoint extrapolation method with arbitrary order p.

v:  function to compute vector field
f:  function to compute force  field
t₀: initial time
t₁: final   time
q₀: initial positions
p₀: initial momenta
q₁: final   positions
p₁: final   momenta
s:  number of interpolations (order p=2s+2)

source

GeometricIntegrators.Integrators.midpoint_extrapolation — Method.

Midpoint extrapolation method with arbitrary order p.

v:  function to compute vector field
t₀: initial time
t₁: final   time
x₀: initial value
x₁: final   value
s:  number of interpolations (order p=2s+2)

source

GeometricIntegrators.Integrators.readTableauRKHeaderFromFile — Method.

Reads and parses Tableau metadata from file.

source