Integrators

Integrators

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for special additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Print error for integrators not implemented, yet.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for variational partitioned additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for variational special partitioned additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for variational partitioned Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for diagonally implicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for explicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for fully implicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for singly implicit Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for partitioned additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for special partitioned additive Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for explicit partitioned Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.Integrator — Method.

Create integrator for implicit partitioned Runge-Kutta tableau.

source

GeometricIntegrators.Integrators.IntegratorDIRK — Type.

Diagonally implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorEPRK — Type.

Explicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorERK — Type.

Explicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorFIRK — Type.

Fully implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorIPRK — Type.

Implicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSARK — Type.

Special Additive Runge Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSIRK — Type.

Singly implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorSPARK — Type.

Special Partitioned Additive Runge Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPARK — Type.

Variational partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRK — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpMidpoint — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpStandard — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVPRKpSymmetric — Type.

Variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorVSPARK — Type.

Variational special partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.initialize! — Method.

Initialize integrator for initial conditions m with m₁ ≤ m ≤ m₂ and time step 0.

source

GeometricIntegrators.Integrators.integrate — Function.

Integrate ODE specified by vector field and initial condition with given tableau for ntime time steps and return solution.

source

GeometricIntegrators.Integrators.integrate — Function.

Apply integrator for ntime time steps and return solution.

source

GeometricIntegrators.Integrators.integrate — Function.

Integrate given equation with given tableau for ntime time steps and return solution.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE for initial conditions m with m₁ ≤ m ≤ m₂.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE for all initial conditions.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate DAE with Additive Runge Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE with diagonally implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate partitioned ODE with diagonally implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate DAE with Special Additive Runge Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE with singly implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate partitioned ODE with singly implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate partitioned DAE with Special Additive Runge Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate DAE with variational partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate! — Method.

Integrate ODE for initial conditions m with m₁ ≤ m ≤ m₂ for time steps n with n₁ ≤ n ≤ n₂.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of variational partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of variational special partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of fully implicit Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of implicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of projected variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Solvers.function_stages! — Method.

Compute stages of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.IntegratorARK — Type.

Additive Runge Kutta integrator.

source

GeometricIntegrators.Integrators.IntegratorPARK — Type.

Implicit partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersFIRK — Type.

Parameters for right-hand side function of fully implicit Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersIPRK — Type.

Parameters for right-hand side function of implicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersPARK — Type.

Parameters for right-hand side function of partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPARK — Type.

Parameters for right-hand side function of variational partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPRK — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPRKpMidpoint — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPRKpStandard — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVPRKpSymmetric — Type.

Parameters for right-hand side function of variational partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.NonlinearFunctionParametersVSPARK — Type.

Parameters for right-hand side function of variational special partitioned additive Runge-Kutta methods.

source

GeometricIntegrators.Integrators.aitken_neville — Method.

Compute p(x) where p is the unique polynomial of degree length(xi), such that p(x[i]) = y[i]) for all i.

ti: interpolation nodes
xi: interpolation values
t:  evaluation point
x:  evaluation value

source

GeometricIntegrators.Integrators.computeStageP! — Method.

Compute P stages of explicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.computeStageQ! — Method.

Compute Q stages of explicit partitioned Runge-Kutta methods.

source

GeometricIntegrators.Integrators.euler_extrapolation — Method.

Euler extrapolation method with arbitrary order p.

v:  function to compute vector field
t₀: initial time
t₁: final   time
x₀: initial value
x₁: final   value
s:  number of interpolations (order p=s+1)

TODO This is probably broken!

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate DAE with partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with variational partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate DAE with variational special partitioned additive Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate partitioned ODE with explicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with explicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with fully implicit Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.integrate_step! — Method.

Integrate ODE with implicit partitioned Runge-Kutta integrator.

source

GeometricIntegrators.Integrators.midpoint_extrapolation — Method.

Midpoint extrapolation method with arbitrary order p.

v:  function to compute vector field
f:  function to compute force  field
t₀: initial time
t₁: final   time
q₀: initial positions
p₀: initial momenta
q₁: final   positions
p₁: final   momenta
s:  number of interpolations (order p=2s+2)

source

GeometricIntegrators.Integrators.midpoint_extrapolation — Method.

Midpoint extrapolation method with arbitrary order p.

v:  function to compute vector field
t₀: initial time
t₁: final   time
x₀: initial value
x₁: final   value
s:  number of interpolations (order p=2s+2)

source