Methods

GeometricIntegrators.Methods.BackwardEuler — Type

Fully implicit Runge-Kutta method with TableauBackwardEuler.

Reference:

Augustin-Louis Cauchy.
Équations différentielles ordinaires. Cours inédit (fragment). Douzième leçon.
Ed. Christian Gilain, Etudes Vivantes, 1981.
Page 102, Equation (5), Θ=1.

GeometricIntegrators.Methods.CMDVI — Type

Midpoint Degenerate Variational Integrator.

GeometricIntegrators.Methods.CTDVI — Type

Trapezoidal Degenerate Variational Integrator.

GeometricIntegrators.Methods.CrankNicolson — Type

Diagonally implicit Runge-Kutta method with TableauCrankNicolson.

Reference:

J. Crank and P. Nicolson.
A practical method for numerical evaluation of solutions of partial differential equations of the heat-conduction type.
Mathematical Proceedings of the Cambridge Philosophical Society, Volume 43, Issue 1, Pages 50-67, 1947.
doi: 10.1017/S0305004100023197

GeometricIntegrators.Methods.Crouzeix — Type

Diagonally implicit Runge-Kutta method with TableauCrouzeix.

Reference:

M.Crouzeix.
Sur L'approximation des équations différentielles opérationelles linéaires par des méthodes de Runge-Kutta.
Thesis. Université de Paris, 1975.

GeometricIntegrators.Methods.DIRK — Type

Diagonally Implicit Runge-Kutta Method

DIRK(tableau)

GeometricIntegrators.Methods.DVIA — Type

Symplectic Euler-A Degenerate Variational Integrator.

GeometricIntegrators.Methods.DVIB — Type

Symplectic Euler-B Degenerate Variational Integrator.

GeometricIntegrators.Methods.DVRK — Type

Degenerate Variational Runge-Kutta Methods

DVRK(tableau::Tableau)
DVRK(method::RKMethod)

GeometricIntegrators.Methods.DegenerateVPRK — Type

Variational Partitioned Runge-Kutta Method for Degenerate Lagrangians

DegenerateVPRK(method::VPRK)
DegenerateVPRK(tableau::AbstractTableau, args...; kwargs...)

GeometricIntegrators.Methods.EPRK — Type

Explicit Partitioned Runge-Kutta Method

EPRK(tableau)

GeometricIntegrators.Methods.ERK — Type

Explicit Runge-Kutta Method

ERK(tableau)

GeometricIntegrators.Methods.ExplicitEulerRK — Type

Explicit Runge-Kutta method with TableauExplicitEuler.

Reference:

Leonhard Euler.
Institutiones calculi differentialis cum eius vsu in analysi finitorum ac doctrina serierum.
Imp. Acad. Imper. Scient. Petropolitanae, Opera Omnia, Vol.X, [I.6], 1755.
In: Opera Omnia, 1st Series, Volume 11, Institutiones Calculi Integralis. Teubner, Leipzig, Pages 424-434, 1913.
Sectio secunda. Caput VII. De integratione aequationum differentialium per approximationem. Problema 85.

GeometricIntegrators.Methods.ExplicitMidpoint — Type

Explicit Runge-Kutta method with TableauExplicitMidpoint.

Reference:

Carl Runge.
Über die numerische Auflösung von Differentialgleichungen.
Mathematische Annalen, Volume 46, Pages 167-178, 1895.
doi: 10.1007/BF01446807.
Equation (2)

GeometricIntegrators.Methods.FLRK — Type

Formal Lagrangian Runge-Kutta Method

FLRK(tableau::Tableau)
FLRK(method::RKMethod)

GeometricIntegrators.Methods.ForwardEuler — Type

Explicit Runge-Kutta method with TableauForwardEuler.

Reference:

Leonhard Euler.
Institutiones calculi differentialis cum eius vsu in analysi finitorum ac doctrina serierum.
Imp. Acad. Imper. Scient. Petropolitanae, Opera Omnia, Vol.X, [I.6], 1755.
In: Opera Omnia, 1st Series, Volume 11, Institutiones Calculi Integralis. Teubner, Leipzig, Pages 424-434, 1913.
Sectio secunda. Caput VII. De integratione aequationum differentialium per approximationem. Problema 85.

GeometricIntegrators.Methods.Gauss — Type

Fully implicit Runge-Kutta method with TableauGauss.

References:

John C. Butcher.
Implicit Runge-Kutta processes.
Mathematics of Computation, Volume 18, Pages 50-64, 1964.
doi: 10.1090/S0025-5718-1964-0159424-9.

John C. Butcher.
Gauss Methods. 
In: Engquist B. (eds). Encyclopedia of Applied and Computational Mathematics. Springer, Berlin, Heidelberg. 2015.
doi: 10.1007/978-3-540-70529-1_115.

GeometricIntegrators.Methods.Heun2 — Type

Explicit Runge-Kutta method with TableauHeun2.

Reference:

Karl Heun.
Neue Methoden zur approximativen Integration der Differentialgleichungen einer unabhängigen Veränderlichen.
Zeitschrift für Mathematik und Physik, Volume 45, Pages 23-38, 1900.
Algorithm II.

GeometricIntegrators.Methods.Heun3 — Type

Explicit Runge-Kutta method with TableauHeun3.

Reference:

Karl Heun.
Neue Methoden zur approximativen Integration der Differentialgleichungen einer unabhängigen Veränderlichen.
Zeitschrift für Mathematik und Physik, Volume 45, Pages 23-38, 1900.
Algorithm VI.

GeometricIntegrators.Methods.IPRK — Type

Implicit Partitioned Runge-Kutta Method

IPRK(tableau)

GeometricIntegrators.Methods.IRK — Type

Implicit Runge-Kutta Method

IRK(tableau)

GeometricIntegrators.Methods.ImplicitEulerRK — Type

Fully implicit Runge-Kutta method with TableauImplicitEuler.

Reference:

Augustin-Louis Cauchy.
Équations différentielles ordinaires. Cours inédit (fragment). Douzième leçon.
Ed. Christian Gilain, Etudes Vivantes, 1981.
Page 102, Equation (5), Θ=1.

GeometricIntegrators.Methods.ImplicitMidpoint — Type

Fully implicit Runge-Kutta method with TableauImplicitMidpoint.

Reference:

Augustin-Louis Cauchy.
Équations différentielles ordinaires. Cours inédit (fragment). Douzième leçon.
Ed. Christian Gilain, Etudes Vivantes, 1981.
Page 102, Equation (5), Θ=1/2.

GeometricIntegrators.Methods.KraaijevangerSpijker — Type

Diagonally implicit Runge-Kutta method with TableauKraaijevangerSpijker.

Reference:

J. F. B. M. Kraaijevanger and M. N. Spijker.
Algebraic stability and error propagation in Runge-Kutta methods.
Applied Numerical Mathematics, Volume 5, Issues 1-2, Pages 71-87, 1989.
doi: 10.1016/0168-9274(89)90025-1

GeometricIntegrators.Methods.Kutta3 — Type

Explicit Runge-Kutta method with TableauKutta3.

Reference:

Wilhelm Kutta
Beitrag zur Näherungsweisen Integration totaler Differentialgleichungen
Zeitschrift für Mathematik und Physik, Volume 46, Pages 435-453, 1901.
Page 440

GeometricIntegrators.Methods.LobattoIII — Type

Runge-Kutta method with TableauLobattoIII.

References:

John C. Butcher.
Integration processes based on Radau quadrature formulas
Mathematics of Computation, Volume 18, Pages 233-244, 1964.
doi: 10.1090/S0025-5718-1964-0165693-1.

Laurent O. Jay.
Lobatto Methods.
In: Engquist B. (eds). Encyclopedia of Applied and Computational Mathematics. Springer, Berlin, Heidelberg. 2015.
doi: 10.1007/978-3-540-70529-1_123.

GeometricIntegrators.Methods.LobattoIIIA — Type

Runge-Kutta method with TableauLobattoIIIA.

References:

Byron Leonard Ehle
On Padé approximations to the exponential function and a-stable methods for the numerical solution of initial value problems.
Research Report CSRR 2010, Dept. AACS, University of Waterloo, 1969.

Laurent O. Jay.
Lobatto Methods.
In: Engquist B. (eds). Encyclopedia of Applied and Computational Mathematics. Springer, Berlin, Heidelberg. 2015.
doi: 10.1007/978-3-540-70529-1_123

GeometricIntegrators.Methods.LobattoIIIAIIIB — Type

Partitioned Runge-Kutta method with TableauLobattoIIIA for $q$ and TableauLobattoIIIB for $p$.

GeometricIntegrators.Methods.LobattoIIIAIIIĀ — Type

Partitioned Runge-Kutta method with TableauLobattoIIIA for $q$ and TableauLobattoIIIĀ for $p$.

GeometricIntegrators.Methods.LobattoIIIB — Type

Runge-Kutta method with TableauLobattoIIIB.

References:

Byron Leonard Ehle.
On Padé approximations to the exponential function and a-stable methods for the numerical solution of initial value problems.
Research Report CSRR 2010, Dept. AACS, University of Waterloo, 1969.

Laurent O. Jay.
Lobatto Methods.
In: Engquist B. (eds). Encyclopedia of Applied and Computational Mathematics. Springer, Berlin, Heidelberg. 2015.
doi: 10.1007/978-3-540-70529-1_123.

GeometricIntegrators.Methods.LobattoIIIBIIIA — Type

Partitioned Runge-Kutta method with TableauLobattoIIIB for $q$ and TableauLobattoIIIA for $p$.

GeometricIntegrators.Methods.LobattoIIIBIIIB̄ — Type

Partitioned Runge-Kutta method with TableauLobattoIIIB for $q$ and TableauLobattoIIIB̄ for $p$.

GeometricIntegrators.Methods.LobattoIIIC — Type

Runge-Kutta method with TableauLobattoIIIC.

References:

F. H. Chipman.
A-stable Runge-Kutta processes.
BIT, Volume 11, Pages 384-388, 1971.
doi: 10.1007/BF01939406.

Laurent O. Jay.
Lobatto Methods.
In: Engquist B. (eds). Encyclopedia of Applied and Computational Mathematics. Springer, Berlin, Heidelberg. 2015.
doi: 10.1007/978-3-540-70529-1_123.

GeometricIntegrators.Methods.LobattoIIICIIIC̄ — Type

Partitioned Runge-Kutta method with TableauLobattoIIIC for $q$ and TableauLobattoIIIC̄ for $p$.

GeometricIntegrators.Methods.LobattoIIIC̄IIIC — Type

Partitioned Runge-Kutta method with TableauLobattoIIIC̄ for $q$ and TableauLobattoIIIC for $p$.

GeometricIntegrators.Methods.LobattoIIID — Type

Runge-Kutta method with TableauLobattoIIID.

References:

R.P.K. Chan.
On symmetric Runge-Kutta methods of high order.
Computing, Volume 45, Pages 301-309, 1990.
doi: 10.1007/BF02238798

Laurent O. Jay.
Lobatto Methods.
In: Engquist B. (eds). Encyclopedia of Applied and Computational Mathematics. Springer, Berlin, Heidelberg. 2015.
doi: 10.1007/978-3-540-70529-1_123.

GeometricIntegrators.Methods.LobattoIIIDIIID̄ — Type

Partitioned Runge-Kutta method with TableauLobattoIIID for $q$ and TableauLobattoIIID̄ for $p$.

GeometricIntegrators.Methods.LobattoIIIE — Type

Runge-Kutta method with TableauLobattoIIIE.

References:

R.P.K. Chan.
On symmetric Runge-Kutta methods of high order.
Computing, Volume 45, Pages 301-309, 1990.
doi: 10.1007/BF02238798

Laurent O. Jay.
Lobatto Methods.
In: Engquist B. (eds). Encyclopedia of Applied and Computational Mathematics. Springer, Berlin, Heidelberg. 2015.
doi: 10.1007/978-3-540-70529-1_123.

GeometricIntegrators.Methods.LobattoIIIEIIIĒ — Type

Partitioned Runge-Kutta method with TableauLobattoIIIE for $q$ and TableauLobattoIIIĒ for $p$.

GeometricIntegrators.Methods.LobattoIIIF — Type

Runge-Kutta method with TableauLobattoIIIF.

References:

Wang Fangzong and Liao Xiaobing.
A Class of Lobatto Methods of Order 2s.
Journal of Applied Mathematics, Volume 46, Pages 6-10, 2016.

GeometricIntegrators.Methods.LobattoIIIFIIIF̄ — Type

Partitioned Runge-Kutta method with TableauLobattoIIIF for $q$ and TableauLobattoIIIF̄ for $p$.

GeometricIntegrators.Methods.LobattoIIIF̄ — Type

Runge-Kutta method with TableauLobattoIIIF̄.

GeometricIntegrators.Methods.LobattoIIIF̄IIIF — Type

Partitioned Runge-Kutta method with TableauLobattoIIIF̄ for $q$ and TableauLobattoIIIF for $p$.

GeometricIntegrators.Methods.LobattoIIIG — Type

Runge-Kutta method with TableauLobattoIIIG.

GeometricIntegrators.Methods.LobattoIIIGIIIḠ — Type

Partitioned Runge-Kutta method with TableauLobattoIIIG for $q$ and TableauLobattoIIIḠ for $p$.

GeometricIntegrators.Methods.PartitionedGauss — Type

Partitioned Runge-Kutta method TableauGauss for both $q$ and $p$.

References:

John C. Butcher.
Implicit Runge-Kutta processes.
Mathematics of Computation, Volume 18, Pages 50-64, 1964.
doi: 10.1090/S0025-5718-1964-0159424-9.

John C. Butcher.
Gauss Methods. 
In: Engquist B. (eds). Encyclopedia of Applied and Computational Mathematics. Springer, Berlin, Heidelberg. 2015.
doi: 10.1007/978-3-540-70529-1_115.

GeometricIntegrators.Methods.ProjectedVPRK — Type

Projected VPRK Method

A projected Variational Runge-Kutta Method consists of a VPRK method and a projection method:

ProjectedVPRK(vprk::VPRKMethod, projection::ProjectionMethod)
ProjectedVPRK(tableau::AbstractTableau, projection::ProjectionMethod, args...; kwargs...) = ProjectedVPRK(VPRK(tableau, args...; kwargs...), projection)

GeometricIntegrators.Methods.QinZhang — Type

Diagonally implicit Runge-Kutta method with TableauQinZhang.

Reference:

M.-Z. Qin and M.-Q. Zhang.
Symplectic Runge-Kutta algorithms for Hamilton systems.
Journal of Computational Mathematics, Supplementary Issue, Pages 205-215, 1992.

GeometricIntegrators.Methods.RK21 — Type

Explicit Runge-Kutta method with TableauRK21.

Reference:

Karl Heun.
Neue Methoden zur approximativen Integration der Differentialgleichungen einer unabhängigen Veränderlichen.
Zeitschrift für Mathematik und Physik, Volume 45, Pages 23-38, 1900.
Algorithm II.

GeometricIntegrators.Methods.RK22 — Type

Explicit Runge-Kutta method with TableauRK22.

Reference:

Carl Runge
Über die numerische Auflösung von Differentialgleichungen.
Mathematische Annalen, Volume 46, Pages 167-178, 1895.
doi: 10.1007/BF01446807.
Equation (3)

GeometricIntegrators.Methods.RK31 — Type

Explicit Runge-Kutta method with TableauRK31.

Reference:

John C. Butcher
Numerical Methods for Ordinary Differential Equations. Wiley, 2016.
Page 99

GeometricIntegrators.Methods.RK32 — Type

Explicit Runge-Kutta method with TableauRK32.

Reference:

Wilhelm Kutta
Beitrag zur Näherungsweisen Integration totaler Differentialgleichungen
Zeitschrift für Mathematik und Physik, Volume 46, Pages 435-453, 1901.
Page 440

GeometricIntegrators.Methods.RK4 — Type

Explicit Runge-Kutta method with TableauRK4.

Reference:

Wilhelm Kutta
Beitrag zur Näherungsweisen Integration totaler Differentialgleichungen
Zeitschrift für Mathematik und Physik, Volume 46, Pages 435-453, 1901.
Page 443

GeometricIntegrators.Methods.RK41 — Type

Explicit Runge-Kutta method with TableauRK41.

Reference:

Wilhelm Kutta
Beitrag zur Näherungsweisen Integration totaler Differentialgleichungen
Zeitschrift für Mathematik und Physik, Volume 46, Pages 435-453, 1901.
Page 443

GeometricIntegrators.Methods.RK416 — Type

Explicit Runge-Kutta method with TableauRK416.

Reference:

Wilhelm Kutta
Beitrag zur Näherungsweisen Integration totaler Differentialgleichungen
Zeitschrift für Mathematik und Physik, Volume 46, Pages 435-453, 1901.
Page 443

GeometricIntegrators.Methods.RK42 — Type

Explicit Runge-Kutta method with TableauRK42.

Reference:

John C. Butcher
Numerical Methods for Ordinary Differential Equations. Wiley, 2016.
Page 102

GeometricIntegrators.Methods.RK438 — Type

Explicit Runge-Kutta method with TableauRK438.

Reference:

Wilhelm Kutta
Beitrag zur Näherungsweisen Integration totaler Differentialgleichungen
Zeitschrift für Mathematik und Physik, Volume 46, Pages 435-453, 1901.
Page 441

GeometricIntegrators.Methods.RK5 — Type

Explicit Runge-Kutta method with TableauRK5.

Reference:

John C. Butcher
Numerical Methods for Ordinary Differential Equations. Wiley, 2016.
Page 103

GeometricIntegrators.Methods.RadauIA — Type

Runge-Kutta method with TableauRadauIA.

References:

Byron Leonard Ehle
On Padé approximations to the exponential function and a-stable methods for the numerical solution of initial value problems.
Research Report CSRR 2010, Dept. AACS, University of Waterloo, 1969.

GeometricIntegrators.Methods.RadauIB — Type

Runge-Kutta method with TableauRadauIB.

Reference:

Sun Geng
Construction of high order symplectic Runge-Kutta methods
Journal of Computational Mathematics, Volume 11, Pages 250-260, 1993.

GeometricIntegrators.Methods.RadauIIA — Type

Runge-Kutta method with TableauRadauIIA.

References:

Byron Leonard Ehle
On Padé approximations to the exponential function and a-stable methods for the numerical solution of initial value problems.
Research Report CSRR 2010, Dept. AACS, University of Waterloo, 1969.

Owe Axelsson.
A class of A-stable methods.
BIT, Volume 9, Pages 185-199, 1969.
doi: 10.1007/BF01946812.

Ernst Hairer and Gerhard Wanner.
Radau Methods.
In: Engquist B. (eds). Encyclopedia of Applied and Computational Mathematics. Springer, Berlin, Heidelberg. 2015.
doi: 10.1007/978-3-540-70529-1_139.

GeometricIntegrators.Methods.RadauIIB — Type

Runge-Kutta method with TableauRadauIIB.

Reference:

Sun Geng
Construction of high order symplectic Runge-Kutta methods
Journal of Computational Mathematics, Volume 11, Pages 250-260, 1993.

GeometricIntegrators.Methods.Ralston2 — Type

Explicit Runge-Kutta method with TableauRalston2.

Reference:

Anthony Ralston.
Runge-Kutta Methods with Minimum Error Bounds.
Mathematics of Computation, Volume 16, Pages 431-437, 1962.
doi: 10.1090/S0025-5718-1962-0150954-0.
Equation (3.5)

GeometricIntegrators.Methods.Ralston3 — Type

Explicit Runge-Kutta method with TableauRalston3.

Reference:

Anthony Ralston.
Runge-Kutta Methods with Minimum Error Bounds.
Mathematics of Computation, Volume 16, Pages 431-437, 1962.
doi: 10.1090/S0025-5718-1962-0150954-0.
Equation (4.10)

GeometricIntegrators.Methods.Runge2 — Type

Explicit Runge-Kutta method with TableauRunge2.

Reference:

Carl Runge
Über die numerische Auflösung von Differentialgleichungen.
Mathematische Annalen, Volume 46, Pages 167-178, 1895.
doi: 10.1007/BF01446807.
Equation (3)

GeometricIntegrators.Methods.SRK3 — Type

Fully implicit Runge-Kutta method with TableauSRK3.

Reference:

Shan Zhao and Guo-Wei Wei.
A unified discontinuous Galerkin framework for time integration.
Mathematical Methods in the Applied Sciences, Volume 37, Issue 7, Pages 1042-1071, 2014.
doi: 10.1002/mma.2863.

GeometricIntegrators.Methods.SSPRK2 — Type

Explicit Runge-Kutta method with TableauSSPRK2.

Reference:

Chi-Wang Shu, Stanley Osher.
Efficient implementation of essentially non-oscillatory shock-capturing schemes.
Journal of Computational Physics, Volume 77, Issue 2, Pages 439-471, 1988.
doi: 10.1016/0021-9991(88)90177-5.
Equation (2.16)

GeometricIntegrators.Methods.SSPRK3 — Type

Explicit Runge-Kutta method with TableauSSPRK3.

Reference:

Chi-Wang Shu, Stanley Osher.
Efficient implementation of essentially non-oscillatory shock-capturing schemes.
Journal of Computational Physics, Volume 77, Issue 2, Pages 439-471, 1988.
doi: 10.1016/0021-9991(88)90177-5.
Equation (2.18)

GeometricIntegrators.Methods.SymplecticEulerA — Type

SymplecticEulerA

Symplectic Euler method using explicit Euler for q and implicit Euler for p.

GeometricIntegrators.Methods.SymplecticEulerB — Type

SymplecticEulerB

Symplectic Euler method using implicit Euler for q and explicit Euler for p.

GeometricIntegrators.Methods.VPRK — Type

Variational Partitioned Runge-Kutta Method

VPRK(tableau::PartitionedTableau, d=nothing)
VPRK(tableau::Tableau, args...; kwargs...) = VPRK(PartitionedTableau(tableau), args...; kwargs...)
VPRK(tableau1::Tableau, tableau2::Tableau, args...; kwargs...) = VPRK(PartitionedTableau(tableau1, tableau2), args...; kwargs...)

GeometricIntegrators.Methods.VPRKGauss — Type

Variational Partitioned Runge-Kutta Method that uses Gauss-Legendre coefficients for both $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIII — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIII for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIA — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIA for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIAIIIB — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIA for the coefficients $a_{ij}$ and TableauLobattoIIIB for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIAIIIĀ — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIA for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIB — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIB for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIBIIIA — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIB for the coefficients $a_{ij}$ and TableauLobattoIIIA for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIBIIIB̄ — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIB for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIC — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIC for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIICIIIC̄ — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIC for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIC̄IIIC — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIC̄ for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIID — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIID for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIDIIID̄ — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIID for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIE — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIE for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIEIIIĒ — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIE for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIF — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIF for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIFIIIF̄ — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIF for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIF̄ — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIF̄ for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIF̄IIIF — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIF̄ for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIG — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIG for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKLobattoIIIGIIIḠ — Type

Variational Partitioned Runge-Kutta Method that uses TableauLobattoIIIG for the coefficients $a_{ij}$ and its symplectic conjugate for the coefficients $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKRadauIIA — Type

Variational Partitioned Runge-Kutta Method that uses TableauRadauIIA for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPRKRadauIIB — Type

Variational Partitioned Runge-Kutta Method that uses TableauRadauIIB for both coefficients $a_{ij}$ and $\bar{a}_{ij}$.

GeometricIntegrators.Methods.VPSRK3 — Type

Variational Partitioned Runge-Kutta Method that uses the symmetric TableauSRK3 coefficients with 3 stages for both $a_{ij}$ and $\bar{a}_{ij}$.

Base.show — Method

Base.show(io::IO, ml::MethodList)
Base.show(io::IO, ::MIME"text/markdown", ml::MethodList)

Pretty-print MethodList.

GeometricIntegrators.Methods.PRK — Method

Partitioned Runge-Kutta Method

PRK(tableau)

Returns an explicit or implicit partitioned Runge-Kutta method depending on the tableau.

GeometricIntegrators.Methods.RK — Method

Runge-Kutta Method

RK(tableau)

Returns an explicit, implicit or diagonally implicit Runge-Kutta method depending on the tableau.

GeometricIntegrators.Methods.VPRKpInternal — Method

VPRK integrator with projection on internal stages

GeometricIntegrators.Methods.VPRKpLegendre — Method

VPRK integrator with Legendre projection

GeometricIntegrators.Methods.VPRKpMidpoint — Method

VPRK integrator with Midpoint projection

GeometricIntegrators.Methods.VPRKpSecondary — Method

VPRK integrator with projection on secondary constraint

GeometricIntegrators.Methods.VPRKpStandard — Method

VPRK integrator with standard projection

GeometricIntegrators.Methods.VPRKpSymmetric — Method

VPRK integrator with symmetric projection

GeometricIntegrators.Methods.VPRKpSymplectic — Method

VPRK integrator with symplectic projection

GeometricIntegrators.Methods.VPRKpVariational — Method

VPRK integrator with variational projection

GeometricIntegrators.Methods.VPRKpVariationalP — Method

VPRK integrator with variational projection on P

GeometricIntegrators.Methods.VPRKpVariationalQ — Method

VPRK integrator with variational projection on Q