Solutions

GeometricIntegrators.Solutions.AtomicSolution — Type

Abstract atomistic or single-step solution.

Constructors:

AtomicSolution(equation)
AtomicSolution(solution)
AtomicSolution(solution, integrator)

Automatically construct the appropriate atomistic solution based on the given equation or solution type. If an integrator is provided as, the internal field of the atomic solution is constructed according to the internal state of the integrator as obtained from the function get_internal_variables.

GeometricIntegrators.Solutions.AtomicSolutionDAE — Type

Atomic solution for an DAE.

Parameters

DT: data type

TT: time step type

AT: array type

IT: internal variable types

Fields

t: time of current time step

t̄: time of previous time step

q: current solution of q

q̄: previous solution of q

q̃: compensated summation error of q

λ: current solution of λ

λ̄: previous solution of λ

v: vector field of q

v̄: vector field of q̄

u: projective vector field of q

ū: projective vector field of q̄

internal: internal variables of the integrator (e.g., internal stages of a Runge-Kutta methods or solver output)

Constructors

AtomicSolutionDAE{DT,TT,AT,IT}(nd, nm, internal::IT)
AtomicSolutionDAE{DT,TT,AT,IT}(t::TT, q::AT, λ::AT, internal::IT)
AtomicSolutionDAE(DT, TT, AT, nd, nm, internal::IT=NamedTuple())
AtomicSolutionDAE(t::TT, q::AT, λ::AT, internal::IT=NamedTuple())

nd: dimension of the state vector

nm: dimension of the constraint

GeometricIntegrators.Solutions.AtomicSolutionODE — Type

Atomic solution for an ODE.

Parameters

DT: data type

TT: time step type

AT: array type

IT: internal variable types

Fields

t: time of current time step

t̄: time of previous time step

q: current solution of q

q̄: previous solution of q

q̃: compensated summation error of q

v: vector field of q

v̄: vector field of q̄

internal: internal variables of the integrator (e.g., internal stages of a Runge-Kutta methods or solver output)

Constructors

AtomicSolutionODE{DT,TT,AT,IT}(nd, internal::IT)
AtomicSolutionODE{DT,TT,AT,IT}(t::TT, q::AT, internal::IT)
AtomicSolutionODE(DT, TT, AT, nd, internal::IT=NamedTuple())
AtomicSolutionODE(t::TT, q::AT, internal::IT=NamedTuple())

nd: dimension of the state vector

GeometricIntegrators.Solutions.AtomicSolutionPDAE — Type

Atomic solution for an PDAE.

Parameters

DT: data type

TT: time step type

AT: array type

IT: internal variable types

Fields

t: time of current time step

t̄: time of previous time step

q: current solution of q

q̄: previous solution of q

q̃: compensated summation error of q

p: current solution of p

p̄: previous solution of p

p̃: compensated summation error of p

λ: current solution of λ

λ̄: previous solution of λ

v: vector field of q

v̄: vector field of q̄

f: vector field of p

f̄: vector field of p̄

u: projective vector field of q

ū: projective vector field of q̄

g: projective vector field of p

ḡ: projective vector field of p̄

internal: internal variables of the integrator (e.g., internal stages of a Runge-Kutta methods or solver output)

Constructors

AtomicSolutionPDAE{DT,TT,AT,IT}(nd, nm, internal::IT)
AtomicSolutionPDAE{DT,TT,AT,IT}(t::TT, q::AT, p::AT, λ::AT, internal::IT)
AtomicSolutionPDAE(DT, TT, AT, nd, nm, internal::IT=NamedTuple())
AtomicSolutionPDAE(t::TT, q::AT, p::AT, λ::AT, internal::IT=NamedTuple())

nd: dimension of the state vector

nm: dimension of the constraint

GeometricIntegrators.Solutions.AtomicSolutionPODE — Type

Atomic solution for an PODE.

Parameters

DT: data type

TT: time step type

AT: array type

IT: internal variable types

Fields

t: time of current time step

t̄: time of previous time step

q: current solution of q

q̄: previous solution of q

q̃: compensated summation error of q

p: current solution of p

p̄: previous solution of p

p̃: compensated summation error of p

v: vector field of q

v̄: vector field of q̄

f: vector field of p

f̄: vector field of p̄

internal: internal variables of the integrator (e.g., internal stages of a Runge-Kutta methods or solver output)

Constructors

AtomicSolutionPODE{DT,TT,AT,IT}(nd, internal::IT)
AtomicSolutionPODE{DT,TT,AT,IT}(t::TT, q::AT, p::AT, internal::IT)
AtomicSolutionPODE(DT, TT, AT, nd, internal::IT=NamedTuple())
AtomicSolutionPODE(t::TT, q::AT, p::AT, internal::IT=NamedTuple())

nd: dimension of the state vector

GeometricIntegrators.Solutions.Solution — Type

Solution(equation, Δt, ntime; kwargs...)

Create the appropriate Solution for the given equation type for a simulation with ntime time steps of step size Δt.

GeometricIntegrators.Solutions.SolutionDAE — Type

SolutionDAE: Solution of a differential algebraic equation

Contains all fields necessary to store the solution of an DAE.

Fields

nd: dimension of the dynamical variable $q$

nm: dimension of the constraint submanifold

nt: number of time steps to store

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ni] with q[:,0,:] the initial conditions

λ: Lagrange multiplier λ[nd, nt+1, ni]

ntime: number of time steps to compute

nsave: store every nsave'th time step (default: 1)

nwrite: save data to disk after every nwrite'th time step (default: ntime)

counter: counter for copied solution entries

woffset: counter for file offset

h5: HDF5 file for storage

Constructors

SSolutionDAE(equation, Δt, ntimesteps; nsave=DEFAULT_NSAVE, nwrite=DEFAULT_NWRITE, filename=nothing)
SSolutionDAE(t::TimeSeries, q::SDataSeries, λ::SDataSeries, ntimesteps)
SSolutionDAE(file::String)
PSolutionDAE(equation, Δt, ntimesteps; nsave=DEFAULT_NSAVE, nwrite=DEFAULT_NWRITE, filename=nothing)
PSolutionDAE(t::TimeSeries, q::PDataSeries, λ::PDataSeries, ntimesteps)
PSolutionDAE(file::String)

The constructors SSolutionDAE create a SolutionDAE with internal data structures for serial simulations (i.e., standard arrays), while the constructors PSolutionDAE create a SolutionDAE with internal data structures for parallel simulations (i.e., shared arrays).

The usual way to initialise a Solution is by passing an equation, which for SolutionDAE has to be an DAE, a time step Δt and the number of time steps ntimesteps. The optional parameters nsave and nwrite determine the intervals for storing the solution and writing to file, i.e., if nsave > 1 only every nsave'th solution is actually stored, and every nwrite'th time step the solution is stored to disk.

The other constructors, either passing a TimeSeries and two DataSeries or a filename are used to read data from previous simulations.

GeometricIntegrators.Solutions.SolutionODE — Type

SolutionODE: Solution of an ordinary differential equation

Contains all fields necessary to store the solution of an ODE.

Fields

nd: dimension of the dynamical variable $q$

nt: number of time steps to store

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ni] with q[:,0,:] the initial conditions

ntime: number of time steps to compute

nsave: store every nsave'th time step (default: 1)

nwrite: save data to disk after every nwrite'th time step (default: ntime)

counter: counter for copied solution entries

woffset: counter for file offset

h5: HDF5 file for storage

Constructors

SSolutionODE(equation, Δt, ntimesteps; nsave=DEFAULT_NSAVE, nwrite=DEFAULT_NWRITE, filename=nothing)
SSolutionODE(t::TimeSeries, q::SDataSeries, ntimesteps)
SSolutionODE(file::String)
PSolutionODE(equation, Δt, ntimesteps; nsave=DEFAULT_NSAVE, nwrite=DEFAULT_NWRITE, filename=nothing)
PSolutionODE(t::TimeSeries, q::PDataSeries, ntimesteps)
PSolutionODE(file::String)

The constructors SSolutionODE create a SolutionODE with internal data structures for serial simulations (i.e., standard arrays), while the constructors PSolutionODE create a SolutionODE with internal data structures for parallel simulations (i.e., shared arrays).

The usual way to initialise a Solution is by passing an equation, which for SolutionODE has to be an ODE or SODE, a time step Δt and the number of time steps ntimesteps. The optional parameters nsave and nwrite determine the intervals for storing the solution and writing to file, i.e., if nsave > 1 only every nsave'th solution is actually stored, and every nwrite'th time step the solution is stored to disk.

The other constructors, either passing a TimeSeries and a DataSeries or a filename are used to read data from previous simulations.

GeometricIntegrators.Solutions.SolutionPDAE — Type

SolutionPDAE: Solution of a partitioned differential algebraic equation

Contains all fields necessary to store the solution of an PDAE.

Fields

nd: dimension of the dynamical variable $q$

nm: dimension of the constraint submanifold

nt: number of time steps to store

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ni] with q[:,0,:] the initial conditions

p: solution p[nd, nt+1, ni] with p[:,0,:] the initial conditions

λ: Lagrange multiplier λ[nd, nt+1, ni]

ntime: number of time steps to compute

nsave: store every nsave'th time step (default: 1)

nwrite: save data to disk after every nwrite'th time step (default: ntime)

counter: counter for copied solution entries

woffset: counter for file offset

h5: HDF5 file for storage

Constructors

SSolutionPDAE(equation, Δt, ntimesteps; nsave=DEFAULT_NSAVE, nwrite=DEFAULT_NWRITE, filename=nothing)
SSolutionPDAE(t::TimeSeries, q::SDataSeries, p::SDataSeries, λ::SDataSeries, ntimesteps)
SSolutionPDAE(file::String)
PSolutionPDAE(equation, Δt, ntimesteps; nsave=DEFAULT_NSAVE, nwrite=DEFAULT_NWRITE, filename=nothing)
PSolutionPDAE(t::TimeSeries, q::PDataSeries, p::PDataSeries, λ::PDataSeries, ntimesteps)
PSolutionPDAE(file::String)

The constructors SSolutionPDAE create a SolutionPDAE with internal data structures for serial simulations (i.e., standard arrays), while the constructors PSolutionPDAE create a SolutionPDAE with internal data structures for parallel simulations (i.e., shared arrays).

The usual way to initialise a Solution is by passing an equation, which for SolutionPDAE has to be an PDAE, HDAE, IDAE or LDAE, a time step Δt and the number of time steps ntimesteps. The optional parameters nsave and nwrite determine the intervals for storing the solution and writing to file, i.e., if nsave > 1 only every nsave'th solution is actually stored, and every nwrite'th time step the solution is stored to disk.

The other constructors, either passing a TimeSeries and three DataSeries or a filename are used to read data from previous simulations.

GeometricIntegrators.Solutions.SolutionPODE — Type

SolutionPODE: Solution of a partitioned ordinary differential equation

Contains all fields necessary to store the solution of an PODE.

Fields

nd: dimension of the dynamical variable $q$

nt: number of time steps to store

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ni] with q[:,0,:] the initial conditions

p: solution p[nd, nt+1, ni] with p[:,0,:] the initial conditions

ntime: number of time steps to compute

nsave: store every nsave'th time step (default: 1)

nwrite: save data to disk after every nwrite'th time step (default: ntime)

counter: counter for copied solution entries

woffset: counter for file offset

h5: HDF5 file for storage

Constructors

SSolutionPODE(equation, Δt, ntimesteps; nsave=DEFAULT_NSAVE, nwrite=DEFAULT_NWRITE, filename=nothing)
SSolutionPODE(t::TimeSeries, q::SDataSeries, p::SDataSeries, ntimesteps)
SSolutionPODE(file::String)
PSolutionPODE(equation, Δt, ntimesteps; nsave=DEFAULT_NSAVE, nwrite=DEFAULT_NWRITE, filename=nothing)
PSolutionPODE(t::TimeSeries, q::PDataSeries, p::PDataSeries, ntimesteps)
PSolutionPODE(file::String)

The constructors SSolutionPODE create a SolutionPODE with internal data structures for serial simulations (i.e., standard arrays), while the constructors PSolutionPODE create a SolutionPODE with internal data structures for parallel simulations (i.e., shared arrays).

The usual way to initialise a Solution is by passing an equation, which for SolutionPODE has to be an PODE, HODE, IODE or LODE, a time step Δt and the number of time steps ntimesteps. The optional parameters nsave and nwrite determine the intervals for storing the solution and writing to file, i.e., if nsave > 1 only every nsave'th solution is actually stored, and every nwrite'th time step the solution is stored to disk.

The other constructors, either passing a TimeSeries and two DataSeries or a filename are used to read data from previous simulations.

GeometricIntegrators.Common.write_to_hdf5 — Function

Append solution to HDF5 file.

GeometricIntegrators.Common.write_to_hdf5 — Function

Append solution to HDF5 file.

GeometricIntegrators.Common.write_to_hdf5 — Method

Creates HDF5 file, writes solution to file, and closes file.

GeometricIntegrators.Common.write_to_hdf5 — Method

writetohdf5: Wrapper for saving Solution to HDF5 file.

GeometricIntegrators.Solutions.ParallelSolution — Function

ParallelSolution(equation, Δt, ntime; kwargs...)

Create the appropriate ParallelSolution for the given equation type for a simulation with ntime time steps of step size Δt.

GeometricIntegrators.Solutions.copy_solution! — Method

copy_solution!(solution, atomic_solution, n, m)

Copy solution for time step n and initial condition m from atomic solution to solution object.

GeometricIntegrators.Solutions.createHDF5 — Method

createHDF5: Creates or opens HDF5 file.

GeometricIntegrators.Solutions.create_hdf5! — Method

Creates HDF5 file and initialises datasets for solution object.

GeometricIntegrators.Solutions.create_hdf5 — Method

Creates HDF5 file and initialises datasets for deterministic solution object.

GeometricIntegrators.Solutions.save_attributes — Method

save_attributes: Saves attributes of Deterministic Solutions to HDF5 file.

GeometricIntegrators.Solutions.save_attributes — Method

save_attributes: Saves common attributes of Solution to HDF5 file.