Solutions

Solutions

GeometricIntegrators.Solutions.PSolutionPDAE — Type.

Parallel Solution of a partitioned differential algebraic equation.

source

GeometricIntegrators.Solutions.SSolutionPDAE — Type.

Serial Solution of a partitioned differential algebraic equation.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for PSDE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for implicit ODE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for ODE and split ODE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for partitioned DAE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for implicit DAE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for partitioned ODE.

source

GeometricIntegrators.Solutions.Solution — Type.

Print error for solutions of equations not implemented, yet.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for variational ODE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for SDE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for DAE.

source

GeometricIntegrators.Solutions.SolutionDAE — Type.

Solution of a differential algebraic equation.

source

GeometricIntegrators.Solutions.SolutionODE — Type.

SolutionODE: Solution of an ordinary differential equation

Contains all fields necessary to store the solution of an ODE.

Fields

nd: dimension of the dynamical variable $q$

nt: number of time steps to store

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ni] with q[:,0,:] the initial conditions

ntime: number of time steps to compute

nsave: save every nsave'th time step

source

GeometricIntegrators.Solutions.SolutionPODE — Type.

Solution of a partitioned ordinary differential equation.

source

GeometricIntegrators.Solutions.SolutionPSDE — Type.

SolutionPSDE: Solution of a partitioned stochastic differential equation

Contains all fields necessary to store the solution of a PSDE or SPSDE

Fields

conv: type of the solution: "strong" or "weak"

nd: dimension of the dynamical variable $q$

nm: dimension of the Wiener process

nt: number of time steps to store

ns: number of sample paths

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ns, ni] with q[:,0,:,:] the initial conditions

p: solution p[nd, nt+1, ns, ni] with p[:,0,:,:] the initial conditions

W: Wiener process driving the stochastic processes q and p

K: integer parameter defining the truncation of the increments of the Wiener process (for strong solutions),

  A = √(2 K Δt |log Δt|) due to Milstein & Tretyakov; if K=0 no truncation

ntime: number of time steps to compute

nsave: save every nsave'th time step

source

GeometricIntegrators.Solutions.SolutionSDE — Type.

SolutionSDE: Solution of a stochastic differential equation

Contains all fields necessary to store the solution of an SDE.

Fields

conv: type of the solution: "strong" or "weak"

nd: dimension of the dynamical variable $q$

nm: dimension of the Wiener process

nt: number of time steps to store

ns: number of sample paths

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ns, ni] with q[:,0,:,:] the initial conditions

W: Wiener process driving the stochastic process q

K: integer parameter defining the truncation of the increments of the Wiener process (for strong solutions),

  A = √(2 K Δt |log Δt|) due to Milstein & Tretyakov; if K=0 no truncation

ntime: number of time steps to compute

nsave: save every nsave'th time step

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file. offset - start writing q at the position offset+2 offset2- start writing ΔW, ΔZ at the position offset2+1

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file. offset - start writing q at the position offset+2 offset2- start writing ΔW, ΔZ at the position offset2+1

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for SDE solution object. It is implemented as one fucntion for all NQ and NW cases, rather than several separate cases as was done for SolutionODE. nt - the total number of time steps to store ntime - the total number of timesteps to be computed

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for SDE solution object. It is implemented as one fucntion for all NQ and NW cases, rather than several separate cases as was done for SolutionODE. nt - the total number of time steps to store ntime - the total number of timesteps to be computed

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for DAE solution object.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for DAE solution object.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for ODE solution object.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for ODE solution object.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for PDAE solution object with single initial condition.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for PDAE solution object with multiple initial conditions.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for PODE solution object.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for PODE solution object.

source

GeometricIntegrators.Solutions.get_initial_conditions! — Method.

copies the m-th initial condition for the k-th sample path from sol.q to q

source

GeometricIntegrators.Solutions.createHDF5 — Function.

createHDF5: Creates or opens HDF5 file. A version for StochasticSolution. It does not create attributes and does not write the time array t, like the version above does. Instead these are set in create_hdf5(), so that arrays larger than currently held in the solution structure can be created in the file. In the future it would be better to rewrite the function above, so that it is universal for all solution structures.

source

GeometricIntegrators.Solutions.createHDF5 — Function.

createHDF5: Creates or opens HDF5 file.

source

GeometricIntegrators.Solutions.writeSolutionToHDF5 — Method.

Creates HDF5 file, writes solution to file, and closes file.

source