Solutions

GeometricSolutions.EnsembleSolution — Type

EnsembleSolution: Collection of all solutions of an EnsembleProblem.

GeometricSolutions.GeometricSolution — Type

GeometricSolution: Solution of a geometric differential equation

Contains all fields necessary to store the solution of a GeometricProblem.

Fields

t: time steps

s: NamedTuple of DataSeries for each solution component

step: store every step'th time step (default: 1)

nstore: number of time steps to store

offset: offset of current time step

Constructors

GeometricSolution(problem, step = 1)

The usual way to initialise a Solution is by passing a GeometricEquations.EquationProblem, which can for example be an GeometricEquations.ODEProblem or GeometricEquations.PODEProblem. The optional parameter step determines the intervals for storing the solution, i.e., if step > 1 only every step'th solution is actually stored.

Solution Steps

GeometricIntegrators.Integrators.SolutionStep — Type

Abstract atomic or single-step solution.

Constructors:

SolutionStep(equation)
SolutionStep(solution)
SolutionStep(solution, integrator)

Automatically construct the appropriate atomic solution based on the given equation or solution type. If an integrator is provided as, the internal field of the atomic solution is constructed according to the internal state of the integrator as obtained from the function internal_variables.

Base.copy! — Method

Copy the initial conditions of a EquationProblem to the current state of an atomic solution.

GeometricIntegrators.Integrators.current — Function

Returns a NamedTuple with the solution of the current time step.

GeometricIntegrators.Integrators.previous — Function

Returns a NamedTuple with the solution of the previous time step.

GeometricIntegrators.Integrators.SolutionStepDAE — Type

Solution step for a DAEProblem.

Parameters

DT: data type

TT: time step type

AT: array type

IT: internal variable types

Fields

t: time of current time step

t̄: time of previous time steps

q: current solution of q

λ: current solution of λ

μ: current solution of μ

q̄: previous solutions of q

λ̄: previous solutions of λ

μ̄: previous solution of μ

v: vector field of q

v̄: vector field of q̄

u: projective vector field of q

ū: projective vector field of q̄

q̃: compensated summation error of q

internal: internal variables of the integrator (e.g., internal stages of a Runge-Kutta methods or solver output)

Constructors

SolutionStepDAE(t::TT, q::AT, λ::AT, internal::IT=NamedTuple())

GeometricIntegrators.Integrators.SolutionStepODE — Type

Solution step for an ODEProblem.

Parameters

DT: data type

TT: time step type

AT: array type

IT: internal variable types

Fields

t: time of current time step

t̄: time of previous time step

q: current solution of q

q̄: previous solution of q

v: vector field of q

v̄: vector field of q̄

q̃: compensated summation error of q

internal: internal variables of the integrator (e.g., internal stages of a Runge-Kutta methods or solver output)

Constructors

SolutionStepODE(t::TT, q::AT; nhistory=1, internal::IT=NamedTuple())

GeometricIntegrators.Integrators.SolutionStepPDAE — Type

Solution step for a PDAEProblem.

Parameters

DT: data type

TT: time step type

AT: array type

IT: internal variable types

Fields

t: time of current time step

t̄: time of previous time steps

q: current solution of q

p: current solution of p

λ: current solution of λ

μ: current solution of μ

q̄: previous solution of q

p̄: previous solution of p

λ̄: previous solution of λ

μ̄: previous solution of μ

v: vector field of q

f: vector field of p

v̄: vector field of q̄

f̄: vector field of p̄

u: projective vector field of q

ū: projective vector field of q̄

g: projective vector field of p

ḡ: projective vector field of p̄

q̃: compensated summation error of q

p̃: compensated summation error of p

internal: internal variables of the integrator (e.g., internal stages of a Runge-Kutta methods or solver output)

Constructors

SolutionStepPDAE(t::TT, q::AT, p::AT, λ::AT; nhistory=1, internal::IT=NamedTuple())

GeometricIntegrators.Integrators.SolutionStepPODE — Type

Solution step for a PODEProblem.

Parameters

DT: data type

TT: time step type

AT: array type

IT: internal variable types

Fields

t: time of current time step

t̄: time of previous time step

q: current solution of q

p: current solution of p

q̄: previous solution of q

p̄: previous solution of p

v: vector field of q

f: vector field of p

v̄: vector field of q̄

f̄: vector field of p̄

q̃: compensated summation error of q

p̃: compensated summation error of p

internal: internal variables of the integrator (e.g., internal stages of a Runge-Kutta methods or solver output)

Constructors

SolutionStepPODE(t::TT, q::AT, p::AT; nhistory=1, internal::IT=NamedTuple())