Solutions

GeometricSolutions.GeometricSolution — Type

GeometricSolution: Solution of an ordinary differential equation

Contains all fields necessary to store the solution of an ODE.

Fields

nt: number of time steps to store

t: time steps

s: NamedTuple of DataSeries for each solution component

ntime: number of time steps to compute

nsave: store every nsave'th time step (default: 1)

Constructors

GeometricSolution(problem; nsave=DEFAULT_NSAVE)
GeometricSolution(t::TimeSeries, q::DataSeries, ntimesteps)

The usual way to initialise a Solution is by passing an equation, which for GeometricSolution has to be an ODEProblem or SODEProblem, a time step Δt and the number of time steps ntimesteps. The optional parameters nsave and nwrite determine the intervals for storing the solution and writing to file, i.e., if nsave > 1 only every nsave'th solution is actually stored, and every nwrite'th time step the solution is stored to disk.

The other constructors, either passing a TimeSeries and a DataSeries or a filename are used to read data from previous simulations.

Solution Steps

GeometricIntegrators.Solutions.SolutionStep — Type

Abstract atomic or single-step solution.

Constructors:

SolutionStep(equation)
SolutionStep(solution)
SolutionStep(solution, integrator)

Automatically construct the appropriate atomic solution based on the given equation or solution type. If an integrator is provided as, the internal field of the atomic solution is constructed according to the internal state of the integrator as obtained from the function get_internal_variables.

GeometricIntegrators.Solutions.SolutionStepDAE — Type

Solution step for a DAEProblem.

Parameters

DT: data type

TT: time step type

AT: array type

IT: internal variable types

Fields

t: time of current time step

t̄: time of previous time step

q: current solution of q

q̄: previous solution of q

q̃: compensated summation error of q

λ: current solution of λ

λ̄: previous solution of λ

v: vector field of q

v̄: vector field of q̄

u: projective vector field of q

ū: projective vector field of q̄

internal: internal variables of the integrator (e.g., internal stages of a Runge-Kutta methods or solver output)

Constructors

SolutionStepDAE(t::TT, q::AT, λ::AT, internal::IT=NamedTuple())

GeometricIntegrators.Solutions.SolutionStepODE — Type

Solution step for an ODEProblem.

Parameters

DT: data type

TT: time step type

AT: array type

IT: internal variable types

Fields

t: time of current time step

t̄: time of previous time step

q: current solution of q

q̄: previous solution of q

q̃: compensated summation error of q

v: vector field of q

v̄: vector field of q̄

internal: internal variables of the integrator (e.g., internal stages of a Runge-Kutta methods or solver output)

Constructors

SolutionStepODE(t::TT, q::AT, internal::IT=NamedTuple())

GeometricIntegrators.Solutions.SolutionStepPDAE — Type

Solution step for a PDAEProblem.

Parameters

DT: data type

TT: time step type

AT: array type

IT: internal variable types

Fields

t: time of current time step

t̄: time of previous time step

q: current solution of q

q̄: previous solution of q

q̃: compensated summation error of q

p: current solution of p

p̄: previous solution of p

p̃: compensated summation error of p

λ: current solution of λ

λ̄: previous solution of λ

v: vector field of q

v̄: vector field of q̄

f: vector field of p

f̄: vector field of p̄

u: projective vector field of q

ū: projective vector field of q̄

g: projective vector field of p

ḡ: projective vector field of p̄

internal: internal variables of the integrator (e.g., internal stages of a Runge-Kutta methods or solver output)

Constructors

SolutionStepPDAE(t::TT, q::AT, p::AT, λ::AT, internal::IT=NamedTuple())

GeometricIntegrators.Solutions.SolutionStepPODE — Type

Solution step for a PODEProblem.

Parameters

DT: data type

TT: time step type

AT: array type

IT: internal variable types

Fields

t: time of current time step

t̄: time of previous time step

q: current solution of q

q̄: previous solution of q

q̃: compensated summation error of q

p: current solution of p

p̄: previous solution of p

p̃: compensated summation error of p

v: vector field of q

v̄: vector field of q̄

f: vector field of p

f̄: vector field of p̄

internal: internal variables of the integrator (e.g., internal stages of a Runge-Kutta methods or solver output)

Constructors

SolutionStepPODE(t::TT, q::AT, p::AT, internal::IT=NamedTuple())

Base.copy! — Method

Copy the initial conditions of a GeometricProblem to the current state of an atomic solution.

GeometricIntegrators.Solutions.current — Function

Returns a NamedTuple with the solution of the current time step.

GeometricIntegrators.Solutions.previous — Function

Returns a NamedTuple with the solution of the previous time step.