Solutions

Solutions

GeometricIntegrators.Solutions.PSolutionPDAE — Type.

Parallel Solution of a partitioned differential algebraic equation.

source

GeometricIntegrators.Solutions.SSolutionPDAE — Type.

Serial Solution of a partitioned differential algebraic equation.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for variational ODE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for ODE and split ODE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for DAE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for implicit DAE.

source

GeometricIntegrators.Solutions.Solution — Type.

Print error for solutions of equations not implemented, yet.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for implicit ODE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for partitioned DAE.

source

GeometricIntegrators.Solutions.Solution — Type.

Create solution for partitioned ODE.

source

GeometricIntegrators.Solutions.SolutionDAE — Type.

Solution of a differential algebraic equation.

source

GeometricIntegrators.Solutions.SolutionODE — Type.

SolutionODE: Solution of an ordinary differential equation

Contains all fields necessary to store the solution of an ODE.

Fields

nd: dimension of the dynamical variable $q$

nt: number of time steps to store

ni: number of initial conditions

t: time steps

q: solution q[nd, nt+1, ni] with q[:,0,:] the initial conditions

ntime: number of time steps to compute

nsave: save every nsave'th time step

source

GeometricIntegrators.Solutions.SolutionPODE — Type.

Solution of a partitioned ordinary differential equation.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.CommonFunctions.write_to_hdf5 — Method.

Append solution to HDF5 file.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for DAE solution object.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for DAE solution object.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for ODE solution object.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for ODE solution object.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for PDAE solution object with single initial condition.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for PDAE solution object with multiple initial conditions.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for PODE solution object.

source

GeometricIntegrators.Solutions.create_hdf5 — Method.

Creates HDF5 file and initialises datasets for PODE solution object.

source

GeometricIntegrators.Solutions.createHDF5 — Function.

createHDF5: Creates or opens HDF5 file.

source

GeometricIntegrators.Solutions.writeSolutionToHDF5 — Method.

Creates HDF5 file, writes solution to file, and closes file.

source