Tableaus

GeometricIntegrators.Tableaus.getTableauBurrageCL — Method

Tableau for the explicit 4-stage CL method due to K. Burrage and P. Burrage Method cited in Eq. (56) in K. Burrage, P. Burrage (1996) "High strong order explicit Runge-Kutta methods for stochastic ordinary differential equations". According to the paper, the method has strong order 1.5 for one-dimensional Brownian motion. Reduces to the classical R-K method of order 4 when noise is zero.

GeometricIntegrators.Tableaus.getTableauBurrageE1 — Method

Tableau for the explicit 4-stage E1 method due to K. Burrage and P. Burrage Method cited in Eq. (4.2)-(4.3) in K. Burrage, P. Burrage (2000) "Order conditions for stochastic Runge-Kutta methods by B-series". According to the paper, the method has strong order 1.0 for one-dimensional Brownian motion.

GeometricIntegrators.Tableaus.getTableauBurrageG5 — Method

Tableau for the explicit 5-stage G5 method due to K. Burrage and P. Burrage Method cited in Section 4 of K. Burrage, P. Burrage (2000) "Order conditions for stochastic Runge-Kutta methods by B-series". According to the paper, the method has strong order 1.5 for one-dimensional Brownian motion.

GeometricIntegrators.Tableaus.getTableauBurrageR2 — Method

Tableau for the explicit 2-stage R2 method due to K. Burrage and P. Burrage Method cited in Eq. (51) in K. Burrage, P. Burrage (1996) "High strong order explicit Runge-Kutta methods for stochastic ordinary differential equations". According to the paper, the method has strong order 1.0 for one-dimensional Brownian motion

GeometricIntegrators.Tableaus.getTableauERK4 — Method

Tableau for explicit Runge-Kutta method of order four (1/6 rule)

GeometricIntegrators.Tableaus.getTableauERK438 — Method

Tableau for explicit Runge-Kutta method of order four (3/8 rule)

GeometricIntegrators.Tableaus.getTableauExplicitEuler — Method

Tableau for explicit Euler method

GeometricIntegrators.Tableaus.getTableauExplicitMidpoint — Method

Tableau for explicit midpoint method

GeometricIntegrators.Tableaus.getTableauGLRK — Method

Gauss-Legendre Runge-Kutta

GeometricIntegrators.Tableaus.getTableauGLRKpSymplectic — Method

Tableau for Gauss-Legendre method with s stages and symplectic projection.

GeometricIntegrators.Tableaus.getTableauHPARKGLRK — Method

Tableau for Gauss-Legendre HPARK method with s stages.

GeometricIntegrators.Tableaus.getTableauHPARKLobIIIAIIIB2 — Method

Tableau for Gauss-Lobatto IIIA-IIIB HPARK method with two stages.

GeometricIntegrators.Tableaus.getTableauHPARKLobIIIAIIIB3 — Method

Tableau for Gauss-Lobatto IIIA-IIIB HPARK method with three stages.

GeometricIntegrators.Tableaus.getTableauHPARKLobIIIAIIIB4 — Method

Tableau for Gauss-Lobatto IIIA-IIIB HPARK method with four stages.

GeometricIntegrators.Tableaus.getTableauHSPARKGLRKpSymmetric — Method

Tableau for Gauss-Legendre method with s stages and symplectic projection.

GeometricIntegrators.Tableaus.getTableauHSPARKLobIIIAIIIB2pSymmetric — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with two stages and symmetric projection.

GeometricIntegrators.Tableaus.getTableauHSPARKLobIIIAIIIB3pSymmetric — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with three stages and symmetric projection.

GeometricIntegrators.Tableaus.getTableauHSPARKLobIIIAIIIB4pSymmetric — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with four stages and symmetric projection.

GeometricIntegrators.Tableaus.getTableauHeun — Method

Tableau for Heun's method

GeometricIntegrators.Tableaus.getTableauImplicitEuler — Method

Implicit Euler

GeometricIntegrators.Tableaus.getTableauImplicitMidpoint — Method

Implicit Midpoint

GeometricIntegrators.Tableaus.getTableauKutta — Method

Tableau for Kutta's method of order three

GeometricIntegrators.Tableaus.getTableauLobIIIA2 — Method

Gauss-Lobatto-IIIA Runge-Kutta, s=2

GeometricIntegrators.Tableaus.getTableauLobIIIA3 — Method

Gauss-Lobatto-IIIA Runge-Kutta, s=3

GeometricIntegrators.Tableaus.getTableauLobIIIA4 — Method

Gauss-Lobatto-IIIA Runge-Kutta, s=4

GeometricIntegrators.Tableaus.getTableauLobIIIAIIIB2pSymplectic — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with two stages and symplectic projection.

GeometricIntegrators.Tableaus.getTableauLobIIIAIIIB3pSymplectic — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with three stages and symplectic projection.

GeometricIntegrators.Tableaus.getTableauLobIIIAIIIB4pSymplectic — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with four stages and symplectic projection.

GeometricIntegrators.Tableaus.getTableauLobIIIB2 — Method

Gauss-Lobatto-IIIB Runge-Kutta, s=2

GeometricIntegrators.Tableaus.getTableauLobIIIB3 — Method

Gauss-Lobatto-IIIB Runge-Kutta, s=3

GeometricIntegrators.Tableaus.getTableauLobIIIB4 — Method

Gauss-Lobatto-IIIB Runge-Kutta, s=4

GeometricIntegrators.Tableaus.getTableauLobIIIC2 — Method

Gauss-Lobatto-IIIC Runge-Kutta, s=2

GeometricIntegrators.Tableaus.getTableauLobIIIC3 — Method

Gauss-Lobatto-IIIC Runge-Kutta, s=3

GeometricIntegrators.Tableaus.getTableauLobIIIC4 — Method

Gauss-Lobatto-IIIC Runge-Kutta, s=4

GeometricIntegrators.Tableaus.getTableauLobIIID2 — Method

Gauss-Lobatto-IIID Runge-Kutta, s=2

GeometricIntegrators.Tableaus.getTableauLobIIID3 — Method

Gauss-Lobatto-IIID Runge-Kutta, s=3

GeometricIntegrators.Tableaus.getTableauLobIIID4 — Method

Gauss-Lobatto-IIID Runge-Kutta, s=4

GeometricIntegrators.Tableaus.getTableauLobIIIE2 — Method

Gauss-Lobatto-IIIE Runge-Kutta, s=2

GeometricIntegrators.Tableaus.getTableauLobIIIE3 — Method

Gauss-Lobatto-IIIE Runge-Kutta, s=3

GeometricIntegrators.Tableaus.getTableauLobIIIE4 — Method

Gauss-Lobatto-IIIE Runge-Kutta, s=4

GeometricIntegrators.Tableaus.getTableauLobIIIF2 — Method

Gauss-Lobatto-IIIF Runge-Kutta, s=2

GeometricIntegrators.Tableaus.getTableauLobIIIF3 — Method

Gauss-Lobatto-IIIF Runge-Kutta, s=3

GeometricIntegrators.Tableaus.getTableauLobIIIF4 — Method

Gauss-Lobatto-IIIF Runge-Kutta, s=4

GeometricIntegrators.Tableaus.getTableauLobIIIG2 — Method

Gauss-Lobatto-IIIG Runge-Kutta, s=2

GeometricIntegrators.Tableaus.getTableauLobIIIG3 — Method

Gauss-Lobatto-IIIG Runge-Kutta, s=3

GeometricIntegrators.Tableaus.getTableauLobIIIG4 — Method

Gauss-Lobatto-IIIG Runge-Kutta, s=4

GeometricIntegrators.Tableaus.getTableauLobattoIIIAIIIB2 — Method

Tableau for Gauss-Lobatto IIIAIIIB method with s=2 stages

GeometricIntegrators.Tableaus.getTableauLobattoIIIBIIIA2 — Method

Tableau for Gauss-Lobatto IIIBIIIA method with s=2 stages

GeometricIntegrators.Tableaus.getTableauModifiedStochasticStormerVerlet — Function

Tableau for the 2-stage modified stochastic LobattoIIIA-IIIB method Tableau for the 2-stage modified stochastic LobattoIIIA-IIIB method Satisfies the conditions for Lagrange-d'Alembert integrators and the conditions for convergence of order 1.0 for one Wiener process

GeometricIntegrators.Tableaus.getTableauPlaten — Method

Tableau for the explicit Platen method Platen's method cited in Eq. (52) in K. Burrage, P. Burrage (1996) "High strong order explicit Runge-Kutta methods for stochastic ordinary differential equations". According to the paper, the method has strong order 1.0 for one-dimensional Brownian motion. Appears to have a rather poor long-time performance.

GeometricIntegrators.Tableaus.getTableauRadIIA2 — Method

Gauss-Radau-IIA Runge-Kutta, s=2

GeometricIntegrators.Tableaus.getTableauRadIIA3 — Method

Gauss-Radau-IIA Runge-Kutta, s=3

GeometricIntegrators.Tableaus.getTableauRosslerRS1 — Method

Tableau for the explicit 4-stage RS1 method due to Andreas Rossler Method cited in Table 5.2 in Andreas Rossler, "Second order Runge-Kutta methods for Stratonovich stochastic differential equations", BIT Numerical Mathematics (2007) 47 According to the paper, the method has weak order 2.0.

GeometricIntegrators.Tableaus.getTableauRosslerRS2 — Method

Tableau for the explicit 4-stage RS2 method due to Andreas Rossler Method cited in Table 5.3 in Andreas Rossler, "Second order Runge-Kutta methods for Stratonovich stochastic differential equations", BIT Numerical Mathematics (2007) 47 According to the paper, the method has weak order 2.0.

GeometricIntegrators.Tableaus.getTableauRunge — Method

Tableau for Runge's method

GeometricIntegrators.Tableaus.getTableauSPARKGLRK — Method

SPARK tableau for Gauss-Legendre Runge-Kutta method with s stages.

GeometricIntegrators.Tableaus.getTableauSPARKGLRKLobIIIAIIIB — Function

SPARK tableau for Gauss-Legendre/Gauss-Lobatto methods.

GeometricIntegrators.Tableaus.getTableauSPARKLob — Method

SPARK tableau for Gauss-Lobatto methods.

GeometricIntegrators.Tableaus.getTableauSPARKLobIIIAIIIB — Method

SPARK tableau for Gauss-Lobatto IIIA-IIIB method with s stages.

GeometricIntegrators.Tableaus.getTableauSRK3 — Method

Gauss-Legendre Runge-Kutta, s=3

GeometricIntegrators.Tableaus.getTableauSRKw1 — Function

Tableau for the 1-stage SRKw1 method due to Wang, Hong & Xu Method cited in Wang, Hong, Xu, "Construction of Symplectic Runge-Kutta Methods for Stochastic Hamiltonian Systems", Commun. Comput. Phys. 21(1), 2017 According to the paper, the method has weak order 1.0.

GeometricIntegrators.Tableaus.getTableauSRKw2 — Function

Tableau for the 4-stage SRKw2 method due to Wang, Hong & Xu Method cited in Wang, Hong, Xu, "Construction of Symplectic Runge-Kutta Methods for Stochastic Hamiltonian Systems", Commun. Comput. Phys. 21(1), 2017 According to the paper, the method has weak order 2.0 when applied to systems driven by one-dimensional noise.

GeometricIntegrators.Tableaus.getTableauStochasticDIRK — Function

Tableau for the 2-stage stochastic symplectic DIRK method Tableau for the stochastic symplectic DIRK method Satisfies the conditions for Lagrange-d'Alembert integrators. Satisfies the conditions for strong convergence of order 1.0 for one Wiener process

GeometricIntegrators.Tableaus.getTableauStochasticEuler — Method

Tableau for the explicit 1-stage stochastic Euler method

GeometricIntegrators.Tableaus.getTableauStochasticGLRK — Method

Tableau for the s-stage Gauss-Lobatto SFIRK method

GeometricIntegrators.Tableaus.getTableauStochasticHeun — Method

Tableau for the explicit 2-stage stochastic Heun method

GeometricIntegrators.Tableaus.getTableauStochasticLobIIIABD2 — Method

Tableau for the 2-stage stochastic LobattoIIIA-IIIB-IIID method Tableau for the 2-stage stochastic LobattoIIIA-IIIB-IIID method (based on the deterministic LobattoIIIA-IIIB-IIID due to L. Jay) It satisfies the conditions for convergence of order 1.0 for one Wiener process, but it doesn't satisfy the conditions for Lagrange-d'Alembert integrators

GeometricIntegrators.Tableaus.getTableauStochasticStormerVerlet — Method

Tableau for the 2-stage stochastic LobattoIIA-IIB method (Stormer-Verlet)

GeometricIntegrators.Tableaus.getTableauStochasticSymplecticEuler — Method

Tableau for the stochastic symplectic Euler method Tableau for the stochastic symplectic Euler method Satisfies the conditions for Lagrange-d'Alembert integrators. Satisfies the conditions for strong convergence of order 1.0 for one Wiener process for special choices of the stochastic Hamiltonians and forces, e.g., h=h(q), f=0.

GeometricIntegrators.Tableaus.getTableauSymplecticEulerA — Method

Tableau for symplectic Euler-A method

GeometricIntegrators.Tableaus.getTableauSymplecticEulerB — Method

Tableau for symplectic Euler-B method

GeometricIntegrators.Tableaus.getTableauVPGLRK — Method

Tableau for variational Gauss-Legendre method with s stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIA2 — Method

Tableau for variational Gauss-Lobatto IIIA-IIIB method with two stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIA3 — Method

Tableau for variational Gauss-Lobatto IIIA-IIIB method with three stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIA4 — Method

Tableau for variational Gauss-Lobatto IIIA-IIIB method with four stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIAIIIA2 — Method

Tableau for Gauss-Lobatto IIIA-IIIA method with two stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIAIIIA3 — Method

Tableau for Gauss-Lobatto IIIA-IIIA method with three stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIAIIIA4 — Method

Tableau for Gauss-Lobatto IIIA-IIIA method with four stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIB2 — Method

Tableau for variational Gauss-Lobatto IIIA-IIIB method with two stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIB3 — Method

Tableau for variational Gauss-Lobatto IIIA-IIIB method with three stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIB4 — Method

Tableau for variational Gauss-Lobatto IIIA-IIIB method with four stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIC2 — Method

Tableau for variational Gauss-Lobatto IIIC-III method with two stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIC3 — Method

Tableau for variational Gauss-Lobatto IIIC-III method with three stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIC4 — Method

Tableau for variational Gauss-Lobatto IIIC-III method with four stages

GeometricIntegrators.Tableaus.getTableauVPLobIIID2 — Method

Tableau for variational Gauss-Lobatto IIID method with two stages

GeometricIntegrators.Tableaus.getTableauVPLobIIID3 — Method

Tableau for variational Gauss-Lobatto IIID method with three stages

GeometricIntegrators.Tableaus.getTableauVPLobIIID4 — Method

Tableau for variational Gauss-Lobatto IIID method with four stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIE2 — Method

Tableau for variational Gauss-Lobatto IIIE method with two stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIE3 — Method

Tableau for variational Gauss-Lobatto IIIE method with three stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIE4 — Method

Tableau for variational Gauss-Lobatto IIIE method with four stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIF2 — Method

Tableau for variational Gauss-Lobatto IIIF method with two stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIF3 — Method

Tableau for variational Gauss-Lobatto IIIF method with three stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIF4 — Method

Tableau for variational Gauss-Lobatto IIIF method with four stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIG2 — Method

Tableau for variational Gauss-Lobatto IIIG method with two stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIG3 — Method

Tableau for variational Gauss-Lobatto IIIG method with three stages

GeometricIntegrators.Tableaus.getTableauVPLobIIIG4 — Method

Tableau for variational Gauss-Lobatto IIIG method with four stages

GeometricIntegrators.Tableaus.getTableauVPRadIIAIIA2 — Method

Tableau for Gauss-Radau IIA-IIA method with two stages

GeometricIntegrators.Tableaus.getTableauVPRadIIAIIA3 — Method

Tableau for Gauss-Radau IIA-IIA method with three stages

GeometricIntegrators.Tableaus.getTableauVPSRK3 — Method

Tableau for variational symmetric Runge-Kutta method with 3 stages

GeometricIntegrators.Tableaus.getTableauVSPARKGLRKpMidpoint — Method

Tableau for Gauss-Legendre method with s stages and midpoint projection.

GeometricIntegrators.Tableaus.getTableauVSPARKGLRKpSymmetric — Method

Tableau for Gauss-Legendre method with s stages and symplectic projection.

GeometricIntegrators.Tableaus.getTableauVSPARKLobIIIAIIIB2pSymmetric — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with two stages and symmetric projection.

GeometricIntegrators.Tableaus.getTableauVSPARKLobIIIAIIIB3pSymmetric — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with three stages and symmetric projection.

GeometricIntegrators.Tableaus.getTableauVSPARKLobIIIAIIIB4pSymmetric — Method

Tableau for Gauss-Lobatto IIIA-IIIB method with four stages and symmetric projection.