Tableaus

Tableaus

GeometricIntegrators.Tableaus.getTableauBurrageCL — Method.

Tableau for the explicit 4-stage CL method due to K. Burrage and P. Burrage Method cited in Eq. (56) in K. Burrage, P. Burrage (1996) "High strong order explicit Runge-Kutta methods for stochastic ordinary differential equations". According to the paper, the method has strong order 1.5 for one-dimensional Brownian motion. Reduces to the classical R-K method of order 4 when noise is zero.

source

GeometricIntegrators.Tableaus.getTableauBurrageE1 — Method.

Tableau for the explicit 4-stage E1 method due to K. Burrage and P. Burrage Method cited in Eq. (4.2)-(4.3) in K. Burrage, P. Burrage (2000) "Order conditions for stochastic Runge-Kutta methods by B-series". According to the paper, the method has strong order 1.0 for one-dimensional Brownian motion.

source

GeometricIntegrators.Tableaus.getTableauBurrageG5 — Method.

Tableau for the explicit 5-stage G5 method due to K. Burrage and P. Burrage Method cited in Section 4 of K. Burrage, P. Burrage (2000) "Order conditions for stochastic Runge-Kutta methods by B-series". According to the paper, the method has strong order 1.5 for one-dimensional Brownian motion.

source

GeometricIntegrators.Tableaus.getTableauBurrageR2 — Method.

Tableau for the explicit 2-stage R2 method due to K. Burrage and P. Burrage Method cited in Eq. (51) in K. Burrage, P. Burrage (1996) "High strong order explicit Runge-Kutta methods for stochastic ordinary differential equations". According to the paper, the method has strong order 1.0 for one-dimensional Brownian motion

source

GeometricIntegrators.Tableaus.getTableauERK4 — Method.

Tableau for explicit Runge-Kutta method of order four (1/6 rule)

source

GeometricIntegrators.Tableaus.getTableauERK438 — Method.

Tableau for explicit Runge-Kutta method of order four (3/8 rule)

source

GeometricIntegrators.Tableaus.getTableauExplicitEuler — Method.

Tableau for explicit Euler method

source

GeometricIntegrators.Tableaus.getTableauExplicitMidpoint — Method.

Tableau for explicit midpoint method

source

GeometricIntegrators.Tableaus.getTableauGLRKpSymmetric — Method.

Tableau for Gauss-Legendre method with s stages and symplectic projection.

source

GeometricIntegrators.Tableaus.getTableauGLRKpSymplectic — Method.

Tableau for Gauss-Legendre method with s stages and symplectic projection.

source

GeometricIntegrators.Tableaus.getTableauHeun — Method.

Tableau for Heun's method

source

GeometricIntegrators.Tableaus.getTableauImplicitEuler — Method.

Implicit Euler

source

GeometricIntegrators.Tableaus.getTableauImplicitMidpoint — Method.

Implicit Midpoint

source

GeometricIntegrators.Tableaus.getTableauKutta — Method.

Tableau for Kutta's method of order three

source

GeometricIntegrators.Tableaus.getTableauLobIIIA2 — Method.

Gauss-Lobatto-IIIA Runge-Kutta, s=2

source

GeometricIntegrators.Tableaus.getTableauLobIIIA3 — Method.

Gauss-Lobatto-IIIA Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauLobIIIA4 — Method.

Gauss-Lobatto-IIIA Runge-Kutta, s=4

source

GeometricIntegrators.Tableaus.getTableauLobIIIAIIIB2pSymmetric — Method.

Tableau for Gauss-Lobatto IIIA-IIIB method with two stages and symmetric projection.

source

GeometricIntegrators.Tableaus.getTableauLobIIIAIIIB2pSymplectic — Method.

Tableau for Gauss-Lobatto IIIA-IIIB method with two stages and symplectic projection.

source

GeometricIntegrators.Tableaus.getTableauLobIIIAIIIB3pSymmetric — Method.

Tableau for Gauss-Lobatto IIIA-IIIB method with three stages and symmetric projection.

source

GeometricIntegrators.Tableaus.getTableauLobIIIAIIIB3pSymplectic — Method.

Tableau for Gauss-Lobatto IIIA-IIIB method with three stages and symplectic projection.

source

GeometricIntegrators.Tableaus.getTableauLobIIIB2 — Method.

Gauss-Lobatto-IIIB Runge-Kutta, s=2

source

GeometricIntegrators.Tableaus.getTableauLobIIIB3 — Method.

Gauss-Lobatto-IIIB Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauLobIIIB4 — Method.

Gauss-Lobatto-IIIB Runge-Kutta, s=4

source

GeometricIntegrators.Tableaus.getTableauLobIIIC2 — Method.

Gauss-Lobatto-IIIC Runge-Kutta, s=2

source

GeometricIntegrators.Tableaus.getTableauLobIIIC3 — Method.

Gauss-Lobatto-IIIC Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauLobIIIC4 — Method.

Gauss-Lobatto-IIIC Runge-Kutta, s=4

source

GeometricIntegrators.Tableaus.getTableauLobIIID2 — Method.

Gauss-Lobatto-IIID Runge-Kutta, s=2

source

GeometricIntegrators.Tableaus.getTableauLobIIID3 — Method.

Gauss-Lobatto-IIID Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauLobIIID4 — Method.

Gauss-Lobatto-IIID Runge-Kutta, s=4

source

GeometricIntegrators.Tableaus.getTableauLobIIIE2 — Method.

Gauss-Lobatto-IIIE Runge-Kutta, s=2

source

GeometricIntegrators.Tableaus.getTableauLobIIIE3 — Method.

Gauss-Lobatto-IIIE Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauLobIIIE4 — Method.

Gauss-Lobatto-IIIE Runge-Kutta, s=4

source

GeometricIntegrators.Tableaus.getTableauLobIIIF2 — Method.

Gauss-Lobatto-IIIF Runge-Kutta, s=2

source

GeometricIntegrators.Tableaus.getTableauLobIIIF3 — Method.

Gauss-Lobatto-IIIF Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauLobIIIF4 — Method.

Gauss-Lobatto-IIIF Runge-Kutta, s=4

source

GeometricIntegrators.Tableaus.getTableauLobIIIG2 — Method.

Gauss-Lobatto-IIIG Runge-Kutta, s=2

source

GeometricIntegrators.Tableaus.getTableauLobIIIG3 — Method.

Gauss-Lobatto-IIIG Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauLobIIIG4 — Method.

Gauss-Lobatto-IIIG Runge-Kutta, s=4

source

GeometricIntegrators.Tableaus.getTableauLobattoIIIAIIIB2 — Method.

Tableau for Gauss-Lobatto IIIAIIIB method with s=2 stages

source

GeometricIntegrators.Tableaus.getTableauLobattoIIIBIIIA2 — Method.

Tableau for Gauss-Lobatto IIIBIIIA method with s=2 stages

source

GeometricIntegrators.Tableaus.getTableauModifiedStochasticStormerVerlet — Function.

Tableau for the 2-stage modified stochastic LobattoIIIA-IIIB method Tableau for the 2-stage modified stochastic LobattoIIIA-IIIB method Satisfies the conditions for Lagrange-d'Alembert integrators and the conditions for convergence of order 1.0 for one Wiener process

source

GeometricIntegrators.Tableaus.getTableauPlaten — Method.

Tableau for the explicit Platen method Platen's method cited in Eq. (52) in K. Burrage, P. Burrage (1996) "High strong order explicit Runge-Kutta methods for stochastic ordinary differential equations". According to the paper, the method has strong order 1.0 for one-dimensional Brownian motion. Appears to have a rather poor long-time performance.

source

GeometricIntegrators.Tableaus.getTableauRadIIA2 — Method.

Gauss-Radau-IIA Runge-Kutta, s=2

source

GeometricIntegrators.Tableaus.getTableauRadIIA3 — Method.

Gauss-Radau-IIA Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauRosslerRS1 — Method.

Tableau for the explicit 4-stage RS1 method due to Andreas Rossler Method cited in Table 5.2 in Andreas Rossler, "Second order Runge-Kutta methods for Stratonovich stochastic differential equations", BIT Numerical Mathematics (2007) 47 According to the paper, the method has weak order 2.0.

source

GeometricIntegrators.Tableaus.getTableauRosslerRS2 — Method.

Tableau for the explicit 4-stage RS2 method due to Andreas Rossler Method cited in Table 5.3 in Andreas Rossler, "Second order Runge-Kutta methods for Stratonovich stochastic differential equations", BIT Numerical Mathematics (2007) 47 According to the paper, the method has weak order 2.0.

source

GeometricIntegrators.Tableaus.getTableauSRK3 — Method.

Gauss-Legendre Runge-Kutta, s=3

source

GeometricIntegrators.Tableaus.getTableauSRKw1 — Function.

Tableau for the 1-stage SRKw1 method due to Wang, Hong & Xu Method cited in Wang, Hong, Xu, "Construction of Symplectic Runge-Kutta Methods for Stochastic Hamiltonian Systems", Commun. Comput. Phys. 21(1), 2017 According to the paper, the method has weak order 1.0.

source

GeometricIntegrators.Tableaus.getTableauSRKw2 — Function.

Tableau for the 4-stage SRKw2 method due to Wang, Hong & Xu Method cited in Wang, Hong, Xu, "Construction of Symplectic Runge-Kutta Methods for Stochastic Hamiltonian Systems", Commun. Comput. Phys. 21(1), 2017 According to the paper, the method has weak order 2.0 when applied to systems driven by one-dimensional noise.

source

GeometricIntegrators.Tableaus.getTableauStochasticDIRK — Function.

Tableau for the 2-stage stochastic symplectic DIRK method Tableau for the stochastic symplectic DIRK method Satisfies the conditions for Lagrange-d'Alembert integrators. Satisfies the conditions for strong convergence of order 1.0 for one Wiener process

source

GeometricIntegrators.Tableaus.getTableauStochasticGLRK — Method.

Tableau for the s-stage Gauss-Lobatto SFIRK method

source

GeometricIntegrators.Tableaus.getTableauStochasticHeun — Method.

Tableau for the explicit 2-stage stochastic Heun method

source

GeometricIntegrators.Tableaus.getTableauStochasticLobIIIABD2 — Method.

Tableau for the 2-stage stochastic LobattoIIIA-IIIB-IIID method Tableau for the 2-stage stochastic LobattoIIIA-IIIB-IIID method (based on the deterministic LobattoIIIA-IIIB-IIID due to L. Jay) It satisfies the conditions for convergence of order 1.0 for one Wiener process, but it doesn't satisfy the conditions for Lagrange-d'Alembert integrators

source

GeometricIntegrators.Tableaus.getTableauStochasticStormerVerlet — Method.

Tableau for the 2-stage stochastic LobattoIIA-IIB method (Stormer-Verlet)

source

GeometricIntegrators.Tableaus.getTableauStochasticSymplecticEuler — Method.

Tableau for the stochastic symplectic Euler method Tableau for the stochastic symplectic Euler method Satisfies the conditions for Lagrange-d'Alembert integrators. Satisfies the conditions for strong convergence of order 1.0 for one Wiener process for special choices of the stochastic Hamiltonians and forces, e.g., h=h(q), f=0.

source

GeometricIntegrators.Tableaus.getTableauSymplecticEulerA — Method.

Tableau for symplectic Euler-A method

source

GeometricIntegrators.Tableaus.getTableauSymplecticEulerB — Method.

Tableau for symplectic Euler-B method

source

GeometricIntegrators.Tableaus.getTableauVPGLRK — Method.

Tableau for variational Gauss-Legendre method with s stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIA2 — Method.

Tableau for variational Gauss-Lobatto IIIA-IIIB method with two stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIA3 — Method.

Tableau for variational Gauss-Lobatto IIIA-IIIB method with three stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIA4 — Method.

Tableau for variational Gauss-Lobatto IIIA-IIIB method with four stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIAIIIA2 — Method.

Tableau for Gauss-Lobatto IIIA-IIIA method with two stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIAIIIA3 — Method.

Tableau for Gauss-Lobatto IIIA-IIIA method with three stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIAIIIA4 — Method.

Tableau for Gauss-Lobatto IIIA-IIIA method with four stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIB2 — Method.

Tableau for variational Gauss-Lobatto IIIA-IIIB method with two stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIB3 — Method.

Tableau for variational Gauss-Lobatto IIIA-IIIB method with three stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIB4 — Method.

Tableau for variational Gauss-Lobatto IIIA-IIIB method with four stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIC2 — Method.

Tableau for variational Gauss-Lobatto IIIC-III method with two stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIC3 — Method.

Tableau for variational Gauss-Lobatto IIIC-III method with three stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIC4 — Method.

Tableau for variational Gauss-Lobatto IIIC-III method with four stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIID2 — Method.

Tableau for variational Gauss-Lobatto IIID method with two stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIID3 — Method.

Tableau for variational Gauss-Lobatto IIID method with three stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIID4 — Method.

Tableau for variational Gauss-Lobatto IIID method with four stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIE2 — Method.

Tableau for variational Gauss-Lobatto IIIE method with two stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIE3 — Method.

Tableau for variational Gauss-Lobatto IIIE method with three stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIE4 — Method.

Tableau for variational Gauss-Lobatto IIIE method with four stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIF2 — Method.

Tableau for variational Gauss-Lobatto IIIF method with two stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIF3 — Method.

Tableau for variational Gauss-Lobatto IIIF method with three stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIF4 — Method.

Tableau for variational Gauss-Lobatto IIIF method with four stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIG2 — Method.

Tableau for variational Gauss-Lobatto IIIG method with two stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIG3 — Method.

Tableau for variational Gauss-Lobatto IIIG method with three stages

source

GeometricIntegrators.Tableaus.getTableauVPLobIIIG4 — Method.

Tableau for variational Gauss-Lobatto IIIG method with four stages

source

GeometricIntegrators.Tableaus.getTableauVPRadIIAIIA2 — Method.

Tableau for Gauss-Radau IIA-IIA method with two stages

source

GeometricIntegrators.Tableaus.getTableauVPRadIIAIIA3 — Method.

Tableau for Gauss-Radau IIA-IIA method with three stages

source

GeometricIntegrators.Tableaus.getTableauVPSRK3 — Method.

Tableau for variational symmetric Runge-Kutta method with 3 stages

source